### 2019-03-22 #1 @ocefpaf
Submit erddapy for review
Submitting Author: Filipe Fernandes (@ocefpaf)
Repository Link:  https://github.com/ioos/erddapy
Version submitted: v0.4.0
Reviewer 1: @jlpalomino @choldgraf
Version accepted: TBD

---

```
erddapy is a python wrapper to ERDDAP's RESTful web services to create valid ERDDAP URLs for data requests, searches, metadata queries, etc.
```

- Please indicate which [category or categories](https://www.pyopensci.org/dev_guide/peer_review/aims_scope.html) this package falls under:

\* Please fill out a pre-submission inquiry before submitting a data visualization package. For more info, see [this section](https://www.pyopensci.org/dev_guide/peer_review/aims_scope.html#notes-on-categories) of our guidebook.

- Explain how the and why the package falls under these categories (briefly, 1-2 sentences):

erddapy is a tool to construct URLs for retrieving data from ERDDAP servers.

The target audience is anyone who needs to access that in ERDDAP servers, they are usually Met/Ocean scientists who need near-real time access to data. erddapy does not have a direct scientific application b/c that varies based on the data available. It is essentially a retrieval tool. However, ERDDAP servers are essentially scientific data servers, erddapy's goal is to make access to that data easier.

Not to my knowledge.

-   If you made a pre-submission enquiry, please paste the link to the corresponding issue, forum post, or other discussion, or @tag the editor you contacted:

For details about the pyOpenSci packaging requirements, see our [packaging guide](https://www.pyopensci.org/dev_guide/packaging/packaging_guide.html). Confirm each of the following by checking the box.  This package:

- [X] has an [OSI approved license](https://opensource.org/licenses)
- [X] contains a vignette notebook with examples of its essential functions and uses.

- [ ] The package has an **obvious research application** according to JOSS's definition in their [submission requirements](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements). Be aware that completing the pyOpenSci review process **does not** guarantee acceptance to JOSS. Be sure to read their submission requirements (linked above) if you are interested in submitting to JOSS.

- [X] I agree to abide by [pyOpenSci's Code of Conduct](https://www.pyopensci.org/dev_guide/peer_review/coc.html) during the review process and in maintaining my package should it be accepted.
