### 2019-04-03 #1 @ocefpaf
> @ocefpaf my apologies for missing this submission almost 2 weeks ago! in our meeting tomorrow, we will discuss who the reviewers will be for this submission. I will serve as the editor for it but also can review if need be!!

No problem. This is a test submission after all :smile:

BTW, one thing that comes to mind would be to set up a bot, similar to JOSS' bot, so the person submitting the package can get some communication as soon as s/he submits the PR. It can be something like: "Thanks for submitting, please read this and that while you wait for a review..."

> OPEN CALL: Does anyone want to review this submission?

We could have a review team to ping here.
Something like @pyOpenSci/reviewers.

> Also just to confirm, Filipe, you are NOT interested in JOSS at this point in time?

Not at this moment.

PS: I won't be able to make it to tomorrow's meeting.
I'm at AnacondaCon in case someone from the pyOpenSci team is here and wants to meet.
