### 2019-04-03 #1 @lwasser
@ocefpaf these are all great suggestions!! we will look into Joss' Bot. @kysolvik can you kindly look into what they have setup to provide an automated response??

I also really like the idea of having a set of reviewers that could be pinged. We might need to think that through given in the future we could have many submissions at a time. BUT let's keep talking as these are excellent suggestions.

enjoy the conference, Filipe!! We will catch you via zoom at the next meeting but i'm sure we will chat online before then.
