### 2019-04-03 #1 @kysolvik
JOSS has an excellent bot named Whedon: https://github.com/openjournals/whedon. It definitely seems to be pretty helpful. We could look into creating something similar!

In the mean time, it looks like rOpenSci just has the current editor in chief watch issues in their software-review repo and assign them as they come up. This could be an option until we're able to set up a bot.
