### 2019-04-09 #1 @lwasser
woo hoo! ok so i am going to serve as the editor for this submission. As a part of this process, i am reviewing  the dev_guide closely and making edits / changes to it. pr soon to come. Below are the checks that i'm looking for

### Editor checks:

- [x ] **Fit**: The package meets criteria for [fit](https://pyopensci.github.io/dev_guide/peer_review/aims_scope.html#package-categories) and [overlap](https://pyopensci.github.io/dev_guide/peer_review/aims_scope.html#package-categories).
DATA ACCESS (API WRAPPER - JOSS WOULD NOT REVIEW THIS i think??)
- [x ] **Automated tests:** Package has a testing suite and is tested via Travis-CI or another CI service.
TRAVIS is being used
- [x ] **License:** The package has an OSI accepted license
BSD-3

---

#### Editor comments

This package is ready for review! @choldgraf @jlpalomino @npch are you still willing to serve as reviewers? Please note that a review could be technical in nature (looking at tests, code, etc etc) or could focus more on usability and documentation. i think ideally we'd hit both via all of our reviewers.

If you are still able to review, please let me know what time frame works for you? I was going to suggest 2-3 weeks. Does that sound reasonable? if so i'll select a date.

---

Reviewers:
Due date:
