### 2019-04-09 #1 @choldgraf
Sounds reasonable to me - @jlpalomino how would you like to handle the joint review? Should we prepare it together then submit at once? Or we can just review independently and have a channel of communication somewhere? I'm happy w/ whatever
