### 2019-04-11 #1 @lwasser
thank you @jlpalomino @choldgraf as you go through the review process, will you please submit PR's to update docs as you find issues? thank you in advance!!!
