### 2019-04-11 #1 @lwasser
wonderful. thank you @npch !! i'll add you as a formal reviewer now! All please be sure to use the [reviewer template](https://www.pyopensci.org/dev_guide/appendices/templates#review-template) when you submit your review!!
