### 2019-04-12 7a9e0fe
CI improvements
