### 2019-04-29 #1 @npch
No - the README could be clearer on the target audience for erdapy (e.g. who are the types of researchers most likely to use ERDDAP) and what ERDDAP is (link to ERDDAP info and summarise what it's used for). Much of this information is on the erddapy website, so perhaps summarise and then link?

No - I expected to see short explicit installation instructions, i.e. `python setup.py install --user` or `pip install erddapy` (to install from PyPI). Also, see later for problematic dependencies.

Partially - there is a good example in the README, but I would have liked to have had it also show what the expected outputs would be (e.g. by showing the first five rows of the data frame produced).

This kind of information is on the [website](https://ioos.github.io/erddapy/quick_intro.html#Longer-introduction) but it isn't linked directly from the README, and I think it's important in the README to show the expected outputs of the example, so the example should go as far as showing a sensible output.

- [ ] **Function Documentation:** for all user-facing functions

Unclear - documentation appears to be being built, but I can't find it (this might be a problem on my system).

- [ ] **Examples** for all user-facing functions

They are on the [website](https://ioos.github.io/erddapy/quick_intro.html#Longer-introduction), but should be better linked from the README.

- [ ] **Community guidelines** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer`.

No -submitting bug reports are covered but it is not clear how others could (or should) contribute to erddapy. This could just be a simple "get in touch via email" or "submit a pull request".

#### For packages co-submitting to JOSS

Unable to confirm as it's not documented. However appears to work using `python setup.py install --user` but not `pip install erddapy`.

Yes, able to connect to the test ERDDAP server and retrieve information.

Yes, but please note the following couple of things that need to be addressed:
- Tests required installation of iris, but this seems to be an ambiguous requirement. See: https://pypi.org/project/iris/
- pytzdata is required for tests, but was not in the requirements-dev.txt.

- [X] **Continuous Integration:** Has continuous integration, such as Travis CI, AppVeyor, CircleCI, and/or others.

Yes - building on Travis CI.

- [ ] **Packaging guidelines**: The package conforms to the pyOpenSci [packaging guidelines]((../packaging/packaging_guide).

Summary of things missing to conform to pyOpenSci packaging guidelines:
- Explicit links to further documentation in README [good]
- Installation instructions [good]
- Links to vignettes in documentation [better/best]
- Improved demonstration usage [better/best]
- Citation information (there as badge, but could be explicit in readme) [better/best]

- First pass on review: 2 hours

---

See inline comments above.

This package looks pretty close to a great pyOpenSci submission - a few places where there is information to be linked, but is available elsewhere, and some things to be checked and corrected.
