### 2019-04-29 #1 @choldgraf
Just a quite note here that I had a bit of a hard time finding the link to the reviewer template...we should make sure that links for reviewers to use are clearly highlighted in the templates for starting the review process etc!
