### 2019-04-30 506e53a
some improvements to travis-ci config
