+++ .travis.yml
@@ -9,25 +9,25 @@ env:
 matrix:
   fast_finish: true
   include:
-    - name: "default"
-      env: TEST_TARGET=default PY=3.7
+    - name: "python-3.6"
+      env: PY=3.6
+    - name: "python-3.7"
+      env: PY=3.7
     - name: "coding_standards"
-      env: TEST_TARGET=coding_standards PY=3.7
+      env: PY=3.7
     - name: "docs"
-      env: TEST_TARGET=docs PY=3.7
+      env: PY=3.7
 
 before_install:
   # Install miniconda and create TEST env.
   - |
-    wget http://bit.ly/miniconda -O miniconda.sh
+    wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh
     bash miniconda.sh -b -p $HOME/miniconda
     export PATH="$HOME/miniconda/bin:$PATH"
     conda config --set always_yes yes --set changeps1 no --set show_channel_urls true
     conda update --quiet conda
     conda config --add channels conda-forge --force
-    conda install pycryptosat
     conda config --set channel_priority strict
-    conda config --set sat_solver pycryptosat
     conda config --set safety_checks disabled
     conda create --name TEST python=$PY --file requirements.txt --file requirements-dev.txt
     source activate TEST
@@ -42,18 +42,23 @@ script:
       pushd /tmp && pytest -n 2 -rxs --cov=erddapy tests && popd ;
     fi
 
-  - if [[ $TEST_TARGET == 'coding_standards' ]]; then
+  - if [[ $TRAVIS_JOB_NAME == 'coding_standards' ]]; then
       pytest --flake8 -m flake8 ;
     fi
 
-  - if [[ $TEST_TARGET == 'docs' ]]; then
-      set -e ;
+    - |
+    if [[ $TRAVIS_JOB_NAME == 'docs' ]]; then
+      set -e
       cp notebooks/{quick_intro.ipynb,searchfor.ipynb} docs/source/ ;
-      pushd docs ;
-      make clean html linkcheck ;
-      popd ;
-      python -m doctr deploy --sync .;
-      python -m doctr deploy --sync --no-require-master --built-docs docs/build/html "docs-$TRAVIS_BRANCH" ;
+      pushd docs
+      make clean html linkcheck
+      popd
+      if [[ -z "$TRAVIS_TAG" ]]; then 
+        python -m doctr deploy --build-tags --key-path github_deploy_key.enc --built-docs docs/_build/html dev
+      else
+        python -m doctr deploy --build-tags --key-path github_deploy_key.enc --built-docs docs/_build/html "version-$TRAVIS_TAG"
+        python -m doctr deploy --build-tags --key-path github_deploy_key.enc --built-docs docs/_build/html .
+      fi
     fi
 
 doctr:
