### 2019-05-02 #1 @lwasser
hey @npch thank you for your review!!
@choldgraf  NOTED...
What do you guys think about creating a thread here: https://pyopensci.discourse.group/c/review-process  for feedback on the review process? that way everyone can see it??

@choldgraf where would you like to see the review template starting the process?
Also just checking in on your / @jlpalomino review? is that coming along?

Thanks all!!
@ocefpaf please note Neils comments above re the review !! The next step will be to address the comments
