### 2019-05-02 #1 @jlpalomino
- Both the README and the docs should be expanded to provide more details for users new to ERDDAP. The README can be improved with additional details on ERDDAP such as the full name for ERDDAP and a brief description of what it is for unfamiliar users. This information could be copied from the docs (https://ioos.github.io/erddapy/), which contains a great first sentence briefly introducing ERDDAP. The docs can be improved with an explicit hyperlink to the ERDDAP site from the text "Environmental Research Division’s Data Access Program (ERDDAP)".

    - [ ] ISSUE: Add details on ERDDAP to README
        - In the README, the ERDDAP acronym should be spelled out as "Environmental Research Division’s Data Access Program (ERDDAP)" and should have an explicit hyperlink to ERDDAP site. In addition, the README should contain a brief description of ERDDAP and an explicit link to the docs for the users that want to learn more information about both ERDDAP and ERDDAPY (similar to first statement provided on docs landing page).

    - [ ] ISSUE: Add hyperlink for ERDDAP to docs
        - On the docs landing page, the text "Environmental Research Division’s Data Access Program (ERDDAP)" in the docs should also be expliclty hyperlinked to ERDDAP site (https://coastwatch.pfeg.noaa.gov/erddap/index.html).

    -  No installation instructions for package are provided. jlpalomino successfully installed package with `pip` using `pip install erddapy` and was able to run the example in README (i.e. connect to ERDDAP server and get data). In addition, we were able to set up the development environment using the packages listed in requirements-dev.txt, but had to add `pytz` which is not listed in requirements-dev.txt. Chris had difficulty properly installing the environment needed to run erddapy.

    - [ ] ISSUE: Add installation instructions to README and docs
        - There are no installation instructions for erddapy in the docs or in the README. A short section in the README and docs called "Installation" with the proper `pip install` command would be helpful as well as a command for installing in development mode. It would also be helpful to provide the necessary configuration files to re-create the environment for erddapy in Binder.

    - [ ] ISSUE: Add `pytz` to `requirements-dev.txt`.
        - `pytz` needs to be added to `requirements-dev.txt` for tests to run successfully

    - The README contains a good start of an example and it runs successfully locally. The example could be improved with comments explaining the code/workflow, some description of the purpose of the example (e.g. the kind of data that it is querying), and displaying the output of the code example. In addition, an example could be provided to show how to get list of the shortcuts available servers (as shown in the docs) and a brief description of what each one contains.

    - The docs contain more detailed vignettes but they are not easy to find, as most are located in the Longer Introduction section of the docs, which is quite long. The docs could be improved by renaming the Quick Introduction to Quick Start and by breaking up the various code sections in the Longer Introduction into separate vignettes with specific titles, similar to the existing examples like OPeNDAP response.

    - [ ] ISSUE: Expand example code in README
        - The existing example code in README can be improved with:
          - [ ] comments explaining code/workflow
          - [ ] brief description of purpose of the example (e.g. the kind of data that it is querying)
          - [ ] display of the output of the code
        - Add a new code example to show how to get list of the shortcuts available servers (as shown in the docs) and a brief description of what each one contains.

    - [ ] ISSUE: Rename and restructure Quick Introduction and Longer Introduction in the docs
        - The docs can be improved by renaming Quick Introduction to Quick Start and by breaking up the various code sections in the Longer Introduction into separate vignettes with specific titles, similar to the existing examples like OPeNDAP response.
        -
- [ ] **Function Documentation:** for all user-facing functions

    - User-facing functions contain appropriate docstrings in erddapy.py, though utilities.py is less documented (note that many functions in utilities.py are not user-facing). The docs for erddapy (https://ioos.github.io/erddapy/) also provide solid documentation for more functions, with the exception of the `to_` functions which have minimal information provided other than links to source code. At a minimum, it would be helpful to provide a list of keyword arguments for each.

    - Finding the link to erddapy functionality in the docs was not obvious, as it was listed as erddapy above Quick introduction in the table of contents and appeared as a subtitle with code formatting, rather than a link (https://ioos.github.io/erddapy/erddapy.html).

    - [ ] ISSUE: Add docstrings to user-facing functions in utilities.py
        - User-facing functions in utilities.py should contain apprpriopriate docstrings (e.g. parse_dates).

    - [ ] ISSUE: Expand documentation for `to_` functions in `erddapy`
        - The documentation for `to_iris`, `to_pandas`, and `to_xarray` should be expanded to contain a list of keyword arguments for each (e.g. rather than "Accepts any iris.load_raw keyword arguments.")

    - [ ] ISSUE: Rename and remove code formatting for `erddapy` in docs table of contents
        - `erddapy` should have code formatting removed and be renamed to API Documentation (or something similarly descriptive) to identify it as the link to the functionality documentation (https://ioos.github.io/erddapy/erddapy.html)

- [ ] **Examples** for all user-facing functions

    - The class definition in erddapy.py contains an example, but the function docstrings do not. There are good examples that could be mirrored from the README and the docs.

    - [ ] ISSUE: Add examples to docstrings for user-facing functions
        - The class definition in erddapy.py contains an example, but the function docstrings do not. Code examples can be mirrored from the existing examples in the README and the docs.

- [ ] **Community guidelines** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer`.

    - Currently, there is only a link to the issues pages to submit bugs, etc. Package maintainer should be explicit about whether they are interested in contributions. If package maintainer is open to contributions, then the README or CONTRIBUTING should include contribution guidelines.

    - [ ] ISSUE: Add explicit statement about contributions to README
        -  README should contain an explicit statement about whether package maintainer is interested in contributions. If package maintainer is open to contributions, then the README or CONTRIBUTING needs to be modified to include contribution guidelines.

#### For packages co-submitting to JOSS

    - Response
        - Note that we could not find any installation instructions for the package. We tried installing with `pip` using `pip install erddapy` and this did work. In addition, the requirements needed to run the examples are often difficult to install (for example, cartopy is fairly difficult to install using pip).

    - [ ] ISSUE: Add installation instructions
        - There are no installation instructions for erddapy in the documentation or in the README. A quick section called "Installation" that had the proper `pip install` command would be helpful, as well as a command for installing in development mode.
        - In addition, some packages are fairly difficult to install with pip (specifically, cartopy and iris, any non-python geoscience package such as gdal which is a dependency for many geospatial Python packages). There should be instructions that will work "out of the box" on multiple platforms (or whichever are listed as supported).

    - Yes - jlpalomino was able to connect to erddapy servers and retrieve data using the example provided in README. Chris wasn't able to get the environment properly-installed, so couldn't confirm that the functionality of the package worked as advertised.

- ~~[ ] **Performance:** Any performance claims of the software been confirmed.~~
    - N/A

    - In general, the tests look to be in good shape with one hiccup. We tried running the tests locally and ran into a little problem - `pytz` is needed but not listed in `requirements-dev.txt`. After that, tests ran successfully locally.

    - [ ] ISSUE: Add `pytz` to requirements-dev.txt

- [ ] **Continuous Integration:** Has continuous integration, such as Travis CI, AppVeyor, CircleCI, and/or others.
    - Travis seems to be running the test suite and the badge shows that all looks well. Code coverage seems to be configured, but it's unclear where the coverage reports exist. This should be documented as well.

    - [ ] ISSUE: Add a badge for code coverage in README and doc
        - Currently, there isn't any indication of the code coverage, which makes it hard to gauge the effectiveness of the tests. It would be great to use a service such as codecov and display as a badge in the readme/docs.

- [ ] **Packaging guidelines**: The package conforms to the pyOpenSci [packaging guidelines]((../packaging/packaging_guide).

    - There's no extra stuff to cover here that hasn't been covered above.

Estimated hours spent reviewing: 2-3 for each reviewer

---

erddapy is a well-scoped tool that has a lot of great functionality. The documentation for the tool has good content, but could be improved with some re-structuring to make content easier to find and follow. There were some basic pieces of information missing that should be added (such as an "installation" section) and standardized across the README and the documentation for the package.
