### 2019-05-20 #1 @lwasser
ok wonderful. thank you @ocefpaf i hope your 15 days away were fun rather than work!! i understanding the catch up game. Please get in touch if you need anything to begin to respond to the reviewers or from pyopensci in general!
