### 2019-05-23 8faa663
add install instructions
