### 2019-05-23 #1 @ocefpaf
>  I hope your 15 days away were fun rather than work!

Both actually. It was fun work.

Here are the answers for the review: https://github.com/pyOpenSci/software-review/issues/1#issuecomment-488816307. I'll respond to https://github.com/pyOpenSci/software-review/issues/1#issuecomment-487607456 later in separately.

> ## Package Review
> _Please check off boxes as applicable, and elaborate in comments below. Your review is not limited to these topics, as described in the reviewer guide_
>
> * [x]  As the reviewer I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).
>
> #### Documentation
> The package includes all the following forms of documentation:
>
> * [ ]  **A statement of need** clearly stating problems the software is designed to solve and its target audience in README
>
>   * Both the README and the docs should be expanded to provide more details for users new to ERDDAP. The README can be improved with additional details on ERDDAP such as the full name for ERDDAP and a brief description of what it is for unfamiliar users. This information could be copied from the docs ([ioos.github.io/erddapy](https://ioos.github.io/erddapy/)), which contains a great first sentence briefly introducing ERDDAP. The docs can be improved with an explicit hyperlink to the ERDDAP site from the text "Environmental Research Division’s Data Access Program (ERDDAP)".
>   * [ ]  ISSUE: Add details on ERDDAP to README
>     * In the README, the ERDDAP acronym should be spelled out as "Environmental Research Division’s Data Access Program (ERDDAP)" and should have an explicit hyperlink to ERDDAP site. In addition, the README should contain a brief description of ERDDAP and an explicit link to the docs for the users that want to learn more information about both ERDDAP and ERDDAPY (similar to first statement provided on docs landing page).

There is not a full, ERDDAP is not an acronym. It is just ERDDAP.
Regarding the details I believe that this part of the README has what you are
saying (same info from the docs that you reference):

```
Easier access to scientific data.

erddapy takes advantage of ERDDAP's RESTful web services and creates the ERDDAP URL for any request, like searching for datasets, acquiring metadata, downloading the data, etc.

What is ERDDAP? ERDDAP unifies the different types of data servers and offers a consistent way to get the data in multiple the formats.
```

Not sure how that can be improved without sending users to ERDDAP's documentation.
(Link added in https://github.com/ioos/erddapy/pull/78)

BTW I don't think users will find erddapy without knowing what ERDDAP is.
It is a wrapper package and the nature of wrapper packages is that
they are tied to the hip with the package they are wrapping.

>   * [ ]  ISSUE: Add hyperlink for ERDDAP to docs
>     * On the docs landing page, the text "Environmental Research Division’s Data Access Program (ERDDAP)" in the docs should also be expliclty hyperlinked to ERDDAP site ([coastwatch.pfeg.noaa.gov/erddap/index.html](https://coastwatch.pfeg.noaa.gov/erddap/index.html)).

I guess the reviewer was confused by the information about ERDDAP online.
The one that the reviewer found is wrong and/or outdated.
Here is what the author has to say about the name:

```
What does the acronym "ERDDAP" stand for?
"ERDDAP" used to be an acronym, but it outgrew that original description. Now, please just think of it as a name, not an acronym.
```

The correct link is already in the docs.
See The link is in: https://github.com/ioos/erddapy/blob/master/docs/source/index.rst#L31

> * [ ]  **Installation instructions:** for the development version of package and any non-standard dependencies in README
>
>   * No installation instructions for package are provided. jlpalomino successfully installed package with `pip` using `pip install erddapy` and was able to run the example in README (i.e. connect to ERDDAP server and get data). In addition, we were able to set up the development environment using the packages listed in requirements-dev.txt, but had to add `pytz` which is not listed in requirements-dev.txt. Chris had difficulty properly installing the environment needed to run erddapy.
>   * [ ]  ISSUE: Add installation instructions to README and docs
>
>     * There are no installation instructions for erddapy in the docs or in the README. A short section in the README and docs called "Installation" with the proper `pip install` command would be helpful as well as a command for installing in development mode. It would also be helpful to provide the necessary configuration files to re-create the environment for erddapy in Binder.

Fixed in https://github.com/ioos/erddapy/pull/79.

>   * [ ]  ISSUE: Add `pytz` to `requirements-dev.txt`.
>
>     * `pytz` needs to be added to `requirements-dev.txt` for tests to run successfully

`pytz` is a dependency of a dependency (`iris`), no need to add it there.
I guess that the issue was with the `pip` installation of `iris`.

> * [ ]  **Vignette(s)** demonstrating major functionality that runs successfully locally
>
>   * The README contains a good start of an example and it runs successfully locally. The example could be improved with comments explaining the code/workflow, some description of the purpose of the example (e.g. the kind of data that it is querying), and displaying the output of the code example. In addition, an example could be provided to show how to get list of the shortcuts available servers (as shown in the docs) and a brief description of what each one contains.
>   * The docs contain more detailed vignettes but they are not easy to find, as most are located in the Longer Introduction section of the docs, which is quite long. The docs could be improved by renaming the Quick Introduction to Quick Start and by breaking up the various code sections in the Longer Introduction into separate vignettes with specific titles, similar to the existing examples like OPeNDAP response.
>   * [ ]  ISSUE: Expand example code in README
>
>     * The existing example code in README can be improved with:
>
>       * [ ]  comments explaining code/workflow
>       * [ ]  brief description of purpose of the example (e.g. the kind of data that it is querying)
>       * [ ]  display of the output of the code
>     * Add a new code example to show how to get list of the shortcuts available servers (as shown in the docs) and a brief description of what each one contains.

I do not grasp the concept of Vignette in README and I do not want to make the README more verbose. We have docs for that. I can expand more on this in a call but, IMO, READMEs should have a very quick intro/example or no example at all, and links and references to the docs.

>   * [ ]  ISSUE: Rename and restructure Quick Introduction and Longer Introduction in the docs
>
>     * The docs can be improved by renaming Quick Introduction to Quick Start and by breaking up the various code sections in the Longer Introduction into separate vignettes with specific titles, similar to the existing examples like OPeNDAP response.

It is not clear to me how that would improve the docs.
It is meant to be a single workflow, that users of ERDDAP usually do,
and separating into sections may actually make it confusing.

The OPeNDAP repose is separate b/c it does not belong in that workflow.

> * [ ]  **Function Documentation:** for all user-facing functions
>
>   * User-facing functions contain appropriate docstrings in erddapy.py, though utilities.py is less documented (note that many functions in utilities.py are not user-facing). The docs for erddapy ([ioos.github.io/erddapy](https://ioos.github.io/erddapy/)) also provide solid documentation for more functions, with the exception of the `to_` functions which have minimal information provided other than links to source code. At a minimum, it would be helpful to provide a list of keyword arguments for each.
>   * Finding the link to erddapy functionality in the docs was not obvious, as it was listed as erddapy above Quick introduction in the table of contents and appeared as a subtitle with code formatting, rather than a link ([ioos.github.io/erddapy/erddapy.html](https://ioos.github.io/erddapy/erddapy.html)).
>   * [ ]  ISSUE: Add docstrings to user-facing functions in utilities.py
>
>     * User-facing functions in utilities.py should contain apprpriopriate docstrings (e.g. parse_dates).

All users facing functions in `utilities.py` have docstrings. Also, the docs you mention are autogenerated from those docstrings. I'm confused with your comment above.

>   * [ ]  ISSUE: Expand documentation for `to_` functions in `erddapy`
>
>     * The documentation for `to_iris`, `to_pandas`, and `to_xarray` should be expanded to contain a list of keyword arguments for each (e.g. rather than "Accepts any iris.load_raw keyword arguments.")

Thanks for the suggestion but that does not belong here. The docstrings mention the upstream documentation and copying that here will make the `to_` functions unnecessarily verbose.

>   * [ ]  ISSUE: Rename and remove code formatting for `erddapy` in docs table of contents
>
>     * `erddapy` should have code formatting removed and be renamed to API Documentation (or something similarly descriptive) to identify it as the link to the functionality documentation ([ioos.github.io/erddapy/erddapy.html](https://ioos.github.io/erddapy/erddapy.html))

I do not follow this comment. That is the documentation for the `erddapy.py` module and has the right name. Also, the code formatting there standard for python documentation and the theme used.

> * [ ]  **Examples** for all user-facing functions
>
>   * The class definition in erddapy.py contains an example, but the function docstrings do not. There are good examples that could be mirrored from the README and the docs.
>   * [ ]  ISSUE: Add examples to docstrings for user-facing functions
>
>     * The class definition in erddapy.py contains an example, but the function docstrings do not. Code examples can be mirrored from the existing examples in the README and the docs.

I don't see the point of that. Functions docstrings should be concise, classes and docs with examples. The Python machinery for investigating the docs online or interactive take advantage of that and repeating information can bloat the code base with stale and out of sync examples.

> * [ ]  **Community guidelines** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer`.
>
>   * Currently, there is only a link to the issues pages to submit bugs, etc. Package maintainer should be explicit about whether they are interested in contributions. If package maintainer is open to contributions, then the README or CONTRIBUTING should include contribution guidelines.
>   * [ ]  ISSUE: Add explicit statement about contributions to README
>
>     * README should contain an explicit statement about whether package maintainer is interested in contributions. If package maintainer is open to contributions, then the README or CONTRIBUTING needs to be modified to include contribution guidelines.

The code is OSS and the code if hosted on GitHub, so I don't think we should explicitly say we are open to contributions when that is obvious IMO.
Regarding a CONTRIBUTING file, NOAA has a policy on that and I'll check with them how we can added it here.

> #### Functionality
> * [ ]  **Installation:** Installation succeeds as documented.
>
>   * Response
>
>     * Note that we could not find any installation instructions for the package. We tried installing with `pip` using `pip install erddapy` and this did work. In addition, the requirements needed to run the examples are often difficult to install (for example, cartopy is fairly difficult to install using pip).
>     * to run example code and tests, jlpalomino was able to install erddapy, pendulum, pytz, and xarray in a conda environment as defined in [earthlab/earth-analytics-python-env:environment.yml@`master`](https://github.com/earthlab/earth-analytics-python-env/blob/master/environment.yml) [](/earthlab/earth-analytics-python-env/blob/HEAD@{2019-05-02T20:16:20Z}/environment.yml)
>   * [ ]  ISSUE: Add installation instructions
>
>     * There are no installation instructions for erddapy in the documentation or in the README. A quick section called "Installation" that had the proper `pip install` command would be helpful, as well as a command for installing in development mode.
>     * In addition, some packages are fairly difficult to install with pip (specifically, cartopy and iris, any non-python geoscience package such as gdal which is a dependency for many geospatial Python packages). There should be instructions that will work "out of the box" on multiple platforms (or whichever are listed as supported).

Fixed in https://github.com/ioos/erddapy/pull/79.
I do not want to add a development more installation b/c devs usually have different opinions on how to do that and I prefer to let them use the tools they are more comfortable with.

The reviewer had some trouble with the `iris` name being different on `PyPI` vs the original package name. I'll add a note on that but, b/c that is development, I will not add in the README. (The `iris` authors are trying to get the name on PyPI and that problem should go away soon.)

Regarding the platforms supported. I think that is beyond the scope of this library and any possible dev guide we add. All those libraries are `pip` and `conda` install-able with different degrees of difficult and devs choose their tools based on their preferences. Adding such guides we are responsible for maintaining workflows that may go stable quickly. Ideally we should let the best practices for package dev installation out of the libraries.

> * [ ]  **Functionality:** Any functional claims of the software been confirmed.
>
>   * Yes - jlpalomino was able to connect to erddapy servers and retrieve data using the example provided in README. Chris wasn't able to get the environment properly-installed, so couldn't confirm that the functionality of the package worked as advertised.
> * ~[ ] **Performance:** Any performance claims of the software been confirmed.~
>
>   * N/A
> * [ ]  **Automated tests:** Tests cover essential functions of the package and a reasonable range of inputs and conditions. All tests pass on the local machine.
>
>   * In general, the tests look to be in good shape with one hiccup. We tried running the tests locally and ran into a little problem - `pytz` is needed but not listed in `requirements-dev.txt`. After that, tests ran successfully locally.
>   * [ ]  ISSUE: Add `pytz` to requirements-dev.txt

Should be addressed with the iris name disambiguation.
There is no need to add it.
See https://github.com/ioos/erddapy/pull/80

> * [ ]  **Continuous Integration:** Has continuous integration, such as Travis CI, AppVeyor, CircleCI, and/or others.
>
>   * Travis seems to be running the test suite and the badge shows that all looks well. Code coverage seems to be configured, but it's unclear where the coverage reports exist. This should be documented as well.
>   * [ ]  ISSUE: Add a badge for code coverage in README and doc
>
>     * Currently, there isn't any indication of the code coverage, which makes it hard to gauge the effectiveness of the tests. It would be great to use a service such as codecov and display as a badge in the readme/docs.

Is that mandatory for a PyOpenSci project? If not I prefer to not add it.
Code coverage can be helpful for a developer locally but it is a bogus code quality in general. It should be used only by the developers to gauge where the flaws in the docs are, otherwise it is easy to fake a 100% code coverage and pretend the tests are in good health.

> * [ ]  **Packaging guidelines**: The package conforms to the pyOpenSci [packaging guidelines]((../packaging/packaging_guide).
>
>   * There's no extra stuff to cover here that hasn't been covered above.

Thanks to the reviewers. This was an interesting exercise. I believe I have to read the review guidelines and I would love to get some feedback later. Hope to see you all in the next PyOpenSci call!
