### 2019-05-27 #1 @ocefpaf
Responses for review https://github.com/pyOpenSci/software-review/issues/1#issuecomment-487607456:

> ## Package Review
> _Please check off boxes as applicable, and elaborate in comments below. Your review is not limited to these topics, as described in the reviewer guide_
>
> * [x]  As the reviewer I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).
>
> #### Documentation
> The package includes all the following forms of documentation:
>
> * [ ]  **A statement of need** clearly stating problems the software is designed to solve and its target audience in README
>
> No - the README could be clearer on the target audience for erdapy (e.g. who are the types of researchers most likely to use ERDDAP) and what ERDDAP is (link to ERDDAP info and summarise what it's used for). Much of this information is on the erddapy website, so perhaps summarise and then link?

I believe the present summary is enough.
The nature of wrapper libraries is that the target users already know the library being wrapped. If that is not true there is a link to the ERDDAP page to send them to the docs.

> * [ ]  **Installation instructions:** for the development version of package and any non-standard dependencies in README
>
> No - I expected to see short explicit installation instructions, i.e. `python setup.py install --user` or `pip install erddapy` (to install from PyPI). Also, see later for problematic dependencies.

Done. The PRs fixing that are referenced in the other review.

> * [ ]  **Vignette(s)** demonstrating major functionality that runs successfully locally
>
> Partially - there is a good example in the README, but I would have liked to have had it also show what the expected outputs would be (e.g. by showing the first five rows of the data frame produced).
>
> This kind of information is on the [website](https://ioos.github.io/erddapy/quick_intro.html#Longer-introduction) but it isn't linked directly from the README, and I think it's important in the README to show the expected outputs of the example, so the example should go as far as showing a sensible output.

I disagree with that. It will bloat the README and it is beyond the scope. That is why we have the docs. Also, the expected output varies a lot depending on the data queried, so there is no expected output but only an expected type.

> * [ ]  **Function Documentation:** for all user-facing functions
>
> Unclear - documentation appears to be being built, but I can't find it (this might be a problem on my system).
> * [ ]  **Examples** for all user-facing functions
>
> They are on the [website](https://ioos.github.io/erddapy/quick_intro.html#Longer-introduction), but should be better linked from the README.

All functions have docstrings and the docs are build with sphinx in the doc page mention.
They do not belong in the README.

> * [ ]  **Community guidelines** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer`.
>
> No -submitting bug reports are covered but it is not clear how others could (or should) contribute to erddapy. This could just be a simple "get in touch via email" or "submit a pull request".

Does PyOpenSci have templates for this? I'll check later and add some but I need to clear with NOAA first b/c they have some policies on this.

> #### Functionality
> * [ ]  **Installation:** Installation succeeds as documented.
>
> Unable to confirm as it's not documented. However appears to work using `python setup.py install --user` but not `pip install erddapy`.

Added a note about this for `pip` users in https://github.com/ioos/erddapy/pull/80

> * [x]  **Functionality:** Any functional claims of the software been confirmed.
>
> Yes, able to connect to the test ERDDAP server and retrieve information.
>
> * [x]  **Performance:** Any performance claims of the software been confirmed.
>
> Not applicable.
>
> * [x]  **Automated tests:** Tests cover essential functions of the package and a reasonable range of inputs and conditions. All tests pass on the local machine.
>
> Yes, but please note the following couple of things that need to be addressed:
>
> * Tests required installation of iris, but this seems to be an ambiguous requirement. See: [pypi.org/project/iris](https://pypi.org/project/iris/)
> * pytzdata is required for tests, but was not in the requirements-dev.txt.

Also addressed in https://github.com/ioos/erddapy/pull/80 both have to do with `iris` name on PyPI.

> * [x]  **Continuous Integration:** Has continuous integration, such as Travis CI, AppVeyor, CircleCI, and/or others.
>
> Yes - building on Travis CI.
>
> * [ ]  **Packaging guidelines**: The package conforms to the pyOpenSci [packaging guidelines]((../packaging/packaging_guide).
>
> Summary of things missing to conform to pyOpenSci packaging guidelines:
>
> * Explicit links to further documentation in README [good]
> * Installation instructions [good]
> * Links to vignettes in documentation [better/best]
> * Improved demonstration usage [better/best]
> * Citation information (there as badge, but could be explicit in readme) [better/best]
