### 2019-05-28 #1 @lwasser
thank you @ocefpaf for the detailed responses to the reviewers @npch , @choldgraf and @jlpalomino
I will have a look at this in more detail in a bit. it looks like we have some questions with respect to
1. expectations how we document functions, etc
2. expectations about contributing and associated documentation
3. documentation in general - what should go in a readme vs actual docs . (filipe - i'm with you in the "keep the readme simple" comments but do believe in well crafted docs) let's talk more!!

More in a bit - i just wanted to acknowledge and thank you all for the effort put into this review so far.
