### 2019-06-02 #1 @ocefpaf
> * [ ]  as someone who isn't familiar with ERDDAP I am not sure what data it can access. Thus it could be super useful to me as i work with different types of data, but there is no statement that clarifies what ERDDAP is. Imagine ERDDAPY in our list of pyopensci packages. I'd love for someone to better understand what exactly this one does even if they aren't familiar with ERDDAPY. Words like gridded and tabular and noaa and maybe coastal (are all of the data coastal and are they global or US focused??). this would greatly clarify and you could add this all in one sentence i think or maybe two. Could you consider a statement in the readme like (please do edit as you see fit):
>
> ```
>  ERDDAPY makes it easier to access NOAA gridded and tabular scientific datasets in common file formats from Python.
>
> erddapy takes advantage of ERDDAP's RESTful web services and creates the ERDDAP URL for any request, like searching for datasets, acquiring metadata, downloading the data, etc.
>
> What is ERDDAP? ERDDAP unifies the different types of data servers and offers a consistent way to get the data in multiple the formats. For more information on ERDDAP servers please see https://coastwatch.pfeg.noaa.gov/erddap/index.html
> ```
>
> Here are my two cents: i think users will find edrappy now via pyopensci. and you'll get new users but only if they know what it does! :) and gosh it's close - just a few more clarifying words will go a long way.

The problem is that ERDDAP is a generic scientific data server and any attempt of describing the datesets in it will come short. The description you got above is of one ERDDAP instance.
I could create another one with sound data from whales and then the description would be wrong. That is why I prefer the generic (and I know vague) "Easier access to scientific data." I asked the author of ERDDAP about that and he agrees that describing the type of datasets are not desirable but he recommends to add something about the various data sources/outputs though (csv, netCDF, hdf, JSON, wav, etc).

> * [ ]  i see your documentation is there with autodoc enabled! it's great. but how do i find it? when i go to the docs, i don't see it on the side bar.
>
> <img alt="Screen Shot 2019-05-30 at 1 46 22 PM" width="327" src="https://user-images.githubusercontent.com/7649194/58660594-07b4fe80-82e3-11e9-84cb-6a441a4de971.png">
>
> i am used to seeing API documentation on the side bar like this on the sidebar. it would be great if something like that existed in the docs as i wasn't sure how to find it either but did see the effort put into docstring code. This would make it easier for users to find the documentation which has great examples. To address neil's comment i wonder if you could also link to the "full api documentation" in the readme below the getting started example. just a thought Filipe.
>
> <img alt="Screen Shot 2019-05-30 at 1 59 57 PM" width="309" src="https://user-images.githubusercontent.com/7649194/58660686-3d59e780-82e3-11e9-9ada-f4ca248ffd7c.png">

The theme I'm using does not add it to the side-bar. It is here:

![Screenshot from 2019-06-02 17-27-40](https://user-images.githubusercontent.com/950575/58767135-97fe7800-855d-11e9-8844-55686d35a9eb.png)

> * [ ]  Dev setup -- iris
>   @ocefpaf just a question here. i'm used to a suggestion of
>   `pip install requirements-dev.txt` so providing one working example of dev setup but then if people want to use conda or whatever else let them do that. i'm just curious if you could consider something like that to address the iris issue or if a user will ALWAYS have to edit the dev-requirements file based on how they install? it is a bit confusing to me but gosh so is the entire world of channels and package versions and such as things get mixed constantly it seems :) or i'm always battling it anyway.

It is a development dependency, so user are not forced to use anything besides `pip` if they do not want to. Also, even as a development dependency, unless someone is dealing with the `to_iris` method they do not need `iris` installed. I have it there just to test that functionality that method.

Last but not least I will answer with this issue: https://github.com/SciTools/iris/issues/3321

Pretty much iris is impossible to install with `pip`!

> it looks like your default is conda `conda create --name TEST python=$PY --file requirements.txt --file requirements-dev.txt` so maybe you could just provide those instructions and tell people to modify the file if they use something other than conda?

That is not necessary for uses. I'm looking into it for a development guide.

> * [ ]  quick start vs short introduction -- my two cents again - [ioos.github.io/erddapy/quick_intro.html#](https://ioos.github.io/erddapy/quick_intro.html#) i am used to seeing words like quickstart for code vignettes that get things started. when i read introduction i think i'm going to get text about your package. should we have a quick discussion in discourse about this to see what others think? i think the comments are just getting as usability. I do think the usability could be improved a bit ... if you are up for it and if others agree! but not something to hang up a review around. with that said if you named the vignettes in your docs quick start or vignette or something you could just provide a link in the readme and be done with it! then things would be clear.

If PyOpenSci has strong opinions about that I can change but I can't really see how the name of the section can be confusing to the understanding of the docs.

> * [ ]  @npch i'm also confused by this statement
>
> ```
> erddapy should have code formatting removed and be renamed to API Documentation (or something similarly descriptive) to identify it as the link to the functionality documentation (ioos.github.io/erddapy/erddapy.html)
> ```
>
> but maybe i follow... would adding a link to API DOCUMENTATION in the readme and to the docs left side panel address the issue? if not can you kindly clarify?
>
> * [ ]  Community guidelines comment
>   I'd like to make a suggestion. i do understand that OSS on github should by default welcome contributions BUT... let's not assume people understand that is the case! i've seen SO many scientists and students who don't understand this concept. I'd like to see ALL of our pyopensci projects have clear contribution guidelines (whatever they are!). @ocefpaf are you open to that? i hear your point but think it's good to be explicit where we can. and we can be concise about it too.

I guess that as long as PyOpenSci provides a template for those I do not have a problem requiring them for package submissions. What I would like to avoid is: you are missing a CONTRIBUTING.md please add one. And then, let's review your CONTRIBUTING.md...

> ```
> The code is OSS and the code if hosted on GitHub, so I don't think we should explicitly say we are open to contributions when that is obvious IMO.
> Regarding a CONTRIBUTING file, NOAA has a policy on that and I'll check with them how we can added it here.
> ```
>
> ok...i'll look at the next review later but i wanted to address this first review first. @ocefpaf can you kindly see my above comments and tell me if they are things you're willing to consider ? i'm trying to find a middle ground between the reviews and pyopensci goals. But also there are a few topics that we probably should discuss more outside of this review!!
>
> Also as you respond, please note our readme guidelines as we can change things if we need to! i think we could add this checklist to the review template too. would that be useful @npch @choldgraf (not pinging jenny as she is out for two weeks!!)

I think the README below is OK.

> ```
> README
> All packages should have a README.md in their root directory. The README should include, from top to bottom:
>
> The package name
> Badges for continuous integration and test coverage, the badge for pyOpenSci peer-review once it has started (see below), a repostatus.org badge, and any other badges. If the README has many more badges, you might want to consider using a table for badges, see this example, that one and that one. Such a table should be more wide than high.
> Short description of goals of package, with descriptive links to all vignettes (rendered, i.e. readable, cf the documentation website section) unless the package is small and there’s only one vignette repeating the README.
> Installation instructions
> Any additional setup required (authentication tokens, etc)
> Brief demonstration usage
> Direction to more detailed documentation (e.g. your documentation files or website).
> If applicable, how the package compares to other similar packages and/or how it relates to other packages
> Citation information
> ```
