### 2019-06-19 #1 @lwasser
ok all... I need to get this moving. the reason why i've been stalling is that not all reviewer comments were implemented and i'm not sure what the best way forward is to ensure everyone feels satisfied. @choldgraf @npch @ocefpaf  should we have a discussion over in the discourse forum on

1. how much documentation do we require and what types of feedback do we require authors to implement
2. how we create usability requirements in a way that are implementable but not overly intensive to apply
3. what should go in the readme and what we enforce.

it would be good to get this package through the review process i just would like us to make a few decisions on how we make decisions on such things.
