### 2019-06-20 #1 @lwasser
ok all -- we are going to do a reboot of this submission. @ocefpaf today we had a good conversation around package usability and what pyopensci wants to see from packages. There was general agreement that we should move towards packages that are more usable to a larger audience of people. this includes things like clear documentation , readmes, etc etc.

We also agreed that this issue is very large and hard to keep tabs on at this point.

Are you ok with @jlpalomino  and @choldgraf revisiting this issue,, opened a second time so we can start fresh with the new review template that allows for issues to be opened and PR's (if you are open to that). we will then have a very organized list of items for you to work on which will align well with the review that you submitted for martin. does that work for you?
