### 2019-06-21 #1 @choldgraf
I can take another pass from a fresh issue and a new template if that's helpful - I must admit that I have been putting off responding on this particular thread because the issue is long and intimidating to go through :-)
