### 2019-06-24 a74c802
run check-manifest
