+++ .travis.yml
@@ -15,6 +15,8 @@ matrix:
       env: PY=3.7
     - name: "coding_standards"
       env: PY=3.7
+    - name: "tarball"
+      env: PY=3.7
     - name: "docs"
       env: PY=3.7
 
@@ -34,12 +36,17 @@ before_install:
     conda info --all
 
 install:
-  - python setup.py sdist && version=$(python setup.py --version) && pushd dist && pip install --no-deps --force-reinstall erddapy-${version}.tar.gz && popd
+  - pip wheel . -w dist --no-deps
 
 script:
   - if [[ $TRAVIS_JOB_NAME == python-* ]]; then
       cp -r tests/ /tmp ;
       pushd /tmp && pytest -n 2 -rxs --cov=erddapy tests && popd ;
+    fi
+
+  - if [[ $TRAVIS_JOB_NAME == 'tarball' ]]; then
+      check-manifest --verbose ;
+      twine check dist/* ;
     fi
 
   - if [[ $TRAVIS_JOB_NAME == 'coding_standards' ]]; then
