### 2019-06-24 #1 @lwasser
@choldgraf we all agree that this issue has gotten massive. my pr in dev_guide begins to address some of the issues here. https://github.com/pyOpenSci/dev_guide/pull/31 if you'd like to resubmit this and start fresh, that would be wonderful. Essentially in our meeting we agreed that usability is core to pyopensci that is captured in the new edits to the review template.
