+++ setup.cfg
@@ -20,3 +20,27 @@ flake8-ignore =
 [metadata]
 description-file = README.md
 license_file = LICENSE.txt
+
+[isort]
+atomic=true
+force_grid_wrap=0
+include_trailing_comma=true
+lines_after_imports=2
+lines_between_types=1
+multi_line_output=3
+not_skip=__init__.py
+use_parentheses=true
+known_first_party=attr
+
+[check-manifest]
+ignore = 
+    .travis.yml
+    .coveragerc
+    Makefile
+    docs
+    docs/*
+    *.enc
+    notebooks
+    notebooks/*
+    tests
+    tests/*
