### 2019-06-24 0decd47
fix install and tarball check
