### 2019-06-24 dab6da5
do not check server URL
