### 2019-06-25 3d01537
publish to PyPI when bulding the docs
