### 2019-07-15 #1 @lwasser
hey team!! it's been a while since we have revisited this review. Let's try again using the new template and via submitting issues / PR's as it makes sense. @choldgraf @jlpalomino are you game to kickstart that process?? thank you in advance!!  and thank you @ocefpaf and all for your patience!!
