### 2019-07-15 #1 @jlpalomino
Sounds good @lwasser. I'll coordinate with @choldgraf to finish a new review using the revised template.
