### 2019-07-25 ioos/erddapy#85 @choldgraf
Include a link to download the notebooks in the documentation
The quick / longer introduction are both quite useful demonstrations of erddapy! It would be helpful to include a link to the notebooks at the top of the example, e.g. something like

"This example is written in a Jupyter Notebook - [click here](https://github.com/ioos/erddapy/blob/master/notebooks/quick_intro.ipynb) to download the notebook so you can run it locally"

(another option would be to use something like [sphinx gallery](https://sphinx-gallery.readthedocs.io/) to generate the examples, but I think that's a longer fix that isn't strictly necessary
