### 2019-07-25 ioos/erddapy#86 @choldgraf
Call the API page "erddapy API" instead of just "erddapy"
I had a hard time figuring out that the documentation page `erddapy` was actually the API page. In the left "contents" menu, I assumed it was a section header rather than a clickable page. I think this would be clearer if it were called "`erddapy` API" here: https://github.com/ioos/erddapy/blob/master/docs/source/erddapy.rst

what do you think?
