### 2019-07-25 ioos/erddapy#87 @choldgraf
Add links to the documentation in the README
There are a few spots (e.g. at the top, in the examples, etc) where you could direct users to your Sphinx documentation for more in-depth usage information. I think this would help direct people who may not know to look for the documentation link in the top of the GitHub page.
