### 2019-07-26 #1 @jlpalomino
- I really like the explanations of what ERDDAP is in the README and documentation - it gives me a better idea for what role the tool plays, and why this package is useful in helping use that tool.
- [x] **Function Documentation:** for all user-facing functions
- [x] **Examples** for all user-facing functions
- [x] **Community guidelines** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer`.

Readme requirements

- [x] Badges for continuous integration and test coverage, the badge for pyOpenSci peer-review once it has started (see below), a repostatus.org badge, and any other badges. If the README has many more badges, you might want to consider using a table for badges, see this example, that one and that one. Such a table should be more wide than high.
- [x] Short description of goals of package, with descriptive links to all vignettes (rendered, i.e. readable, cf the documentation website section) unless the package is small and there’s only one vignette repeating the README.
- [x] Any additional setup required (authentication tokens, etc)
- [x] Brief demonstration usage
- [ ] Direction to more detailed documentation (e.g. your documentation files or website).
- [x] If applicable, how the package compares to other similar packages and/or how it relates to other packages

- [x] **Continuous Integration:** Has continuous integration, such as Travis CI, AppVeyor, CircleCI, and/or others.

#### For packages co-submitting to JOSS

2-3 per reviewer

---

##### Review from choldgraf

In general, this package does a good job of explaining the problem
that it is trying to solve, as well as basic usage for how to get
started with the tool. The installation and example code worked
well for me! The main improvements could be relatively minor things
in the documentation, such as adding more links between the README
and the website documentation. I also think that the documentation
examples could be split into more "how-to" sections rather than a
single all-encompassing tutorial; however, I think those can be
improved in the future and shouldn't be a blocker for this review.
In all, it appears to be a well-designed tool!

- I like the list of badges at the top!
- Recommend adding a link to the notebook in the docs [#85](https://github.com/ioos/erddapy/issues/85) - [name=chris h]
- Recommend calling the API page "erddapy API" [#86](https://github.com/ioos/erddapy/issues/86) - [name=chris h]
- Add links to the documentation in the readme [#87](https://github.com/ioos/erddapy/issues/87) - [name=chris h]
- It's hard to know if there's an example for each user-facing function.
  What is a good way we can ask people to check this without "ctrl-f"ing
  through the whole documentation?
- There's information on how to get in touch now, though still not official
  guidelines for how to get involved or contribute. If the answer to that
  is "I'd prefer to be the one writing the code, and don't plan on creating
  a big community around this project", I think that's fine, it should just
  be mentioned explicitly. That said, I don't think it's a blocker.
- There's no citation information, is that OK?
- There's no JOSS-specific configuration (e.g. no paper.md). We should
  confirm that the author isn't interested in submitting this to JOSS.
- Another thing I noticed: we don't have any explicit review items
  about code review, style, etc. Not sure where we decided on that
  point, but just wanted to point it out.

##### Review from jlpalomino

I agree with my co-reviewer's comments and open issues for this package. erddapy is a well-scoped tool with strong functionality for accessing ERDDAP data. The installation of the package (using both conda and pip) and the short example worked for me. The installation of requirements-dev.txt also worked (using conda), and while not required, it could be a good addition to include instructions for that installation as well, perhaps in contributing guidelines, if the package author is open to that. I agree with my co-reviewer's comments that it should not a considered a block but it may be good to specify to users whether or not the author is interested in contributions, if there will not be contributing guidelines.

The documentation for the tool has strong content, and it would be great to have a link to it directly in the README, as my co-reviewer suggests. The "quick introduction" is solid and easy to follow, though it would be nice to link here again to where the user can find out more information about the servers/datasets that are accessible with erddapy. Similar to my co-reviewer's comments about the documentation, I think the "longer introduction" could be split into smaller how-to" sections to make specific functionality a little easier to follow and easier to find, but it should not be considered a block and can be completed over time. Overall, the tool is fairly well-documented and easy to implement.
