### 2019-07-26 #1 @jlpalomino
@lwasser @ocefpaf Please see above the review for erddapy by choldgraf and myself using the latest template.
