### 2019-08-18 ea0091d
Closes #87
