### 2019-08-18 ioos/erddapy#86 @ocefpaf
> What do you think?

It is not confusing to me b/c I'm used to project that do not use the wording API for the functions/methods docs. (Old scientist that never learned proper programming don't use the word API.) Adding it does not hurt though. See https://github.com/ioos/erddapy/pull/89.
