### 2019-08-18 ioos/erddapy#86 @choldgraf
I think that works!
