### 2019-08-18 8f9a078
add environment.yml for the binder link
