### 2019-08-18 ioos/erddapy#85 @ocefpaf
Done in #90.
