### 2019-08-18 ioos/erddapy#85 @ocefpaf
> (another option would be to use something like [sphinx gallery](https://sphinx-gallery.readthedocs.io/) to generate the examples, but I think that's a longer fix that isn't strictly necessary

Yeah. I'm planning on re-writing the docs and I'll try `sphinx-gallery`. For now let's just add the link.
