### 2019-08-23 #1 @lwasser
@ocefpaf just a friendly ping to see how if you have time to respond to the latest review . Also please note at the [top of this issue](https://github.com/pyOpenSci/software-review/issues/1#issuecomment-479609702) I did confirm that Filipe is NOT interested in the JOSS component of the review at this time @jlpalomino
