+++ setup.cfg
@@ -1,4 +1,6 @@
 [tool:pytest]
+markers =
+    web: marks tests require connection (deselect with '-m "not web"')
 flake8-max-line-length = 105
 flake8-ignore =
     docs/* ALL
