### 2019-10-29 5d304aa
add comment for the server-side functions tests
