### 2019-10-29 c878162
add docstrings
