### 2019-10-29 7b6c607
fix test docstrings
