### 2019-12-03 fc8157e
pass requests_kwargs to _check_url_response [fixes #100]
