+++ .travis.yml
@@ -13,12 +13,14 @@ matrix:
       env: PY=3.6
     - name: "python-3.7"
       env: PY=3.7
+    - name: "python-3.8"
+      env: PY=3.8
     - name: "coding_standards"
-      env: PY=3.7
+      env: PY=3
     - name: "tarball"
-      env: PY=3.7
+      env: PY=3
     - name: "docs"
-      env: PY=3.7
+      env: PY=3
 
 before_install:
   # Install miniconda and create TEST env.
@@ -85,4 +87,4 @@ deploy:
     repo: ioos/erddapy
     tags: true
     all_branches: master
-    condition: '$TEST_TARGET == "docs"'
+    condition: '$TRAVIS_JOB_NAME == "docs"'
