+++ .travis.yml
@@ -78,11 +78,11 @@ doctr:
 deploy:
   skip_cleanup: true
   provider: pypi
-  user: ocefpaf
+  user: "__token__"
   password:
-    secure: "PTeep8OV7UlqsgSHmI4QNjQNlchdGlI32w15SyK1oPrbmvIMt0bWx8+sOHvfIyzEiuKooA/j+N5wjAnsHZF+BPby634IjSj/ctF1Fbw143TF1m7GpsYMzT9zWRcBk/9/Br8q+w4Uv4HUvtWlP0TV+gTMIbNpDt9/lpmhaEAsbz9ceDv3VT3kZUS989vYHlKAdZ3mmasWfd/WEPZgleiLLwT9Vi3zooLw4RiCIGX5yiVtn6Dvde9HChHBw21gvXN1Dsqoc4aFebcCr2zuTaRgcumGka8ZZaj8pw5p25d9Elp5PyYl2fBQCvGa6VlbJGFhDe6lY1HwvLsOBsrNrPIRnpiyU4oNBpW5qfk29amBs379RvlVkN+w39aMcoObsrIIFigC3ggUz9LcFJbTkRqsnWPLyujxquj1kti2NfTkzqbTKHiXT8nzxa8WPrpsG69ay6HV5MjfEUMzAAUbjiUKWoQTkXsGusXiSUZ4tJtl2O3vrHm5YGYGHLqP79sQSGKOKkAg8tY6r7g1WSAhzC8RH/GzclIqtaOds/+K32kT2PAwi14SWZbA+YNHAmblx5+RGhnJiHp67tK3gX+4Oy6M/L7DDQANLrbJBAi/Lg8v3jGA7i5o1xn5yVKXh90ae57RyILokIfue3t+UNCyQ3ENxRfDCZIanC1VqRZg1MnqfQI="
+    secure: "GUI/SjwCFR4tc7lsc9kzr3ZrZYdb8tHvCBmX1YGpjcSRfio73Me39BbDh47gy2L/fSmJAAohRTmb1SskuGPIbWLGd295c44btOqRgohjzsVdLoJ6Pa8+E4T1NXj7GKP12puu4z+u78czPpOBw/XQHaFFiS6xBgeG8sxby6mhvpZeb7d+3LnkecXoGPlTMYvfKR1veltuhTIfqmAdHaIAl2JXGvJGIQNvYTI2WXzKt6XUSQRSIgqLwhn7dAx89e4BKti4XKE5KrhdsO0cUKw9LOIJh8kNAw3p6GPByy/KIs362c/hr+7MDsdr5SSJ5UxsoM0AEqSg3YXwLCTQbAONzC9KwHp0e7XHkUEgaPjMY2gYj61g54iMmbWeblN57LGUzOgAXF3FBSCoQfSu0GeBF3I5dEQIkoQX2zUI93JNbt2lT/gfcRrIuKWOIJ6HysOaGFHSbE3+pGUIi9pxfbimp7G6c30Eg8qs2D7/b++X8BGaHpGSunSkqKIkQv0jo5ND0JTYinLzUBH2VNkUo7t0b1a6LOUo86ZM+w8YdIMV5mXAzalFrA4XOl9/wsKn382EZZo0voGC7Kxd5E2bV8NYylsXDU+3l0zwSxch7WNcLYLIBEHvX3hX8+a8+b7enmMmjyEv/9/Jt5pNtjaa8q+tDQmy4+ExN/TECw3YhCKsPmc="
   distributions: sdist bdist_wheel
-  upload_docs: no
+  upload_docs: false
   on:
     repo: ioos/erddapy
     tags: true
