+++ .travis.yml
@@ -80,7 +80,7 @@ deploy:
   provider: pypi
   user: ocefpaf
   password:
-    secure: "PTeep8OV7UlqsgSHmI4QNjQNlchdGlI32w15SyK1oPrbmvIMt0bWx8+sOHvfIyzEiuKooA/j+N5wjAnsHZF+BPby634IjSj/ctF1Fbw143TF1m7GpsYMzT9zWRcBk/9/Br8q+w4Uv4HUvtWlP0TV+gTMIbNpDt9/lpmhaEAsbz9ceDv3VT3kZUS989vYHlKAdZ3mmasWfd/WEPZgleiLLwT9Vi3zooLw4RiCIGX5yiVtn6Dvde9HChHBw21gvXN1Dsqoc4aFebcCr2zuTaRgcumGka8ZZaj8pw5p25d9Elp5PyYl2fBQCvGa6VlbJGFhDe6lY1HwvLsOBsrNrPIRnpiyU4oNBpW5qfk29amBs379RvlVkN+w39aMcoObsrIIFigC3ggUz9LcFJbTkRqsnWPLyujxquj1kti2NfTkzqbTKHiXT8nzxa8WPrpsG69ay6HV5MjfEUMzAAUbjiUKWoQTkXsGusXiSUZ4tJtl2O3vrHm5YGYGHLqP79sQSGKOKkAg8tY6r7g1WSAhzC8RH/GzclIqtaOds/+K32kT2PAwi14SWZbA+YNHAmblx5+RGhnJiHp67tK3gX+4Oy6M/L7DDQANLrbJBAi/Lg8v3jGA7i5o1xn5yVKXh90ae57RyILokIfue3t+UNCyQ3ENxRfDCZIanC1VqRZg1MnqfQI="
+    secure: "JZZhudzApHIGqGziiejBVqYM3Lr99zkQhiWngDCHb4fMytGgmODvptrArsENIOC3SS3SBJHw9pIaDj9Dyqt+FKh8DLXjPy190u7QIrStlCQ7395IPRORH5+qiHOGk+I9XZ6zb7RMEG2tNwxOeOmphvQ3RxOt00IuqX4C9QTrf9GCzg//++Xd3RNkwMxgacP8+Sz88KKN77u0IDTUH9AnOQi2Y38Tw70zCFE1qggXStKntnB0d7bf4AxMzFvbn6FxmbkjnT/7i8PcUTav3wtOcT69MRLNazOWv9wvGp/dpps7JwIp1+VFOitW2r9S10PqMD5hu6Pzp+EVBYpGy2FU1SJoyvZ0X6hOio+hQN9W/lyGR0Y7xstBM8Nkvt2ve42TSFZEhw++dnfDqVKfZFL1q81NLC243+sOhvWUBKmHeMgY5cBrWCgvajriE3gwSu5hL67xIrOsmvVNbYiTrH3tP39VZdXhdqQY4r3K6kOnxQWwfRB2OCV6XDAk4+OYst0X1uFdBIhquoJoBX1/XaHvNw3VZUz9JljpAPTxpDfAEDb7F4fZV/5NLFbB5TNRi6FacogZFz5BMJuuFlaCYI1gTIa9IQjXrFquGkyEQGIrJXRBp5PyiH5ZtF7uIkiAeaqFdSeUIJDU1WXxaGORJ/8kyZG0HEgLoaeJPBJxM7J7ZuA="
   distributions: sdist bdist_wheel
   upload_docs: no
   on:
