### 2020-02-12 10a9491
versioneer -> setuptools_scmi, and use setup.cfg
