+++ docs/source/conf.py
@@ -59,10 +59,10 @@ author = "Filipe Fernandes"
 # |version| and |release|, also used in various other places throughout the
 # built documents.
 #
-from erddapy._version import get_versions
+import erddapy
 
-version = release = get_versions()["version"]
-del get_versions
+
+version = release = erddapy__version__
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
