+++ .travis.yml
@@ -47,6 +47,7 @@ script:
     fi
 
   - if [[ $TRAVIS_JOB_NAME == 'tarball' ]]; then
+      python setup.py --version ;
       pip wheel . -w dist --no-deps ;
       check-manifest --verbose ;
       twine check dist/* ;
