+++ docs/source/conf.py
@@ -62,7 +62,7 @@ author = "Filipe Fernandes"
 import erddapy
 
 
-version = release = erddapy__version__
+version = release = erddapy.__version__
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
