### 2020-03-03 f22e08d
setup deps in a single place
