+++ setup.cfg
@@ -33,9 +33,6 @@ install_requires =
     requests
 python_requires = >=3.6
 packages = find:
-setup_requires =
-    setuptools >= 41.2
-    setuptools_scm
 
 [sdist]
 formats = gztar
