+++ pyproject.toml
@@ -1,3 +1,3 @@
 [build-system]
-requires = ["setuptools>=34.4", "wheel"]
+requires = ["setuptools>=41.2", "wheel"]
 build-backend = "setuptools.build_meta"
