+++ .travis.yml
@@ -64,7 +64,7 @@ script:
       pushd docs
       make clean html linkcheck
       popd
-      if [[ -z "$TRAVIS_TAG" ]]; then 
+      if [[ -z "$TRAVIS_TAG" ]]; then
         python -m doctr deploy --build-tags --key-path github_deploy_key.enc --built-docs docs/_build/html dev
       else
         python -m doctr deploy --build-tags --key-path github_deploy_key.enc --built-docs docs/_build/html "version-$TRAVIS_TAG"
