+++ setup.cfg
@@ -28,7 +28,7 @@ classifiers =
 
 [options]
 zip_safe = True
-install_requires = 
+install_requires =
     pandas >=0.20.3
     requests
 python_requires = >=3.6
@@ -49,7 +49,7 @@ use_parentheses=true
 known_first_party=attr
 
 [check-manifest]
-ignore = 
+ignore =
     *.yml
     .coveragerc
     Makefile
