### 2020-03-03 003c335
mypy config
