+++ setup.cfg
@@ -62,3 +62,6 @@ ignore =
     notebooks/*
     tests
     tests/*
+
+[mypy]
+ignore_missing_imports = True
