### 2020-03-03 6a0c501
add more type annotations
