+++ tests/test_url_builder.py
@@ -204,9 +204,9 @@ def test_get_var_by_attr(e):
         )
         == []
     )
-    assert e.get_var_by_attr(
-        dataset_id="pmelTao5dayIso", standard_name="time"
-    ) == ["time"]
+    assert e.get_var_by_attr(dataset_id="pmelTao5dayIso", standard_name="time") == [
+        "time"
+    ]
 
 
 # Test URL modifiers (server-side functions).
@@ -214,9 +214,7 @@ def test_download_url_distinct(e):
     """Check download URL results with and without the distinct option."""
     dataset_id = "gtoppAT"
     variables = ["commonName", "yearDeployed", "serialNumber"]
-    no_distinct_url = e.get_download_url(
-        dataset_id=dataset_id, variables=variables
-    )
+    no_distinct_url = e.get_download_url(dataset_id=dataset_id, variables=variables)
     with_distinct_url = e.get_download_url(
         dataset_id=dataset_id, variables=variables, distinct=True
     )
