### 2020-03-03 8f4cf16
flake8 config in a single place
