+++ setup.cfg
@@ -1,11 +1,3 @@
-[tool:pytest]
-markers =
-    web: marks tests require connection (deselect with '-m "not web"')
-flake8-max-line-length = 105
-flake8-ignore =
-    docs/* ALL
-    erddapy/_version.py ALL
-
 [metadata]
 name = erddapy
 description = Python interface for ERDDAP
@@ -65,3 +57,7 @@ ignore =
 
 [mypy]
 ignore_missing_imports = True
+
+[tool:pytest]
+markers =
+    web: marks tests require connection (deselect with '-m "not web"')
