+++ .pre-commit-config.yaml
@@ -10,6 +10,7 @@ repos:
   hooks:
     - id: flake8
       exclude: docs/source/conf.py
+      args: [--max-line-length=105]
 
 - repo: https://github.com/pre-commit/mirrors-isort
   rev: v4.3.21
