### 2020-03-03 b68ae34
mypy and isort configs in a single place
