+++ setup.cfg
@@ -31,17 +31,6 @@ packages = find:
 [sdist]
 formats = gztar
 
-[isort]
-atomic=true
-force_grid_wrap=0
-include_trailing_comma=true
-lines_after_imports=2
-lines_between_types=1
-multi_line_output=3
-not_skip=__init__.py
-use_parentheses=true
-known_first_party=attr
-
 [check-manifest]
 ignore =
     *.yml
@@ -54,9 +43,6 @@ ignore =
     notebooks/*
     tests
     tests/*
-
-[mypy]
-ignore_missing_imports = True
 
 [tool:pytest]
 markers =
