+++ .pre-commit-config.yaml
@@ -16,7 +16,7 @@ repos:
   rev: v4.3.21
   hooks:
   - id: isort
-    args: [--multi-line=3, --trailing-comma, --force-grid-wrap=0, --use-parentheses, --line-width=88]
+    args: [--multi-line=3, --lines-after-imports=2, --lines-between-types=1, --trailing-comma, --force-grid-wrap=0, --use-parentheses, --line-width=88]
 
 - repo: https://github.com/psf/black
   rev: 19.10b0
@@ -29,3 +29,4 @@ repos:
   hooks:
   - id: mypy
     exclude: docs/source/conf.py
+    args: [--ignore-missing-imports]
