+++ .isort.cfg
@@ -1,0 +1,2 @@
+[settings]
+known_third_party = pandas,pendulum,pkg_resources,pytest,pytz,requests,setuptools
