+++ .pre-commit-config.yaml
@@ -4,6 +4,7 @@ repos:
   hooks:
     - id: trailing-whitespace
     - id: check-ast
+    - id: debug-statements
 
 - repo: https://gitlab.com/pycqa/flake8
   rev: 3.7.9
@@ -16,7 +17,13 @@ repos:
   rev: v4.3.21
   hooks:
   - id: isort
+    additional_dependencies: [toml]
     args: [--project=erddapy, --multi-line=3, --lines-after-imports=2, --lines-between-types=1, --trailing-comma, --force-grid-wrap=0, --use-parentheses, --line-width=88]
+
+- repo: https://github.com/asottile/seed-isort-config
+  rev: v1.9.4
+  hooks:
+    - id: seed-isort-config
 
 - repo: https://github.com/psf/black
   rev: 19.10b0
