+++ .travis.yml
@@ -15,8 +15,6 @@ matrix:
       env: PY=3.7
     - name: "python-3.8"
       env: PY=3.8
-    - name: "coding_standards"
-      env: PY=3
     - name: "tarball"
       env: PY=3
     - name: "docs"
@@ -51,10 +49,6 @@ script:
       pip wheel . -w dist --no-deps ;
       check-manifest --verbose ;
       twine check dist/* ;
-    fi
-
-  - if [[ $TRAVIS_JOB_NAME == 'coding_standards' ]]; then
-      pytest --flake8 -m flake8 ;
     fi
 
   - |
