+++ erddapy/erddapy.py
@@ -8,7 +8,7 @@ Pythonic way to access ERDDAP data
 import copy
 import functools
 
-from typing import Dict, Optional
+from typing import Dict, List, Optional, Tuple, Union
 from urllib.parse import quote_plus
 
 import pandas as pd
@@ -23,7 +23,11 @@ from erddapy.utilities import (
 )
 
 
-def _distinct(url: str, **kwargs) -> str:
+ListLike = Union[List[str], Tuple[str]]
+OptionalStr = Optional[str]
+
+
+def _distinct(url: str, **kwargs: Dict) -> str:
     """
     Sort all of the rows in the results table
     (starting with the first requested variable,
@@ -42,7 +46,7 @@ def _distinct(url: str, **kwargs) -> str:
     return url
 
 
-def _order_by(url: str, variables, by: Optional[str] = None) -> str:
+def _order_by(url: str, variables, by: OptionalStr = None) -> str:
     variables = ",".join(variables)
     return f'{url}&orderBy{by}("{variables}")'
 
@@ -55,10 +59,10 @@ class ERDDAP(object):
         protocol: tabledap or griddap.
 
     Attributes:
-        dataset_id (str): a dataset unique id.
-        variables (:obj:`list`/`tuple`): a list variables to download.
+        dataset_id: a dataset unique id.
+        variables: a list variables to download.
         response: default is HTML.
-        constraints (:obj:`dict`): download constraints, default None (opendap-like url)
+        constraints: download constraints, default None (opendap-like url)
         params and requests_kwargs: `request.get` options
 
     Returns:
@@ -109,10 +113,7 @@ class ERDDAP(object):
     """
 
     def __init__(
-        self,
-        server: str,
-        protocol: Optional[str] = None,
-        response: Optional[str] = "html",
+        self, server: str, protocol: OptionalStr = None, response: str = "html",
     ):
         if server in servers.keys():
             server = servers[server].url
@@ -122,30 +123,30 @@ class ERDDAP(object):
 
         # Initialized only via properties.
         self.constraints: Optional[Dict] = None
-        self.dataset_id: Optional[str] = None
+        self.dataset_id: OptionalStr = None
         self.params: Optional[Dict] = None
         self.requests_kwargs: Dict = {}
-        self.variables: Optional[str] = ""
+        self.variables: Optional[ListLike] = None
 
         # Caching the last `dataset_id` and `variables` list request for quicker multiple accesses,
         # will be overridden when requesting a new `dataset_id`.
-        self._dataset_id: Optional[str] = None
+        self._dataset_id: OptionalStr = None
         self._variables: Dict = {}
 
     def get_search_url(
         self,
-        response=None,
-        search_for=None,
-        protocol=None,
-        items_per_page=1000,
-        page=1,
+        response: OptionalStr = None,
+        search_for: OptionalStr = None,
+        protocol: OptionalStr = None,
+        items_per_page: int = 1000,
+        page: int = 1,
         **kwargs,
-    ):
+    ) -> str:
 
         """The search URL for the `server` endpoint provided.
 
         Args:
-            search_for (str): "Google-like" search of the datasets' metadata.
+            search_for: "Google-like" search of the datasets' metadata.
 
                 - Type the words you want to search for, with spaces between the words.
                     ERDDAP will search for the words separately, not as a phrase.
@@ -161,17 +162,17 @@ class ERDDAP(object):
                     to find datasets from a specific website (usually the start of the datasetID),
                     include (for example) `"datasetID=erd"` in your search.
 
-            response (str): default is HTML.
-            items_per_page (int): how many items per page in the return,
+            response: default is HTML.
+            items_per_page: how many items per page in the return,
                 default is 1000.
[…]
