### 2020-03-04 278da73
fetch server list from a canonical source
