+++ erddapy/utilities.py
@@ -12,77 +12,27 @@ from datetime import datetime
 from typing import Dict, Generator, Optional, Union
 from typing.io import BinaryIO
 
+import pandas as pd
 import pytz
 import requests
 
 from pandas.core.tools.datetimes import parse_time_string
 
 
-_server = namedtuple("server", ["description", "url"])
+@functools.lru_cache(maxsize=None)
+def servers_list():
+    # Replace with the master version when
+    # https://github.com/IrishMarineInstitute/awesome-erddap/pull/14 is merged.
+    # url = "https://raw.githubusercontent.com/IrishMarineInstitute/awesome-erddap/master/erddaps.json"
+    url = "https://raw.githubusercontent.com/ocefpaf/awesome-erddap/short_names/erddaps.json"
+    df = pd.read_json(url)
+    _server = namedtuple("server", ["description", "url"])
+    return {
+        row["short_name"]: _server(row["name"], row["url"]) for k, row in df.iterrows()
+    }
 
-servers = {
-    "MDA": _server(
-        "Marine Domain Awareness (MDA) - Italy",
-        "https://bluehub.jrc.ec.europa.eu/erddap/",
-    ),
-    "MII": _server("Marine Institute - Ireland", "https://erddap.marine.ie/erddap/"),
-    "CSCGOM": _server(
-        "CoastWatch Caribbean/Gulf of Mexico Node",
-        "http://cwcgom.aoml.noaa.gov/erddap/",
-    ),
-    "CSWC": _server(
-        "CoastWatch West Coast Node", "https://coastwatch.pfeg.noaa.gov/erddap/",
-    ),
-    "CeNCOOS": _server(
-        "NOAA IOOS CeNCOOS (Central and Northern California Ocean Observing System)",
-        "http://erddap.axiomalaska.com/erddap/",
-    ),
-    "NERACOOS": _server(
-        "NOAA IOOS NERACOOS (Northeastern Regional Association of Coastal and Ocean Observing Systems)",
-        "http://www.neracoos.org/erddap/",
-    ),
-    "NGDAC": _server(
-        "NOAA IOOS NGDAC (National Glider Data Assembly Center)",
-        "https://data.ioos.us/gliders/erddap/",
-    ),
-    "PacIOOS": _server(
-        "NOAA IOOS PacIOOS (Pacific Islands Ocean Observing System) at the University of Hawaii (UH)",
-        "http://oos.soest.hawaii.edu/erddap/",
-    ),
-    "SECOORA": _server(
-        "NOAA IOOS SECOORA (Southeast Coastal Ocean Observing Regional Association)",
-        "http://erddap.secoora.org/erddap/",
-    ),
-    "NCEI": _server(
-        "NOAA NCEI (National Centers for Environmental Information) / NCDDC",
-        "https://ecowatch.ncddc.noaa.gov/erddap/",
-    ),
-    "OSMC": _server(
-        "NOAA OSMC (Observing System Monitoring Center)",
-        "http://osmc.noaa.gov/erddap/",
-    ),
-    "UAF": _server(
-        "NOAA UAF (Unified Access Framework)", "https://upwell.pfeg.noaa.gov/erddap/",
-    ),
-    "ONC": _server("ONC (Ocean Networks Canada)", "http://dap.onc.uvic.ca/erddap/"),
-    "BMLSC": _server(
-        "UC Davis BML (University of California at Davis, Bodega Marine Laboratory)",
-        "http://bmlsc.ucdavis.edu:8080/erddap/",
-    ),
-    "RTECH": _server("R.Tech Engineering", "https://meteo.rtech.fr/erddap/"),
-    "IFREMER": _server(
[…]
