### 2020-05-11 2684df1
split serial and non-serial tests
