+++ setup.cfg
@@ -47,3 +47,4 @@ ignore =
 [tool:pytest]
 markers =
     web: marks tests require connection (deselect with '-m "not web"')
+    serial: marks tests that cannot be run in parallel (deselect with '-m "not serial"')
