### 2020-05-11 15a4cfb
use ncCF response instead of nc in the netCDF file download methods
