### 2020-05-11 64974b8
fix tests for the extras vars
