### 2020-05-11 b70b6ae
workaround when cube cannot be realised
