### 2020-05-12 617abc4
test _check_url_response
