### 2020-05-21 a13ee24
to_ncCF response and in-memory dataset possible
