### 2020-05-25 a543fb0
fix mandatory deps
