+++ requirements.txt
@@ -1,2 +1,3 @@
 pandas >=0.20.3
+pytz
 requests
