### 2020-06-25 ae891c6
version from file and pre-commit fixes
