+++ erddapy/__init__.py
@@ -1,4 +1,6 @@
-import pkg_resources
+"""
+Easier access to scientific data
+"""
 
 from erddapy.erddapy import ERDDAP, servers
 
@@ -6,6 +8,6 @@ from erddapy.erddapy import ERDDAP, servers
 __all__ = ["ERDDAP", "servers"]
 
 try:
-    __version__ = pkg_resources.get_distribution("erddapy").version
-except Exception:
+    from ._version import __version__
+except ImportError:
     __version__ = "unknown"
