### 2020-06-25 cafc840
remove order_by leftover and pre-commit fixes
