+++ .pre-commit-config.yaml
@@ -5,6 +5,12 @@ repos:
     - id: trailing-whitespace
     - id: check-ast
     - id: debug-statements
+    - id: end-of-file-fixer
+    - id: check-docstring-first
+    - id: check-added-large-files
+    - id: requirements-txt-fixer
+    - id: file-contents-sorter
+      files: requirements-dev.txt
 
 - repo: https://gitlab.com/pycqa/flake8
   rev: 3.7.9
@@ -16,9 +22,9 @@ repos:
 - repo: https://github.com/pre-commit/mirrors-isort
   rev: v4.3.21
   hooks:
-  - id: isort
-    additional_dependencies: [toml]
-    args: [--project=erddapy, --multi-line=3, --lines-after-imports=2, --lines-between-types=1, --trailing-comma, --force-grid-wrap=0, --use-parentheses, --line-width=88]
+    - id: isort
+      additional_dependencies: [toml]
+      args: [--project=erddapy, --multi-line=3, --lines-after-imports=2, --lines-between-types=1, --trailing-comma, --force-grid-wrap=0, --use-parentheses, --line-width=88]
 
 - repo: https://github.com/asottile/seed-isort-config
   rev: v1.9.4
@@ -37,3 +43,22 @@ repos:
   - id: mypy
     exclude: docs/source/conf.py
     args: [--ignore-missing-imports]
+
+- repo: https://github.com/keewis/blackdoc
+  rev: stable
+  hooks:
+    - id: blackdoc
+
+- repo: https://github.com/econchick/interrogate
+  rev: 1.2.0
+  hooks:
+    - id: interrogate
+      exclude: ^(docs|setup.py|tests)
+      args: [--config=pyproject.toml]
+
+- repo: https://github.com/codespell-project/codespell
+  rev: v1.16.0
+  hooks:
+    - id: codespell
+      args:
+        - --quiet-level=2
