### 2020-07-06 89b8a4b
Raise for status and 'promote' check URL
