+++ erddapy/utilities.py
@@ -127,3 +127,10 @@ def _tempnc(data: BinaryIO) -> Generator[str, None, None]:
     finally:
         if tmp is not None:
             tmp.close()
+
+
+def show_iframe(src):
+    """Helper function to show HTML returns."""
+    from IPython.display import IFrame
+
+    return IFrame(src=src, width="100%", height=950)
