### 2020-07-06 b345c28
run the notebook before building the docs
