+++ .travis.yml
@@ -65,10 +65,10 @@ script:
       make clean html linkcheck
       popd
       if [[ -z "$TRAVIS_TAG" ]]; then
-        python -m doctr deploy --build-tags --key-path github_deploy_key.enc --built-docs docs/_build/html dev
+        python -m doctr deploy --build-tags --key-path github_deploy_key.enc --built-docs docs/build/html dev
       else
-        python -m doctr deploy --build-tags --key-path github_deploy_key.enc --built-docs docs/_build/html "version-$TRAVIS_TAG"
-        python -m doctr deploy --build-tags --key-path github_deploy_key.enc --built-docs docs/_build/html .
+        python -m doctr deploy --build-tags --key-path github_deploy_key.enc --built-docs docs/build/html "version-$TRAVIS_TAG"
+        python -m doctr deploy --build-tags --key-path github_deploy_key.enc --built-docs docs/build/html .
       fi
     fi
 
