### 2020-07-07 07adb1a
Fix notebook name and copy them instead of moving.
