### 2020-07-12 aa6bd0f
better exception message
