+++ erddapy/utilities.py
@@ -56,7 +56,10 @@ def urlopen(url, auth: Optional[tuple] = None) -> BinaryIO:
 
     """
     response = requests.get(url, allow_redirects=True, auth=auth)
-    response.raise_for_status()
+    try:
+        response.raise_for_status()
+    except requests.exceptions.HTTPError as err:
+        raise requests.exceptions.HTTPError(f"r.content.decode()") from err
     return io.BytesIO(response.content)
 
 
