+++ .travis.yml
@@ -1,10 +1,9 @@
 language: minimal
 
-sudo: false
-
 env:
   global:
-    - secure: "mn3RZC/TCt+zx8HK1LYFplCAWiSE+19mJ6S7JgVRggbpizbSln73/KtTXvLvZbJHwxK1n5m8yK6gFS3gZBpFKdOnQtkkT+yinaEYFV7eX+3kPxv4KX97spIHHMGKl/jIrewMbLBe4Drz8XcAUdWDWA4iFPrA1tVDhbnGTiZ2NEvMdn4Q4QXWfYhOzfJizSLnjqiuiNYI9cFVERf6mD5S5aVqRd2OivKR85FbCZPzAn6gcxl5XKIzvqFhMTffXXhe1vq4/pWuu+2vj5sb36gAKoh8fPW+WvieuAtKY4VqCZ4GPIA7iMLxqDXkGXD2s5KsBPXN0j9f5zyL7KCDqmYgw08gxa9GM+eEdtC2lQLtGpV2mmoau3hG6FevJkD9qraORoQiul4ZmBwCqznqDvE61MkbMnA0XCutLb/bhFlcrUrLRvCkiF5xpm8t6NNWKI5kqP1MQUlRFCxoaUn+xdwAGKbwXdyC+w9HhZR9lNuotfg0XVLCVTeP3GIWIhuZJMiJsfCpm71HnaFXsq7B9J0mq4wEinIBDeNCJxAZN2phQHH5ElBkl+EyDy+8p3czLLSQ/WhfHVW7Ylv3oRNab5YTFFTKTVY5plGx6nbzHQG6YS9wAPTGEeHyPGheLI0/fpZZ/8m4WLBx/VUYAPpkYe42a8xpdbeoz6lx7Xj3jAmo+Lg="
+    # Doctr deploy key for ioos/erddapy
+    - secure: "K7/ZsuH+mp28YhBgnHARRmneKyMqODpMcrzTBd4tK0rqtzRJYP7fel9JGhD/3mIqyp/E2qCF3Rk/YbonHvmsyjh0yBmyI1BtHIKEcsUgwYCMvkjWYTR0k5LYWqtyEl1pW5RSs5PAAKCcXDdxDoXkmLdJViDiKv6Mv8bnfppYJzPH7Z2xjfr9YxvhMiIngfLxtjPAm1GZKO75Ldhw1uE9eMqdxYHrENvTkbXTfS2o5ERCJppd8o8zeEsnUjE7fDXsvBTYP5Cdng9tYMVCFiIubQXLuYUGKjwVObR7AIczwx5so4yTly0k2gNrXtTpTW+hm6D8VQCGfJAXgfq3zpzKtUwBKpXVAP8K+NWy91eXOyr/rgcmFXjlzsN25/gPonyac92a7OQXRs5nsXKBY1KfUR2LyQapVZ89RpoA6w98X7OWeqvUg7uDuCdYAcsM3zVJMqIHvlr9UpqMvuluRMvX4iq9P/fD5Bq1WT3uJ3olLEXpbcaOeeaBpb3vO8IwiOEaNXR7A9iug3cpz4bP5nd5DET8Ki+lew6bkfiC5SaP+emS9WXD4uoKbiwrL6wz1kAdjOH6Exn4R0q8paJuaORiIfDEXAay39ZwBl3qzvt6Y7O3Y8i9oXG59+UKojSXxkPdF4byWA1nUF6wl4bpC4f/oJMoDtXiUdNv6WM8Jcndwdk="
 
 matrix:
   fast_finish: true
@@ -65,16 +64,12 @@ script:
       make clean html linkcheck
       popd
       if [[ -z "$TRAVIS_TAG" ]]; then
[…]
