### 2020-08-01 30ec942
Update docs after building Travis build 315 of ioos/erddapy
The docs were built from the tag 'v0.6.0' against the commit
4c01f56b88ab1f1fd6ab63d5c1c52c66469af729.

The Travis build that generated this commit is at
https://travis-ci.org/ioos/erddapy/jobs/713988009.

The doctr command that was run is

    /home/travis/miniconda/envs/TEST/lib/python3.8/site-packages/doctr/__main__.py deploy --build-tags --key-path github_deploy_key_ioos_erddapy.enc --built-docs docs/build/html .
