+++ _sources/erddapy.rst.txt
@@ -1,0 +1,7 @@
+:mod:`ERDDAP class API`
+-----------------------
+
+.. automodule:: erddapy
+   :members:
+   :undoc-members:
+   :show-inheritance:
