### 2020-08-06 a5a0b48
Re-adds the ability to pass requests_kwargs to underlying request
Adds several tests to make sure that keyword arguments can be passed appropriately
to requests.get, including simulating passing a timeout and querying through
a slow proxy.

Does not work on the removed request params, as that may be mixed in with auth.

Closes #131
