### 2020-08-06 0320040
Changes to tests for isort and flake8
