### 2020-09-18 88b63e0
pre-commit update and use micromamba in the CIs
