+++ .pre-commit-config.yaml
@@ -1,6 +1,6 @@
 repos:
 - repo: https://github.com/pre-commit/pre-commit-hooks
-  rev: v2.5.0
+  rev: v3.1.0
   hooks:
     - id: trailing-whitespace
     - id: check-ast
@@ -21,25 +21,25 @@ repos:
       args: [--max-line-length=105]
 
 - repo: https://github.com/pre-commit/mirrors-isort
-  rev: v4.3.21
+  rev: v5.5.3
   hooks:
     - id: isort
       additional_dependencies: [toml]
       args: [--project=erddapy, --multi-line=3, --lines-after-imports=2, --lines-between-types=1, --trailing-comma, --force-grid-wrap=0, --use-parentheses, --line-width=88]
 
 - repo: https://github.com/asottile/seed-isort-config
-  rev: v1.9.4
+  rev: v2.1.1
   hooks:
     - id: seed-isort-config
 
 - repo: https://github.com/psf/black
-  rev: 19.10b0
+  rev: 20.8b1
   hooks:
   - id: black
     language_version: python3
 
 - repo: https://github.com/pre-commit/mirrors-mypy
-  rev: v0.761
+  rev: v0.782
   hooks:
   - id: mypy
     exclude: docs/source/conf.py
