+++ erddapy/utilities.py
@@ -141,7 +141,7 @@ def _tempnc(data: BinaryIO) -> Generator[str, None, None]:
     tmp = None
     try:
         tmp = NamedTemporaryFile(suffix=".nc", prefix="erddapy_")
-        tmp.write(data)
+        tmp.write(data.getvalue())
         tmp.flush()
         yield tmp.name
     finally:
