+++ .pre-commit-config.yaml
@@ -1,6 +1,6 @@
 repos:
 - repo: https://github.com/pre-commit/pre-commit-hooks
-  rev: v3.1.0
+  rev: v3.2.0
   hooks:
     - id: trailing-whitespace
     - id: check-ast
@@ -14,23 +14,18 @@ repos:
       files: requirements-dev.txt
 
 - repo: https://gitlab.com/pycqa/flake8
-  rev: 3.7.9
+  rev: 3.8.4
   hooks:
     - id: flake8
       exclude: docs/source/conf.py
       args: [--max-line-length=105]
 
-- repo: https://github.com/pre-commit/mirrors-isort
-  rev: v5.5.3
+- repo: https://github.com/pycqa/isort
+  rev: 5.5.4
   hooks:
     - id: isort
       additional_dependencies: [toml]
       args: [--project=erddapy, --multi-line=3, --lines-after-imports=2, --lines-between-types=1, --trailing-comma, --force-grid-wrap=0, --use-parentheses, --line-width=88]
-
-- repo: https://github.com/asottile/seed-isort-config
-  rev: v2.1.1
-  hooks:
-    - id: seed-isort-config
 
 - repo: https://github.com/psf/black
   rev: 20.8b1
@@ -51,14 +46,14 @@ repos:
     - id: blackdoc
 
 - repo: https://github.com/econchick/interrogate
-  rev: 1.2.0
+  rev: 1.3.1
   hooks:
     - id: interrogate
       exclude: ^(docs|setup.py|tests)
       args: [--config=pyproject.toml]
 
 - repo: https://github.com/codespell-project/codespell
-  rev: v1.16.0
+  rev: v1.17.1
   hooks:
     - id: codespell
       args:
