### 2020-10-04 1240f46
ship a server list for offline usage
