+++ .travis.yml
@@ -43,11 +43,13 @@ script:
       popd
     fi
 
-  - if [[ $TRAVIS_JOB_NAME == 'tarball' ]]; then
-      python setup.py --version ;
-      pip wheel . -w dist --no-deps ;
-      check-manifest --verbose ;
-      twine check dist/* ;
+  - |
+    if [[ $TRAVIS_JOB_NAME == 'tarball' ]]; then
+      python setup.py --version
+      python setup.py sdist
+      pip wheel . -w dist --no-deps
+      check-manifest --verbose
+      twine check dist/*
     fi
 
   - |
