+++ setup.cfg
@@ -22,6 +22,7 @@ classifiers =
 
 [options]
 zip_safe = True
+include_package_data = True
 install_requires =
     pandas >=0.20.3
     requests
