### 2020-10-04 1e5803b
test _nc_dataset()
