+++ tests/test_utilities.py
@@ -11,6 +11,7 @@ from requests.exceptions import HTTPError, ReadTimeout
 
 from erddapy.utilities import (
     _clean_response,
+    _nc_dataset,
     _tempnc,
     check_url_response,
     parse_dates,
@@ -146,7 +147,7 @@ def test_quote_string_constraints():
             assert v.startswith('"') and v.endswith('"')
 
 
-@pytest.mark.serial
+@pytest.mark.web
 def test__tempnc():
     url = "https://data.ioos.us/gliders/erddap/tabledap/cp_336-20170116T1254.nc"
     data = urlopen(url).read()
@@ -157,3 +158,19 @@ def test__tempnc():
         assert tmp.endswith("nc")
     # Check that the file was removed.
     assert not os.path.exists(tmp)
+
+
+@pytest.mark.web
+def test__nc_dataset():
+    """
+    FIXME: we need to test both in-memory and local file.
+    That can be achieve with a different libnetcdf but having two environments for testing is cumbersome.
+    However, it turns out sometimes a server can fail to provide files can be loaded in memory (#137).
+    If we can identify the reason we can use them to test this function on both in-memory and disk.
+    """
+    from netCDF4 import Dataset
+
+    url = "https://data.ioos.us/gliders/erddap/tabledap/cp_336-20170116T1254.nc"
+    auth = None
+    _nc = _nc_dataset(url, auth)
+    assert isinstance(_nc, Dataset)
