### 2020-10-04 2a6b2f6
make urlopen return consistent
