+++ tests/test_utilities.py
@@ -30,6 +30,7 @@ def test_servers():
 
 
 @pytest.mark.web
+@pytest.mark.vcr()
 def test_urlopen():
     """Assure that urlopen is always a BytesIO object."""
     url = "http://erddap.sensors.ioos.us/erddap/tabledap/"
@@ -38,6 +39,7 @@ def test_urlopen():
 
 
 @pytest.mark.web
+@pytest.mark.vcr()
 def test_urlopen_raise():
     """Assure that urlopen will raise for bad URLs."""
     url = "https://developer.mozilla.org/en-US/404"
@@ -46,6 +48,7 @@ def test_urlopen_raise():
 
 
 @pytest.mark.web
+@pytest.mark.vcr()
 def test_urlopen_requests_kwargs():
     """ Test that urlopen can pass kwargs to requests """
     base_url = "http://erddap.sensors.ioos.us/erddap/tabledap/"
@@ -63,6 +66,7 @@ def test_urlopen_requests_kwargs():
 
 
 @pytest.mark.web
+@pytest.mark.vcr()
 def test_check_url_response():
     """Test if a bad request returns HTTPError."""
     bad_request = (
@@ -148,9 +152,10 @@ def test_quote_string_constraints():
 
 
 @pytest.mark.web
+@pytest.mark.vcr()
 def test__tempnc():
     url = "https://data.ioos.us/gliders/erddap/tabledap/cp_336-20170116T1254.nc"
-    data = urlopen(url).read()
+    data = urlopen(url)
     with _tempnc(data) as tmp:
         # Check that the file was exists.
         assert os.path.exists(tmp)
@@ -161,12 +166,13 @@ def test__tempnc():
 
 
 @pytest.mark.web
+@pytest.mark.vcr()
 def test__nc_dataset():
     """
     FIXME: we need to test both in-memory and local file.
     That can be achieve with a different libnetcdf but having two environments for testing is cumbersome.
     However, it turns out sometimes a server can fail to provide files can be loaded in memory (#137).
-    If we can identify the reason we can use them to test this function on both in-memory and disk.
+    If we identify the reason we can use them to test this function on both in-memory and disk options.
     """
     from netCDF4 import Dataset
 
