+++ tests/cassettes/test_get_var_by_attr.yaml
@@ -1,0 +1,156 @@
+interactions:
+- request:
+    body: null
+    headers:
+      Accept:
+      - '*/*'
+      Accept-Encoding:
+      - gzip, deflate
+      Connection:
+      - keep-alive
+      User-Agent:
+      - python-requests/2.24.0
+    method: GET
+    uri: https://upwell.pfeg.noaa.gov/erddap/info/gtoppAT/index.csv
+  response:
+    body:
+      string: !!binary |
+        H4sIAAAAAAAAAL1Ya2/bOBb9Pr+C8JckgKzIdp3UAYJd10mDLtwkk3g6O0UAg5EohVuJ1JCUU//7
+        PaQkP2I6NbDAFgUciefyvs8l9SBfyWxZsuAbVZw+54zc0oIFY2MUf65M83hFDW1hecV+o+1ycDuZ
+        30zvPo2nQZwU8wS4ubG4RwBEFswU/Q+LjVTLvTJmBZkvGht0K96JZVFI4WwgS0bVFStzuWRJQEol
+        rVhAjCzLL1cB0QzC+W1VPDMVEK6vFE8NUx2v4okUCyYMl2Kta3I3frh6DMjkc7cXfgjIH4Jbf4h1
+        XjNDrriO5YKpJVn0wiggt5PJ4zdAI7+KWDEKr+asoDxvlUyXQoRX7JUb808hKQ0zufBKX1MN40Uh
+        tZnbv61wIisEJxgMR+HovOcVS6GzUmx2eAoyJnVJDUI3z6mZF/Rnq+d8EJ6dnR8kxEUr1D0/Dz+O
+        /Ma9kaoEN6voJyxTjOm5kMq8/FJaik1D3wvIW6m1pVEYHSTiNZMhJV7hF65tqFv8vZILnrCEJCiC
+        JXlekvHsFiVL7mnMUx6Tz1y/oHCZJtdiwZUUBcqS5mRKn8kDS6UqoAXykLS1Q+raIce+QjrxWsRF
+        Kv9Qqwp8MabUF6enme2bUKrsdI8Uas5UtkNaybHgBSybsZwVzKANbpl5leqHVz7nMRN6VYSd2Qsj
+        rpsKijAwUmk4RUVCFEsQslo+IXCYpIgwwSN6mAiJH2GYQAyfhF3NWQYjII6O5yJmhBu3ZywRNy6A
+        pnFcKYSX6RAmcoP4EtOofxLo/FqbBEtcP4A4bu/G4wCaahS63hryaKhBUm5sw7uc1Agqlk9CphbJ
+        FWGFYyPgsGQNgFrsqwNY9ANvgSavVCkqzDIg7GepmLbYJ0SyzLmlMbiQVwlC1AJtKUBBwVT8gkf6
+        zHNuli5UKTfCbmDDAE9KqgyPq5wqUlaqlDYkdkXrqmiU18HKebtL2jjZhGgZPAkwbJkzRBg7O3mE
+        Nq3y5tF6ijTYGrKViGoI/Wz3lRnq2P9/Z1a/gltLDC0juocNSnyHqrSsVMw2GuA4lzGCknLMGX/H
+        PMpqv6r3CE4bpImqZC4wsOYL6Hm26VmxgfW/75esnhGIb/+3+YcSKTYMs+2pmwY9rkr8tpR1Qqhi
+        aElZuMLZxwHkGLx2Yo3KFC1CQj6xrBKoRSIFq8sIu3CWJ63htqjcW9veSrjaws4TsEblOuArggHZ
[…]
