+++ tests/cassettes/test_to_xarray_tabledap.yaml
@@ -1,0 +1,172 @@
+interactions:
+- request:
+    body: null
+    headers:
+      Accept:
+      - '*/*'
+      Accept-Encoding:
+      - gzip, deflate
+      Connection:
+      - keep-alive
+      User-Agent:
+      - python-requests/2.24.0
+    method: GET
+    uri: https://upwell.pfeg.noaa.gov/erddap/tabledap/pmelTao5dayIso.ncCF?ISO_6,time&time%3E=248011200.0&time%3C=1548504000.0&latitude%3E=10&latitude%3C=10&longitude%3E=265&longitude%3C=265
+  response:
+    body:
+      string: ''
+    headers:
+      Connection:
+      - close
+      Content-Length:
+      - '0'
+      Content-Type:
+      - text/plain; charset=UTF-8
+      Date:
+      - Sun, 04 Oct 2020 15:45:10 GMT
+      Last-Modified:
+      - Sun, 04 Oct 2020 15:45:10 GMT
+      Location:
+      - https://coastwatch.pfeg.noaa.gov/erddap/tabledap/pmelTao5dayIso.ncCF?ISO_6%2Ctime&time%3E=248011200.0&time%3C=1548504000.0&latitude%3E=10&latitude%3C=10&longitude%3E=265&longitude%3C=265
+      Strict-Transport-Security:
+      - max-age=31536000; includeSubDomains
+      erddap-server:
+      - '2.02'
+      xdods-server:
+      - dods/3.7
+    status:
+      code: 302
+      message: ''
+- request:
+    body: null
+    headers:
+      Accept:
+      - '*/*'
+      Accept-Encoding:
+      - gzip, deflate
+      Connection:
+      - keep-alive
+      User-Agent:
+      - python-requests/2.24.0
+    method: GET
+    uri: https://coastwatch.pfeg.noaa.gov/erddap/tabledap/pmelTao5dayIso.ncCF?ISO_6%2Ctime&time%3E=248011200.0&time%3C=1548504000.0&latitude%3E=10&latitude%3C=10&longitude%3E=265&longitude%3C=265
+  response:
+    body:
+      string: !!binary |
+        H4sIAAAAAAAAAO2ZeXxTxb7AU0qhBCo7VhYdWQukbVJaoJXFmbSFCm2hKbsaT5OT9kiSE89JuoAs
+        FwWxgiKbCBdE7hUEXAAFRFBQEJ+yuSCiPnkoixdwQXEDFe5v5nfSpFiW53t/vD8eH9qcnO8sv/nN
[…]
