+++ tests/cassettes/test_to_pandas.yaml
@@ -1,0 +1,123 @@
+interactions:
+- request:
+    body: null
+    headers:
+      Accept:
+      - '*/*'
+      Accept-Encoding:
+      - gzip, deflate
+      Connection:
+      - keep-alive
+      User-Agent:
+      - python-requests/2.24.0
+    method: GET
+    uri: https://upwell.pfeg.noaa.gov/erddap/tabledap/pmelTao5dayIso.csvp?ISO_6,time&time%3E=248011200.0&time%3C=1548504000.0&latitude%3E=10&latitude%3C=10&longitude%3E=265&longitude%3C=265
+  response:
+    body:
+      string: ''
+    headers:
+      Connection:
+      - close
+      Content-Length:
+      - '0'
+      Content-Type:
+      - text/plain; charset=UTF-8
+      Date:
+      - Sun, 04 Oct 2020 15:45:10 GMT
+      Last-Modified:
+      - Sun, 04 Oct 2020 15:45:10 GMT
+      Location:
+      - https://coastwatch.pfeg.noaa.gov/erddap/tabledap/pmelTao5dayIso.csvp?ISO_6%2Ctime&time%3E=248011200.0&time%3C=1548504000.0&latitude%3E=10&latitude%3C=10&longitude%3E=265&longitude%3C=265
+      Strict-Transport-Security:
+      - max-age=31536000; includeSubDomains
+      erddap-server:
+      - '2.02'
+      xdods-server:
+      - dods/3.7
+    status:
+      code: 302
+      message: ''
+- request:
+    body: null
+    headers:
+      Accept:
+      - '*/*'
+      Accept-Encoding:
+      - gzip, deflate
+      Connection:
+      - keep-alive
+      User-Agent:
+      - python-requests/2.24.0
+    method: GET
+    uri: https://coastwatch.pfeg.noaa.gov/erddap/tabledap/pmelTao5dayIso.csvp?ISO_6%2Ctime&time%3E=248011200.0&time%3C=1548504000.0&latitude%3E=10&latitude%3C=10&longitude%3E=265&longitude%3C=265
+  response:
+    body:
+      string: !!binary |
+        H4sIAAAAAAAAAHWZva4ttw2Fez+FSxu4dyCSEkmlTZXGKeI0aYIUKVK4y/sjS2fPiEMqF9jdB0ra
+        I/4sUn/521//qT//8sev3/77nz/+/fMvf//9z7/+RP3yqXPaN5pzfif+3uR34j+1ht8/fiK92M01
+        qL8ot6vZlDk2JsnYezd94WTtl7v21+L8tha6Rtc+4mT8tpZ+jUadv3Fr7Xuj743fdFxdvMF4Y3vh
+        zhf3OWVTeht3HKyTxtL0th3tmtL6DMzJWC5qnWNjfhvLvNQmy4PxQSn/qc7DOajmD6ZM7mFMb+NO
[…]
