### 2020-10-04 476deb9
Update docs after building Travis build 356 of ioos/erddapy
The docs were built from the tag 'v0.7.2' against the commit
f84e7061f6cc7aa4dd3e793e84ae8c10156f560e.

The Travis build that generated this commit is at
https://travis-ci.org/ioos/erddapy/jobs/732747966.

The doctr command that was run is

    /home/travis/micromamba/envs/TEST/lib/python3.8/site-packages/doctr/__main__.py deploy --build-tags --key-path github_deploy_key_ioos_erddapy.enc --built-docs docs/build/html version-v0.7.2
