+++ version-v0.7.2/_sources/searchfor-output.ipynb.txt
@@ -1,0 +1,322 @@
+{
+ "cells": [
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "# Searching datasets\n",
+    "\n",
+    "\n",
+    "erddapy can wrap the same form-like search capabilities of ERDDAP with the *search_for* keyword."
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 1,
+   "metadata": {
+    "execution": {
+     "iopub.execute_input": "2020-10-04T16:52:15.373064Z",
+     "iopub.status.busy": "2020-10-04T16:52:15.372359Z",
+     "iopub.status.idle": "2020-10-04T16:52:15.896198Z",
+     "shell.execute_reply": "2020-10-04T16:52:15.895643Z"
+    }
+   },
+   "outputs": [],
+   "source": [
+    "from erddapy import ERDDAP\n",
+    "\n",
+    "\n",
+    "e = ERDDAP(\n",
+    "    server=\"https://upwell.pfeg.noaa.gov/erddap\",\n",
+    "    protocol=\"griddap\"\n",
+    ")"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "Single word search."
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 2,
+   "metadata": {
+    "execution": {
+     "iopub.execute_input": "2020-10-04T16:52:15.906104Z",
+     "iopub.status.busy": "2020-10-04T16:52:15.905674Z",
+     "iopub.status.idle": "2020-10-04T16:52:16.593413Z",
+     "shell.execute_reply": "2020-10-04T16:52:16.593042Z"
+    }
+   },
+   "outputs": [
+    {
+     "data": {
+      "text/plain": [
+       "0     erdCAnma_LonPM180\n",
+       "1     erdCMnma_LonPM180\n",
+       "2     erdCBsfc_LonPM180\n",
+       "3     erdC125h_LonPM180\n",
+       "4              erdC125h\n",
+       "5              erdC1nma\n",
+       "6              erdC1sfc\n",
+       "7              erdCA25h\n",
+       "8              erdCAnma\n",
+       "9              erdCAsfc\n",
+       "10             erdCB25h\n",
+       "11             erdCBsfc\n",
+       "12             erdCM25h\n",
+       "13             erdCMnma\n",
+       "14             erdCMsfc\n",
+       "15             erdCS25h\n",
+       "16             erdCSsfc\n",
+       "17    erdC1nma_LonPM180\n",
+       "18    erdC1sfc_LonPM180\n",
+       "19    erdCA25h_LonPM180\n",
+       "20    erdCAsfc_LonPM180\n",
+       "21    erdCB25h_LonPM180\n",
+       "22    erdCM25h_LonPM180\n",
+       "23    erdCMsfc_LonPM180\n",
+       "24    erdCS25h_LonPM180\n",
+       "25    erdCSsfc_LonPM180\n",
+       "26            ucsdHfrA6\n",
+       "27            ucsdHfrE1\n",
+       "28            ucsdHfrE2\n",
+       "29            ucsdHfrE6\n",
+       "30            ucsdHfrH1\n",
+       "31            ucsdHfrP2\n",
+       "32            ucsdHfrP6\n",
+       "33            ucsdHfrW1\n",
+       "34            ucsdHfrW2\n",
+       "35          ucsdHfrW500\n",
+       "36            ucsdHfrW6\n",
+       "Name: Dataset ID, dtype: object"
+      ]
+     },
+     "execution_count": 1,
+     "metadata": {},
[…]
