### 2020-10-05 81784b5
move netcdf handling functions out of utilities
