+++ erddapy/erddapy.py
@@ -11,14 +11,9 @@ from urllib.parse import quote_plus
 
 import pandas as pd
 
+from erddapy.netcdf_handling import _nc_dataset, _tempnc
 from erddapy.servers import servers
-from erddapy.utilities import (
-    _nc_dataset,
-    _tempnc,
-    parse_dates,
-    quote_string_constraints,
-    urlopen,
-)
+from erddapy.utilities import parse_dates, quote_string_constraints, urlopen
 
 
 ListLike = Union[List[str], Tuple[str]]
