### 2020-10-05 3f3b256
move URL handling functions out of utilities
