### 2020-10-05 53f1a04
move functions from utilities to erddapy
