+++ erddapy/erddapy.py
@@ -6,19 +6,57 @@ Pythonic way to access ERDDAP data
 import copy
 import functools
 
+from datetime import datetime
 from typing import Dict, List, Optional, Tuple, Union
 from urllib.parse import quote_plus
 
 import pandas as pd
+import pytz
 
 from erddapy.netcdf_handling import _nc_dataset, _tempnc
 from erddapy.servers import servers
 from erddapy.url_handling import _distinct, urlopen
-from erddapy.utilities import parse_dates, quote_string_constraints
+
+
+try:
+    from pandas.core.indexes.period import parse_time_string
+except ImportError:
+    from pandas._libs.tslibs.parsing import parse_time_string
 
 
 ListLike = Union[List[str], Tuple[str]]
 OptionalStr = Optional[str]
+
+
+def _quote_string_constraints(kwargs: Dict) -> Dict:
+    """
+    For constraints of String variables,
+    the right-hand-side value must be surrounded by double quotes.
+
+    """
+    return {k: f'"{v}"' if isinstance(v, str) else v for k, v in kwargs.items()}
+
+
+def parse_dates(date_time: Union[datetime, str]) -> float:
+    """
+    ERDDAP ReSTful API standardizes the representation of dates as either ISO
+    strings or seconds since 1970, but internally ERDDAPY uses datetime-like
+    objects. `timestamp` returns the expected strings in seconds since 1970.
+
+    """
+    if isinstance(date_time, str):
+        # pandas returns a tuple with datetime, dateutil, and string representation.
+        # we want only the datetime obj.
+        parse_date_time = parse_time_string(date_time)[0]
+    else:
+        parse_date_time = date_time
+
+    if not parse_date_time.tzinfo:
+        parse_date_time = pytz.utc.localize(parse_date_time)
+    else:
+        parse_date_time = parse_date_time.astimezone(pytz.utc)
+
+    return parse_date_time.timestamp()
 
 
 class ERDDAP(object):
@@ -310,7 +348,7 @@ class ERDDAP(object):
             for k, v in _constraints.items():
                 if k.startswith("time"):
                     _constraints.update({k: parse_dates(v)})
-            _constraints = quote_string_constraints(_constraints)
+            _constraints = _quote_string_constraints(_constraints)
             _constraints = "".join([f"&{k}{v}" for k, v in _constraints.items()])
 
             url += f"{_constraints}"
