+++ version-v0.6.0/_sources/erddapy.rst.txt
@@ -1,7 +1,0 @@
-:mod:`ERDDAP class API`
------------------------
-
-.. automodule:: erddapy
-   :members:
-   :undoc-members:
-   :show-inheritance:
