+++ version-v0.6.0/_sources/searchfor-output.ipynb.txt
@@ -1,281 +1,0 @@
-{
- "cells": [
-  {
-   "cell_type": "markdown",
-   "metadata": {},
-   "source": [
-    "# Searching datasets\n",
-    "\n",
-    "\n",
-    "erddapy can wrap the same form-like search capabilities of ERDDAP with the *search_for* keyword."
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": 1,
-   "metadata": {},
-   "outputs": [],
-   "source": [
-    "from erddapy import ERDDAP\n",
-    "\n",
-    "\n",
-    "e = ERDDAP(\n",
-    "    server=\"https://upwell.pfeg.noaa.gov/erddap\",\n",
-    "    protocol=\"griddap\"\n",
-    ")"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "metadata": {},
-   "source": [
-    "Single word search."
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": 2,
-   "metadata": {},
-   "outputs": [
-    {
-     "data": {
-      "text/plain": [
-       "0     erdCAnma_LonPM180\n",
-       "1     erdC125h_LonPM180\n",
-       "2     erdC1nma_LonPM180\n",
-       "3     erdC1sfc_LonPM180\n",
-       "4     erdCAsfc_LonPM180\n",
-       "5     erdCB25h_LonPM180\n",
-       "6     erdCBsfc_LonPM180\n",
-       "7     erdCM25h_LonPM180\n",
-       "8     erdCMnma_LonPM180\n",
-       "9     erdCMsfc_LonPM180\n",
-       "10    erdCS25h_LonPM180\n",
-       "11    erdCSsfc_LonPM180\n",
-       "12    erdCA25h_LonPM180\n",
-       "13             erdC125h\n",
-       "14             erdC1nma\n",
-       "15             erdC1sfc\n",
-       "16             erdCA25h\n",
-       "17             erdCAnma\n",
-       "18             erdCAsfc\n",
-       "19             erdCB25h\n",
-       "20             erdCBsfc\n",
-       "21             erdCM25h\n",
-       "22             erdCMnma\n",
-       "23             erdCMsfc\n",
-       "24             erdCS25h\n",
-       "25             erdCSsfc\n",
-       "26            ucsdHfrP2\n",
-       "27            ucsdHfrA6\n",
-       "28            ucsdHfrH1\n",
-       "29            ucsdHfrP6\n",
-       "30            ucsdHfrW6\n",
-       "31            ucsdHfrE6\n",
-       "32            ucsdHfrW1\n",
-       "33            ucsdHfrW2\n",
-       "34          ucsdHfrW500\n",
-       "35            ucsdHfrE1\n",
-       "36            ucsdHfrE2\n",
-       "Name: Dataset ID, dtype: object"
-      ]
-     },
-     "execution_count": 2,
-     "metadata": {},
-     "output_type": "execute_result"
-    }
-   ],
-   "source": [
-    "import pandas as pd\n",
-    "\n",
-    "search_for = \"HFRadar\"\n",
-    "\n",
-    "url = e.get_search_url(search_for=search_for, response=\"csv\")\n",
-    "\n",
-    "pd.read_csv(url)[\"Dataset ID\"]"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "metadata": {},
-   "source": [
-    "Filtering the search with extra words."
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": 3,
-   "metadata": {},
-   "outputs": [
-    {
-     "data": {
-      "text/plain": [
-       "0    ucsdHfrP2\n",
[…]
