+++ .travis.yml
@@ -60,13 +60,9 @@ script:
       pushd docs
       make clean html linkcheck
       popd
-      if [[ -z "$TRAVIS_TAG" ]]; then
-        python -m doctr deploy --build-tags --key-path github_deploy_key_ioos_erddapy.enc --built-docs docs/build/html dev
-      else
-        python -m doctr deploy --build-tags --key-path github_deploy_key_ioos_erddapy.enc --built-docs docs/build/html "version-$TRAVIS_TAG"
-        python -m doctr deploy --build-tags --key-path github_deploy_key_ioos_erddapy.enc --built-docs docs/build/html .
-      fi
-    fi
+
+      if [ ! -z "$TRAVIS_TAG" ]; then DEPLOY_DIR="$TRAVIS_TAG"; else DEPLOY_DIR="$TRAVIS_BRANCH"; fi
+      doctr deploy --command=doctr-versions-menu --no-require-master --build-tags "$DEPLOY_DIR"
 
 deploy:
   skip_cleanup: true
