### 2020-10-05 9b36322
Update docs after building Travis build 362 of ioos/erddapy
The docs were built from the branch 'master' against the commit
8a6453de2fbc6737fcd5edaf8d38d7b8549da0cd.

The Travis build that generated this commit is at
https://travis-ci.org/ioos/erddapy/jobs/733030705.

The doctr command that was run is

    /home/travis/micromamba/envs/TEST/bin/doctr deploy --command=doctr-versions-menu --no-require-master --build-tags master
