### 2020-10-05 e1947d2
Update docs after building Travis build 363 of ioos/erddapy
The docs were built from the tag 'v0.7.2' against the commit
96be378bae8832bc995363c8804f0d088bce69c8.

The Travis build that generated this commit is at
https://travis-ci.org/ioos/erddapy/jobs/733040583.

The doctr command that was run is

    /home/travis/micromamba/envs/TEST/bin/doctr deploy --command=doctr-versions-menu --no-require-master --build-tags v0.7.2
