### 2020-10-06 f042325
refactor docs
