+++ docs/source/index.rst
@@ -36,8 +36,10 @@ See https://coastwatch.pfeg.noaa.gov/erddap/index.html for more information.
    :maxdepth: 3
    :caption: Contents:
 
-   quick_intro-output.ipynb
-   searchfor-output.ipynb
+   00-quick_intro-output.ipynb
+   01-longer_intro-output.ipynb
+   02-extras-output.ipynb
+   03-searchfor-output.ipynb
    erddapy
 
 Indices and tables
