### 2020-11-02 88fc415
rstrip(/) to avoid workarounds on the client side.
