+++ .pre-commit-config.yaml
@@ -1,6 +1,6 @@
 repos:
 - repo: https://github.com/pre-commit/pre-commit-hooks
-  rev: v3.2.0
+  rev: v3.3.0
   hooks:
     - id: trailing-whitespace
     - id: check-ast
@@ -22,7 +22,7 @@ repos:
       args: [--max-line-length=105]
 
 - repo: https://github.com/pycqa/isort
-  rev: 5.5.4
+  rev: 5.6.4
   hooks:
     - id: isort
       additional_dependencies: [toml]
@@ -35,14 +35,14 @@ repos:
     language_version: python3
 
 - repo: https://github.com/pre-commit/mirrors-mypy
-  rev: v0.782
+  rev: v0.790
   hooks:
   - id: mypy
     exclude: docs/source/conf.py
     args: [--ignore-missing-imports]
 
 - repo: https://github.com/keewis/blackdoc
-  rev: stable
+  rev: v0.2
   hooks:
     - id: blackdoc
 
@@ -63,3 +63,15 @@ repos:
           )$
       args:
         - --quiet-level=2
+
+- repo: https://github.com/asottile/pyupgrade
+  rev: v2.7.3
+  hooks:
+    - id: pyupgrade
+      args:
+        - --py36-plus
+
+- repo: https://github.com/asottile/add-trailing-comma
+  rev: v2.0.1
+  hooks:
+    - id: add-trailing-comma
