### 2020-11-02 a0873ed
Update docs after building Travis build 377 of ioos/erddapy
The docs were built from the branch 'master' against the commit
dd1213d68971865ecc0b69263bd3c4e9246c8c5d.

The Travis build that generated this commit is at
https://travis-ci.org/ioos/erddapy/jobs/740861026.

The doctr command that was run is

    /home/travis/micromamba/envs/TEST/bin/doctr deploy --command=doctr-versions-menu --no-require-master --build-tags master
