+++ master/_sources/03-searchfor-output.ipynb.txt
@@ -15,10 +15,10 @@
    "execution_count": 1,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2020-10-05T16:57:39.757847Z",
-     "iopub.status.busy": "2020-10-05T16:57:39.756644Z",
-     "iopub.status.idle": "2020-10-05T16:57:40.185304Z",
-     "shell.execute_reply": "2020-10-05T16:57:40.184427Z"
+     "iopub.execute_input": "2020-11-02T15:16:17.858816Z",
+     "iopub.status.busy": "2020-11-02T15:16:17.858155Z",
+     "iopub.status.idle": "2020-11-02T15:16:18.166938Z",
+     "shell.execute_reply": "2020-11-02T15:16:18.166479Z"
     }
    },
    "outputs": [],
@@ -44,53 +44,53 @@
    "execution_count": 2,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2020-10-05T16:57:40.195879Z",
-     "iopub.status.busy": "2020-10-05T16:57:40.195422Z",
-     "iopub.status.idle": "2020-10-05T16:57:41.062018Z",
-     "shell.execute_reply": "2020-10-05T16:57:41.061243Z"
+     "iopub.execute_input": "2020-11-02T15:16:18.176011Z",
+     "iopub.status.busy": "2020-11-02T15:16:18.170119Z",
+     "iopub.status.idle": "2020-11-02T15:16:18.929602Z",
+     "shell.execute_reply": "2020-11-02T15:16:18.929929Z"
     }
    },
    "outputs": [
     {
      "data": {
       "text/plain": [
-       "0     erdCAnma_LonPM180\n",
-       "1     erdCMnma_LonPM180\n",
-       "2     erdCBsfc_LonPM180\n",
-       "3     erdC125h_LonPM180\n",
-       "4              erdC125h\n",
-       "5              erdC1nma\n",
-       "6              erdC1sfc\n",
-       "7              erdCA25h\n",
-       "8              erdCAnma\n",
-       "9              erdCAsfc\n",
-       "10             erdCB25h\n",
-       "11             erdCBsfc\n",
-       "12             erdCM25h\n",
-       "13             erdCMnma\n",
-       "14             erdCMsfc\n",
-       "15             erdCS25h\n",
-       "16             erdCSsfc\n",
-       "17    erdC1nma_LonPM180\n",
-       "18    erdC1sfc_LonPM180\n",
-       "19    erdCA25h_LonPM180\n",
-       "20    erdCAsfc_LonPM180\n",
-       "21    erdCB25h_LonPM180\n",
-       "22    erdCM25h_LonPM180\n",
+       "0              erdCA25h\n",
+       "1              erdCAnma\n",
+       "2              erdCAsfc\n",
+       "3              erdCB25h\n",
+       "4              erdCBsfc\n",
+       "5              erdC125h\n",
+       "6              erdC1nma\n",
+       "7              erdC1sfc\n",
+       "8              erdCM25h\n",
+       "9              erdCMnma\n",
+       "10             erdCMsfc\n",
+       "11             erdCS25h\n",
+       "12             erdCSsfc\n",
+       "13    erdC125h_LonPM180\n",
+       "14    erdC1nma_LonPM180\n",
+       "15    erdC1sfc_LonPM180\n",
+       "16    erdCA25h_LonPM180\n",
+       "17    erdCAnma_LonPM180\n",
+       "18    erdCAsfc_LonPM180\n",
+       "19    erdCB25h_LonPM180\n",
[…]
