+++ master/_sources/00-quick_intro-output.ipynb.txt
@@ -1,0 +1,250 @@
+{
+ "cells": [
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "# Quick introduction\n",
+    "\n",
+    "\n",
+    "erddapy is a pure python package and can be installed with conda\n",
+    "\n",
+    "\n",
+    "```shell\n",
+    "conda install --channel conda-forge erddapy\n",
+    "```\n",
+    "\n",
+    " or pip\n",
+    "\n",
+    "```shell\n",
+    "pip install erddapy\n",
+    "```"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "First we need to instantiate the ERDDAP URL constructor for a server.\n",
+    "In this example we will use [https://gliders.ioos.us/erddap](https://gliders.ioos.us/erddap/index.html)."
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 1,
+   "metadata": {
+    "execution": {
+     "iopub.execute_input": "2020-11-02T15:15:44.131754Z",
+     "iopub.status.busy": "2020-11-02T15:15:44.131089Z",
+     "iopub.status.idle": "2020-11-02T15:15:44.588510Z",
+     "shell.execute_reply": "2020-11-02T15:15:44.587928Z"
+    }
+   },
+   "outputs": [],
+   "source": [
+    "from erddapy import ERDDAP\n",
+    "\n",
+    "\n",
+    "e = ERDDAP(\n",
+    "    server=\"NGDAC\", # \"NOAA IOOS NGDAC (National Glider Data Assembly Center)\"\n",
+    "    protocol=\"tabledap\",\n",
+    "    response=\"csv\",\n",
+    ")"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "Now we can populate the object a dataset id, variables of interest, and \n",
+    "its constraints. We can download the csvp response with the `.to_pandas` method."
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 2,
+   "metadata": {
+    "execution": {
+     "iopub.execute_input": "2020-11-02T15:15:44.594703Z",
+     "iopub.status.busy": "2020-11-02T15:15:44.592812Z",
+     "iopub.status.idle": "2020-11-02T15:15:45.322430Z",
+     "shell.execute_reply": "2020-11-02T15:15:45.322055Z"
+    }
+   },
+   "outputs": [
+    {
+     "data": {
+      "text/html": [
+       "<div>\n",
+       "<style scoped>\n",
+       "    .dataframe tbody tr th:only-of-type {\n",
+       "        vertical-align: middle;\n",
+       "    }\n",
+       "\n",
+       "    .dataframe tbody tr th {\n",
+       "        vertical-align: top;\n",
+       "    }\n",
+       "\n",
+       "    .dataframe thead th {\n",
+       "        text-align: right;\n",
+       "    }\n",
+       "</style>\n",
+       "<table border=\"1\" class=\"dataframe\">\n",
+       "  <thead>\n",
+       "    <tr style=\"text-align: right;\">\n",
+       "      <th></th>\n",
+       "      <th>depth (m)</th>\n",
+       "      <th>latitude (degrees_north)</th>\n",
+       "      <th>longitude (degrees_east)</th>\n",
+       "      <th>salinity (1)</th>\n",
+       "      <th>temperature (Celsius)</th>\n",
+       "    </tr>\n",
+       "    <tr>\n",
+       "      <th>time (UTC)</th>\n",
+       "      <th></th>\n",
+       "      <th></th>\n",
[…]
