### 2020-11-03 be80e5e
normilize search terms
