### 2020-11-04 f3df42d
Update docs after building Travis build 381 of ioos/erddapy
The docs were built from the branch 'master' against the commit
aa166c1eaf9a233114bd5709c80195e426564554.

The Travis build that generated this commit is at
https://travis-ci.org/ioos/erddapy/jobs/741352683.

The doctr command that was run is

    /home/travis/micromamba/envs/TEST/bin/doctr deploy --command=doctr-versions-menu --no-require-master --build-tags master
