+++ master/_sources/00-quick_intro-output.ipynb.txt
@@ -34,10 +34,10 @@
    "execution_count": 1,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2020-11-04T00:28:35.786068Z",
-     "iopub.status.busy": "2020-11-04T00:28:35.785315Z",
-     "iopub.status.idle": "2020-11-04T00:28:36.228608Z",
-     "shell.execute_reply": "2020-11-04T00:28:36.228114Z"
+     "iopub.execute_input": "2020-11-04T10:36:40.114017Z",
+     "iopub.status.busy": "2020-11-04T10:36:40.113332Z",
+     "iopub.status.idle": "2020-11-04T10:36:40.470081Z",
+     "shell.execute_reply": "2020-11-04T10:36:40.469513Z"
     }
    },
    "outputs": [],
@@ -65,10 +65,10 @@
    "execution_count": 2,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2020-11-04T00:28:36.235528Z",
-     "iopub.status.busy": "2020-11-04T00:28:36.235096Z",
-     "iopub.status.idle": "2020-11-04T00:28:36.448154Z",
-     "shell.execute_reply": "2020-11-04T00:28:36.448567Z"
+     "iopub.execute_input": "2020-11-04T10:36:40.477245Z",
+     "iopub.status.busy": "2020-11-04T10:36:40.474661Z",
+     "iopub.status.idle": "2020-11-04T10:36:40.726483Z",
+     "shell.execute_reply": "2020-11-04T10:36:40.726033Z"
     }
    },
    "outputs": [
