+++ master/_sources/01-longer_intro-output.ipynb.txt
@@ -16,10 +16,10 @@
    "execution_count": 1,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2020-11-04T00:28:38.240923Z",
-     "iopub.status.busy": "2020-11-04T00:28:38.240181Z",
-     "iopub.status.idle": "2020-11-04T00:28:38.603638Z",
-     "shell.execute_reply": "2020-11-04T00:28:38.603167Z"
+     "iopub.execute_input": "2020-11-04T10:36:42.524947Z",
+     "iopub.status.busy": "2020-11-04T10:36:42.524156Z",
+     "iopub.status.idle": "2020-11-04T10:36:42.853125Z",
+     "shell.execute_reply": "2020-11-04T10:36:42.852683Z"
     }
    },
    "outputs": [
@@ -71,10 +71,10 @@
    "execution_count": 2,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2020-11-04T00:28:38.607547Z",
-     "iopub.status.busy": "2020-11-04T00:28:38.607141Z",
-     "iopub.status.idle": "2020-11-04T00:28:38.610572Z",
-     "shell.execute_reply": "2020-11-04T00:28:38.610134Z"
+     "iopub.execute_input": "2020-11-04T10:36:42.857821Z",
+     "iopub.status.busy": "2020-11-04T10:36:42.856896Z",
+     "iopub.status.idle": "2020-11-04T10:36:42.859183Z",
+     "shell.execute_reply": "2020-11-04T10:36:42.859553Z"
     }
    },
    "outputs": [
@@ -108,10 +108,10 @@
    "execution_count": 3,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2020-11-04T00:28:38.620308Z",
-     "iopub.status.busy": "2020-11-04T00:28:38.613742Z",
-     "iopub.status.idle": "2020-11-04T00:28:38.802400Z",
-     "shell.execute_reply": "2020-11-04T00:28:38.802010Z"
+     "iopub.execute_input": "2020-11-04T10:36:42.869343Z",
+     "iopub.status.busy": "2020-11-04T10:36:42.868870Z",
+     "iopub.status.idle": "2020-11-04T10:36:43.190557Z",
+     "shell.execute_reply": "2020-11-04T10:36:43.190968Z"
     }
    },
    "outputs": [
@@ -155,10 +155,10 @@
    "execution_count": 4,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2020-11-04T00:28:38.806859Z",
-     "iopub.status.busy": "2020-11-04T00:28:38.805954Z",
-     "iopub.status.idle": "2020-11-04T00:28:38.811969Z",
-     "shell.execute_reply": "2020-11-04T00:28:38.812464Z"
+     "iopub.execute_input": "2020-11-04T10:36:43.195493Z",
+     "iopub.status.busy": "2020-11-04T10:36:43.194815Z",
+     "iopub.status.idle": "2020-11-04T10:36:43.199178Z",
+     "shell.execute_reply": "2020-11-04T10:36:43.198824Z"
     }
    },
    "outputs": [
@@ -176,7 +176,7 @@
        "        "
       ],
       "text/plain": [
-       "<IPython.lib.display.IFrame at 0x7f3d10a6c280>"
+       "<IPython.lib.display.IFrame at 0x7f63f0bd69a0>"
       ]
      },
      "execution_count": 1,
@@ -218,10 +218,10 @@
    "execution_count": 5,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2020-11-04T00:28:38.822027Z",
[…]
