### 2020-11-04 520996d
Update docs after building Travis build 382 of ioos/erddapy
The docs were built from the tag 'v0.8.0' against the commit
aa166c1eaf9a233114bd5709c80195e426564554.

The Travis build that generated this commit is at
https://travis-ci.org/ioos/erddapy/jobs/741362487.

The doctr command that was run is

    /home/travis/micromamba/envs/TEST/bin/doctr deploy --command=doctr-versions-menu --no-require-master --build-tags v0.8.0
