### 2020-11-11 27c4fb1
Update docs after building Travis build 386 of ioos/erddapy
The docs were built from the branch 'master' against the commit
f4bb53dc53ae8ea59d52028b755a4821ad16d93e.

The Travis build that generated this commit is at
https://travis-ci.org/ioos/erddapy/jobs/742983234.

The doctr command that was run is

    /home/travis/micromamba/envs/TEST/bin/doctr deploy --command=doctr-versions-menu --no-require-master --build-tags master
