+++ master/_sources/03-searchfor-output.ipynb.txt
@@ -15,10 +15,10 @@
    "execution_count": 1,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2020-11-04T10:37:05.792193Z",
-     "iopub.status.busy": "2020-11-04T10:37:05.791492Z",
-     "iopub.status.idle": "2020-11-04T10:37:06.138684Z",
-     "shell.execute_reply": "2020-11-04T10:37:06.138193Z"
+     "iopub.execute_input": "2020-11-11T17:25:48.316223Z",
+     "iopub.status.busy": "2020-11-11T17:25:48.315502Z",
+     "iopub.status.idle": "2020-11-11T17:25:48.668370Z",
+     "shell.execute_reply": "2020-11-11T17:25:48.667891Z"
     }
    },
    "outputs": [],
@@ -44,24 +44,24 @@
    "execution_count": 2,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2020-11-04T10:37:06.148575Z",
-     "iopub.status.busy": "2020-11-04T10:37:06.148052Z",
-     "iopub.status.idle": "2020-11-04T10:37:06.785271Z",
-     "shell.execute_reply": "2020-11-04T10:37:06.785682Z"
+     "iopub.execute_input": "2020-11-11T17:25:48.678335Z",
+     "iopub.status.busy": "2020-11-11T17:25:48.677923Z",
+     "iopub.status.idle": "2020-11-11T17:25:49.395557Z",
+     "shell.execute_reply": "2020-11-11T17:25:49.395158Z"
     }
    },
    "outputs": [
     {
      "data": {
       "text/plain": [
-       "0              erdCA25h\n",
-       "1              erdCAnma\n",
-       "2              erdCAsfc\n",
-       "3              erdCB25h\n",
-       "4              erdCBsfc\n",
-       "5              erdC125h\n",
-       "6              erdC1nma\n",
-       "7              erdC1sfc\n",
+       "0              erdC125h\n",
+       "1              erdC1nma\n",
+       "2              erdC1sfc\n",
+       "3              erdCA25h\n",
+       "4              erdCAnma\n",
+       "5              erdCAsfc\n",
+       "6              erdCB25h\n",
+       "7              erdCBsfc\n",
        "8              erdCM25h\n",
        "9              erdCMnma\n",
        "10             erdCMsfc\n",
@@ -72,25 +72,25 @@
        "15    erdC1sfc_LonPM180\n",
        "16    erdCA25h_LonPM180\n",
        "17    erdCAnma_LonPM180\n",
-       "18    erdCB25h_LonPM180\n",
-       "19    erdCBsfc_LonPM180\n",
-       "20    erdCM25h_LonPM180\n",
-       "21    erdCMnma_LonPM180\n",
-       "22    erdCMsfc_LonPM180\n",
-       "23    erdCS25h_LonPM180\n",
+       "18    erdCAsfc_LonPM180\n",
+       "19    erdCB25h_LonPM180\n",
+       "20    erdCBsfc_LonPM180\n",
+       "21    erdCM25h_LonPM180\n",
+       "22    erdCMnma_LonPM180\n",
+       "23    erdCMsfc_LonPM180\n",
        "24    erdCSsfc_LonPM180\n",
-       "25    erdCAsfc_LonPM180\n",
+       "25    erdCS25h_LonPM180\n",
        "26            ucsdHfrA6\n",
-       "27            ucsdHfrE1\n",
-       "28            ucsdHfrE2\n",
-       "29            ucsdHfrE6\n",
-       "30            ucsdHfrH1\n",
[…]
