+++ master/_sources/02-extras-output.ipynb.txt
@@ -14,10 +14,10 @@
    "execution_count": 1,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2020-11-04T10:37:00.184880Z",
-     "iopub.status.busy": "2020-11-04T10:37:00.184146Z",
-     "iopub.status.idle": "2020-11-04T10:37:00.537161Z",
-     "shell.execute_reply": "2020-11-04T10:37:00.536354Z"
+     "iopub.execute_input": "2020-11-11T17:25:43.002029Z",
+     "iopub.status.busy": "2020-11-11T17:25:43.001316Z",
+     "iopub.status.idle": "2020-11-11T17:25:43.343933Z",
+     "shell.execute_reply": "2020-11-11T17:25:43.343008Z"
     }
    },
    "outputs": [],
@@ -33,10 +33,10 @@
    "execution_count": 2,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2020-11-04T10:37:00.545797Z",
-     "iopub.status.busy": "2020-11-04T10:37:00.540776Z",
-     "iopub.status.idle": "2020-11-04T10:37:00.848680Z",
-     "shell.execute_reply": "2020-11-04T10:37:00.849083Z"
+     "iopub.execute_input": "2020-11-11T17:25:43.348575Z",
+     "iopub.status.busy": "2020-11-11T17:25:43.347635Z",
+     "iopub.status.idle": "2020-11-11T17:25:43.523945Z",
+     "shell.execute_reply": "2020-11-11T17:25:43.524406Z"
     }
    },
    "outputs": [
@@ -78,10 +78,10 @@
    "execution_count": 3,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2020-11-04T10:37:00.853371Z",
-     "iopub.status.busy": "2020-11-04T10:37:00.852739Z",
-     "iopub.status.idle": "2020-11-04T10:37:03.031791Z",
-     "shell.execute_reply": "2020-11-04T10:37:03.031432Z"
+     "iopub.execute_input": "2020-11-11T17:25:43.530628Z",
+     "iopub.status.busy": "2020-11-11T17:25:43.528440Z",
+     "iopub.status.idle": "2020-11-11T17:25:45.528602Z",
+     "shell.execute_reply": "2020-11-11T17:25:45.528146Z"
     },
     "scrolled": false
    },
@@ -137,10 +137,10 @@
    "execution_count": 4,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2020-11-04T10:37:03.035384Z",
-     "iopub.status.busy": "2020-11-04T10:37:03.034795Z",
-     "iopub.status.idle": "2020-11-04T10:37:03.449298Z",
-     "shell.execute_reply": "2020-11-04T10:37:03.448911Z"
+     "iopub.execute_input": "2020-11-11T17:25:45.533273Z",
+     "iopub.status.busy": "2020-11-11T17:25:45.531536Z",
+     "iopub.status.idle": "2020-11-11T17:25:45.959264Z",
+     "shell.execute_reply": "2020-11-11T17:25:45.958770Z"
     }
    },
    "outputs": [
@@ -568,13 +568,13 @@
        "    time_coverage_end:             2016-09-20T13:32:33Z\n",
        "    time_coverage_start:           2016-09-02T17:05:38Z\n",
        "    title:                         whoi_406-20160902T1700\n",
[…]

[…]
