+++ CHANGES.txt
@@ -1,5 +1,9 @@
 # Changelog
 
+## v0.9.0
+
+- Support ERDDAP 2.10 search syntax (#155)
+
 ## v0.8.0
 
 - Cache urlopen returns for faster access when the query does not change (#147)
