### 2020-11-12 e14d82a
setup python to avoid ubuntu defaults
