+++ .github/workflows/publish.yml
@@ -11,6 +11,11 @@ jobs:
     steps:
     - uses: actions/checkout@v2
 
+    - name: Set up Python
+      uses: actions/setup-python@v1
+      with:
+        python-version: 3.8
+
     - name: Get tags
       run: git fetch --depth=1 origin +refs/tags/*:refs/tags/*
       shell: bash
