### 2020-11-12 8c121cb
Update docs after building Travis build 393 of ioos/erddapy
The docs were built from the tag 'v0.9.0' against the commit
fb70c908a91dfef09419bb4926ed37651b339f92.

The Travis build that generated this commit is at
https://travis-ci.org/ioos/erddapy/jobs/743063995.

The doctr command that was run is

    /home/travis/micromamba/envs/TEST/bin/doctr deploy --command=doctr-versions-menu --no-require-master --build-tags v0.9.0
