+++ v0.9.0/_sources/erddapy.rst.txt
@@ -1,0 +1,7 @@
+:mod:`ERDDAP class API`
+-----------------------
+
+.. automodule:: erddapy.erddapy
+   :members:
+   :undoc-members:
+   :show-inheritance:
