+++ v0.9.0/_sources/02-extras-output.ipynb.txt
@@ -1,0 +1,1114 @@
+{
+ "cells": [
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "## Extra convenience methods for common responses\n",
+    "\n",
+    "### OPeNDAP"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 1,
+   "metadata": {
+    "execution": {
+     "iopub.execute_input": "2020-11-12T01:05:26.554869Z",
+     "iopub.status.busy": "2020-11-12T01:05:26.554147Z",
+     "iopub.status.idle": "2020-11-12T01:05:26.870368Z",
+     "shell.execute_reply": "2020-11-12T01:05:26.869594Z"
+    }
+   },
+   "outputs": [],
+   "source": [
+    "from erddapy import ERDDAP\n",
+    "\n",
+    "\n",
+    "e = ERDDAP(server=\"https://gliders.ioos.us/erddap\")"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 2,
+   "metadata": {
+    "execution": {
+     "iopub.execute_input": "2020-11-12T01:05:26.878938Z",
+     "iopub.status.busy": "2020-11-12T01:05:26.873821Z",
+     "iopub.status.idle": "2020-11-12T01:05:27.191190Z",
+     "shell.execute_reply": "2020-11-12T01:05:27.191588Z"
+    }
+   },
+   "outputs": [
+    {
+     "name": "stdout",
+     "output_type": "stream",
+     "text": [
+      "https://gliders.ioos.us/erddap/tabledap/whoi_406-20160902T1700\n",
+      "Slocum glider dataset gathered as part of the TEMPESTS (The Experiment to Measure and Predict East coast STorm Strength), funded by NOAA through CINAR (Cooperative Institute for the North Atlantic Region).\n"
+     ]
+    }
+   ],
+   "source": [
+    "from netCDF4 import Dataset\n",
+    "\n",
+    "\n",
+    "e.constraints = None\n",
+    "e.protocol = \"tabledap\"\n",
+    "e.dataset_id = \"whoi_406-20160902T1700\"\n",
+    "\n",
+    "opendap_url = e.get_download_url(\n",
+    "    response=\"opendap\",\n",
+    ")\n",
+    "\n",
+    "print(opendap_url)\n",
+    "with Dataset(opendap_url) as nc:\n",
+    "    print(nc.summary)"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "### netCDF Climate and Forecast"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 3,
+   "metadata": {
+    "execution": {
+     "iopub.execute_input": "2020-11-12T01:05:27.195535Z",
+     "iopub.status.busy": "2020-11-12T01:05:27.194933Z",
+     "iopub.status.idle": "2020-11-12T01:05:29.581546Z",
+     "shell.execute_reply": "2020-11-12T01:05:29.581026Z"
+    },
+    "scrolled": false
+   },
+   "outputs": [
+    {
+     "name": "stdout",
+     "output_type": "stream",
+     "text": [
+      "Unidata Dataset Discovery v1.0, COARDS, CF-1.6\n",
+      "<class 'netCDF4._netCDF4.Variable'>\n",
+      "float32 temperature(obs)\n",
+      "    _ChunkSizes: 73\n",
+      "    _FillValue: -999.0\n",
+      "    actual_range: [ 9.9473 22.6177]\n",
[…]

[…]
