### 2020-11-12 67189e7
Update docs after building Travis build 395 of ioos/erddapy
The docs were built from the branch 'master' against the commit
101df117aa51c6cf8457dfa9f180c2e0d4c6cd95.

The Travis build that generated this commit is at
https://travis-ci.org/ioos/erddapy/jobs/743066227.

The doctr command that was run is

    /home/travis/micromamba/envs/TEST/bin/doctr deploy --command=doctr-versions-menu --no-require-master --build-tags master
