+++ master/_sources/03-searchfor-output.ipynb.txt
@@ -15,10 +15,10 @@
    "execution_count": 1,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2020-11-11T17:25:48.316223Z",
-     "iopub.status.busy": "2020-11-11T17:25:48.315502Z",
-     "iopub.status.idle": "2020-11-11T17:25:48.668370Z",
-     "shell.execute_reply": "2020-11-11T17:25:48.667891Z"
+     "iopub.execute_input": "2020-11-12T02:16:33.664578Z",
+     "iopub.status.busy": "2020-11-12T02:16:33.663843Z",
+     "iopub.status.idle": "2020-11-12T02:16:34.008149Z",
+     "shell.execute_reply": "2020-11-12T02:16:34.007666Z"
     }
    },
    "outputs": [],
@@ -44,10 +44,10 @@
    "execution_count": 2,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2020-11-11T17:25:48.678335Z",
-     "iopub.status.busy": "2020-11-11T17:25:48.677923Z",
-     "iopub.status.idle": "2020-11-11T17:25:49.395557Z",
-     "shell.execute_reply": "2020-11-11T17:25:49.395158Z"
+     "iopub.execute_input": "2020-11-12T02:16:34.017735Z",
+     "iopub.status.busy": "2020-11-12T02:16:34.017312Z",
+     "iopub.status.idle": "2020-11-12T02:16:34.757357Z",
+     "shell.execute_reply": "2020-11-12T02:16:34.756983Z"
     }
    },
    "outputs": [
@@ -81,16 +81,16 @@
        "24    erdCSsfc_LonPM180\n",
        "25    erdCS25h_LonPM180\n",
        "26            ucsdHfrA6\n",
-       "27            ucsdHfrE2\n",
-       "28            ucsdHfrE1\n",
-       "29            ucsdHfrP2\n",
-       "30            ucsdHfrP6\n",
-       "31            ucsdHfrW6\n",
-       "32            ucsdHfrE6\n",
-       "33            ucsdHfrW1\n",
-       "34            ucsdHfrW2\n",
-       "35          ucsdHfrW500\n",
-       "36            ucsdHfrH1\n",
+       "27            ucsdHfrW1\n",
+       "28            ucsdHfrW2\n",
+       "29          ucsdHfrW500\n",
+       "30            ucsdHfrE2\n",
+       "31            ucsdHfrE6\n",
+       "32            ucsdHfrE1\n",
+       "33            ucsdHfrH1\n",
+       "34            ucsdHfrP2\n",
+       "35            ucsdHfrP6\n",
+       "36            ucsdHfrW6\n",
        "Name: Dataset ID, dtype: object"
       ]
      },
@@ -121,19 +121,19 @@
    "execution_count": 3,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2020-11-11T17:25:49.405447Z",
-     "iopub.status.busy": "2020-11-11T17:25:49.404985Z",
-     "iopub.status.idle": "2020-11-11T17:25:49.852234Z",
-     "shell.execute_reply": "2020-11-11T17:25:49.851858Z"
+     "iopub.execute_input": "2020-11-12T02:16:34.766994Z",
+     "iopub.status.busy": "2020-11-12T02:16:34.766548Z",
+     "iopub.status.idle": "2020-11-12T02:16:35.233176Z",
+     "shell.execute_reply": "2020-11-12T02:16:35.232807Z"
     }
    },
    "outputs": [
     {
      "data": {
       "text/plain": [
[…]
