### 2020-11-12 fc967ee
Update docs after building Travis build 401 of ioos/erddapy
The docs were built from the branch 'master' against the commit
0391c3f7703b4a2585440d86ca58fee655d7a3a4.

The Travis build that generated this commit is at
https://travis-ci.org/ioos/erddapy/jobs/743236632.

The doctr command that was run is

    /home/travis/micromamba/envs/TEST/bin/doctr deploy --command=doctr-versions-menu --no-require-master --build-tags master
