+++ master/_sources/02-extras-output.ipynb.txt
@@ -14,10 +14,10 @@
    "execution_count": 1,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2020-11-12T02:16:28.237748Z",
-     "iopub.status.busy": "2020-11-12T02:16:28.237031Z",
-     "iopub.status.idle": "2020-11-12T02:16:28.574576Z",
-     "shell.execute_reply": "2020-11-12T02:16:28.573989Z"
+     "iopub.execute_input": "2020-11-12T20:51:33.635281Z",
+     "iopub.status.busy": "2020-11-12T20:51:33.634592Z",
+     "iopub.status.idle": "2020-11-12T20:51:33.961334Z",
+     "shell.execute_reply": "2020-11-12T20:51:33.960533Z"
     }
    },
    "outputs": [],
@@ -33,10 +33,10 @@
    "execution_count": 2,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2020-11-12T02:16:28.578671Z",
-     "iopub.status.busy": "2020-11-12T02:16:28.577973Z",
-     "iopub.status.idle": "2020-11-12T02:16:28.756782Z",
-     "shell.execute_reply": "2020-11-12T02:16:28.757197Z"
+     "iopub.execute_input": "2020-11-12T20:51:33.970249Z",
+     "iopub.status.busy": "2020-11-12T20:51:33.965295Z",
+     "iopub.status.idle": "2020-11-12T20:51:34.287463Z",
+     "shell.execute_reply": "2020-11-12T20:51:34.287051Z"
     }
    },
    "outputs": [
@@ -78,10 +78,10 @@
    "execution_count": 3,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2020-11-12T02:16:28.761574Z",
-     "iopub.status.busy": "2020-11-12T02:16:28.760921Z",
-     "iopub.status.idle": "2020-11-12T02:16:30.957922Z",
-     "shell.execute_reply": "2020-11-12T02:16:30.957545Z"
+     "iopub.execute_input": "2020-11-12T20:51:34.293058Z",
+     "iopub.status.busy": "2020-11-12T20:51:34.291079Z",
+     "iopub.status.idle": "2020-11-12T20:52:00.971545Z",
+     "shell.execute_reply": "2020-11-12T20:52:00.971921Z"
     },
     "scrolled": false
    },
@@ -137,10 +137,10 @@
    "execution_count": 4,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2020-11-12T02:16:30.961525Z",
-     "iopub.status.busy": "2020-11-12T02:16:30.960955Z",
-     "iopub.status.idle": "2020-11-12T02:16:31.373500Z",
-     "shell.execute_reply": "2020-11-12T02:16:31.373114Z"
+     "iopub.execute_input": "2020-11-12T20:52:00.976526Z",
+     "iopub.status.busy": "2020-11-12T20:52:00.974795Z",
+     "iopub.status.idle": "2020-11-12T20:52:01.376868Z",
+     "shell.execute_reply": "2020-11-12T20:52:01.377183Z"
     }
    },
    "outputs": [
@@ -568,13 +568,13 @@
        "    time_coverage_end:             2016-09-20T13:32:33Z\n",
        "    time_coverage_start:           2016-09-02T17:05:38Z\n",
        "    title:                         whoi_406-20160902T1700\n",
[…]
