+++ master/_sources/00-quick_intro-output.ipynb.txt
@@ -34,10 +34,10 @@
    "execution_count": 1,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2020-11-12T02:16:06.631005Z",
-     "iopub.status.busy": "2020-11-12T02:16:06.630304Z",
-     "iopub.status.idle": "2020-11-12T02:16:06.982253Z",
-     "shell.execute_reply": "2020-11-12T02:16:06.981393Z"
+     "iopub.execute_input": "2020-11-12T20:50:49.192702Z",
+     "iopub.status.busy": "2020-11-12T20:50:49.192020Z",
+     "iopub.status.idle": "2020-11-12T20:50:49.521256Z",
+     "shell.execute_reply": "2020-11-12T20:50:49.520446Z"
     }
    },
    "outputs": [],
@@ -65,10 +65,10 @@
    "execution_count": 2,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2020-11-12T02:16:06.988911Z",
-     "iopub.status.busy": "2020-11-12T02:16:06.986933Z",
-     "iopub.status.idle": "2020-11-12T02:16:07.216728Z",
-     "shell.execute_reply": "2020-11-12T02:16:07.217066Z"
+     "iopub.execute_input": "2020-11-12T20:50:49.526244Z",
+     "iopub.status.busy": "2020-11-12T20:50:49.525839Z",
+     "iopub.status.idle": "2020-11-12T20:50:50.064410Z",
+     "shell.execute_reply": "2020-11-12T20:50:50.064049Z"
     }
    },
    "outputs": [
