### 2020-11-19 f6a635b
changes to allow relative constraining
Can now use relative calculations on time and other variable’s when constraining data
