+++ tests/test_url_builder.py
@@ -194,14 +194,14 @@ def test_download_url_distinct(e):
 
 # Test generic sever-side functions
 def test_download_url_server_functions(e):
-    """Check download URL results with and without the distinct option."""
+    """Check download URL results with and without the relative constraint option."""
     dataset_id = "gtoppAT"
     variables = ["commonName", "yearDeployed", "serialNumber"]
-    no_distinct_url = e.get_download_url(dataset_id=dataset_id, variables=variables)
-    with_distinct_url = e.get_download_url(
+    no_relative_constraints_url = e.get_download_url(dataset_id=dataset_id, variables=variables)
+    with_relative_constraints_url = e.get_download_url(
         dataset_id=dataset_id,
         variables=variables,
         relative_constraints={"time=": "max(time)+0minutes"},
     )
-    assert not no_distinct_url.endswith("&time=max(time)+0minutes")
-    assert with_distinct_url.endswith("&time=max(time)+0minutes")
+    assert not no_relative_constraints_url.endswith("&time=max(time)+0minutes")
+    assert with_relative_constraints_url.endswith("&time=max(time)+0minutes")
