### 2020-11-28 5c17b90
Update docs after building Travis build 408 of ioos/erddapy
The docs were built from the branch 'master' against the commit
e8bbca80a7bca811f22fa02260dbca0cae377fc0.

The Travis build that generated this commit is at
https://travis-ci.org/ioos/erddapy/jobs/746504557.

The doctr command that was run is

    /home/travis/micromamba/envs/TEST/bin/doctr deploy --command=doctr-versions-menu --no-require-master --build-tags master
