+++ master/_sources/00-quick_intro-output.ipynb.txt
@@ -34,10 +34,10 @@
    "execution_count": 1,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2020-11-28T22:16:43.247110Z",
-     "iopub.status.busy": "2020-11-28T22:16:43.246424Z",
-     "iopub.status.idle": "2020-11-28T22:16:43.696568Z",
-     "shell.execute_reply": "2020-11-28T22:16:43.696911Z"
+     "iopub.execute_input": "2020-11-28T23:47:39.260943Z",
+     "iopub.status.busy": "2020-11-28T23:47:39.260173Z",
+     "iopub.status.idle": "2020-11-28T23:47:39.606026Z",
+     "shell.execute_reply": "2020-11-28T23:47:39.606389Z"
     }
    },
    "outputs": [],
@@ -65,10 +65,10 @@
    "execution_count": 2,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2020-11-28T22:16:43.703985Z",
-     "iopub.status.busy": "2020-11-28T22:16:43.702043Z",
-     "iopub.status.idle": "2020-11-28T22:16:43.973313Z",
-     "shell.execute_reply": "2020-11-28T22:16:43.973698Z"
+     "iopub.execute_input": "2020-11-28T23:47:39.613702Z",
+     "iopub.status.busy": "2020-11-28T23:47:39.613267Z",
+     "iopub.status.idle": "2020-11-28T23:47:39.971268Z",
+     "shell.execute_reply": "2020-11-28T23:47:39.970763Z"
     }
    },
    "outputs": [
