+++ master/_sources/02-extras-output.ipynb.txt
@@ -14,10 +14,10 @@
    "execution_count": 1,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2020-11-28T22:17:05.251045Z",
-     "iopub.status.busy": "2020-11-28T22:17:05.250341Z",
-     "iopub.status.idle": "2020-11-28T22:17:05.589703Z",
-     "shell.execute_reply": "2020-11-28T22:17:05.589100Z"
+     "iopub.execute_input": "2020-11-28T23:48:14.732771Z",
+     "iopub.status.busy": "2020-11-28T23:48:14.732052Z",
+     "iopub.status.idle": "2020-11-28T23:48:15.083409Z",
+     "shell.execute_reply": "2020-11-28T23:48:15.082549Z"
     }
    },
    "outputs": [],
@@ -33,10 +33,10 @@
    "execution_count": 2,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2020-11-28T22:17:05.598616Z",
-     "iopub.status.busy": "2020-11-28T22:17:05.593435Z",
-     "iopub.status.idle": "2020-11-28T22:17:05.906770Z",
-     "shell.execute_reply": "2020-11-28T22:17:05.906333Z"
+     "iopub.execute_input": "2020-11-28T23:48:15.087649Z",
+     "iopub.status.busy": "2020-11-28T23:48:15.086941Z",
+     "iopub.status.idle": "2020-11-28T23:48:15.402527Z",
+     "shell.execute_reply": "2020-11-28T23:48:15.402012Z"
     }
    },
    "outputs": [
@@ -78,10 +78,10 @@
    "execution_count": 3,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2020-11-28T22:17:05.912271Z",
-     "iopub.status.busy": "2020-11-28T22:17:05.910297Z",
-     "iopub.status.idle": "2020-11-28T22:17:09.008280Z",
-     "shell.execute_reply": "2020-11-28T22:17:09.007897Z"
+     "iopub.execute_input": "2020-11-28T23:48:15.408354Z",
+     "iopub.status.busy": "2020-11-28T23:48:15.406270Z",
+     "iopub.status.idle": "2020-11-28T23:48:26.153132Z",
+     "shell.execute_reply": "2020-11-28T23:48:26.153452Z"
     },
     "scrolled": false
    },
@@ -137,10 +137,10 @@
    "execution_count": 4,
    "metadata": {
     "execution": {
-     "iopub.execute_input": "2020-11-28T22:17:09.012901Z",
-     "iopub.status.busy": "2020-11-28T22:17:09.011177Z",
-     "iopub.status.idle": "2020-11-28T22:17:09.428214Z",
-     "shell.execute_reply": "2020-11-28T22:17:09.428555Z"
+     "iopub.execute_input": "2020-11-28T23:48:26.157091Z",
+     "iopub.status.busy": "2020-11-28T23:48:26.156435Z",
+     "iopub.status.idle": "2020-11-28T23:48:26.579541Z",
+     "shell.execute_reply": "2020-11-28T23:48:26.579955Z"
     }
    },
    "outputs": [
@@ -568,13 +568,13 @@
        "    time_coverage_end:             2016-09-20T13:32:33Z\n",
        "    time_coverage_start:           2016-09-02T17:05:38Z\n",
        "    title:                         whoi_406-20160902T1700\n",
[…]
