+++ docs/source/conf.py
@@ -61,7 +61,6 @@ author = "Filipe Fernandes"
 #
 import erddapy
 
-
 version = release = erddapy.__version__
 
 # The language for content autogenerated by Sphinx. Refer to documentation
