+++ .pre-commit-config.yaml
@@ -1,6 +1,6 @@
 repos:
 - repo: https://github.com/pre-commit/pre-commit-hooks
-  rev: v3.3.0
+  rev: v3.4.0
   hooks:
     - id: trailing-whitespace
     - id: check-ast
@@ -22,11 +22,11 @@ repos:
       args: [--max-line-length=105]
 
 - repo: https://github.com/pycqa/isort
-  rev: 5.6.4
+  rev: 5.7.0
   hooks:
     - id: isort
       additional_dependencies: [toml]
-      args: [--project=erddapy, --multi-line=3, --lines-after-imports=2, --lines-between-types=1, --trailing-comma, --force-grid-wrap=0, --use-parentheses, --line-width=88]
+      args: ["--profile", "black", "--filter-files"]
 
 - repo: https://github.com/psf/black
   rev: 20.8b1
@@ -35,14 +35,14 @@ repos:
     language_version: python3
 
 - repo: https://github.com/pre-commit/mirrors-mypy
-  rev: v0.790
+  rev: v0.812
   hooks:
   - id: mypy
     exclude: docs/source/conf.py
     args: [--ignore-missing-imports]
 
 - repo: https://github.com/keewis/blackdoc
-  rev: v0.3
+  rev: v0.3.3
   hooks:
     - id: blackdoc
 
@@ -65,13 +65,13 @@ repos:
         - --quiet-level=2
 
 - repo: https://github.com/asottile/pyupgrade
-  rev: v2.7.4
+  rev: v2.10.0
   hooks:
     - id: pyupgrade
       args:
         - --py36-plus
 
 - repo: https://github.com/asottile/add-trailing-comma
-  rev: v2.0.1
+  rev: v2.1.0
   hooks:
     - id: add-trailing-comma
