### 2021-03-11 4a912f9
refactor constraint url building into function
