+++ .github/workflows/tests.yml
@@ -35,4 +35,4 @@ jobs:
       shell: bash -l {0}
       run: |
         source activate TEST
-        pytest -n 2 -rxs --cov=erddapy --vcr-record=none tests
+        pytest -rxs --cov=erddapy --vcr-record=none tests
