+++ erddapy/erddapy_url_search.py
@@ -1,0 +1,133 @@
+import copy
+import functools
+from datetime import datetime
+from typing import Dict,List,Optional,Tuple,Union
+from urllib.parse import quote_plus
+import pandas as pd
+import pytz
+from erddapy.netcdf_handling import _nc_dataset,_tempnc
+from erddapy.servers import servers
+from erddapy.url_handling import _distinct,urlopen
+try:
+    from pandas.core.indexes.period import parse_time_string
+except ImportError:
+    from pandas._libs.tslibs.parsing import parse_time_string
+
+ListLike = Union[List[str],Tuple[str]]
+OptionalStr = Optional[str]
+
+def _quote_string_constraints(kwargs:Dict)->Dict:
+    return {k:f'"{v}"' if isinstance(v,str) else v for k,v in kwargs.items()}
+    
+def _format_constraints_url(kwargs:Dict)->str:
+    return "".join([f"&{k}{v}" for k,v in kwargs.items()])
+
+def parse_dates(date_time: Union[datetime,str])->float:
+    if isinstance(date_time,str):
+        parse_date_time = parse_time_string(date_time)[0]
+    else:
+        parse_date_time = date_time
+
+    if not parse_date_time.tzinfo:
+        parse_date_time = pytz.utc.localize(parse_date_time)
+    else:
+        parse_date_time = parse_date_time.astimezone(pytz.utc)
+    return parse_date_time.timestamp()
+
+class ERDDAP:
+    def __init__(
+        self,
+        server: str,
+        protocol: OptionalStr = None,
+        response: str = "html",
+    ):
+        if server in servers.keys():
+            server = servers[server].url
+        self.server = server.rstrip("/")
+        self.protocol = protocol
+        self.response = response
+
+        self.constraints: Optional[Dict] = None
+        self.relative_constraints: Optional[Dict] = None
+        self.server_functions: Optional[Dict] = None
+        self.dataset_id: OptionalStr = None
+        self.requests_kwargs: Dict = {}
+        self.auth: Optional[tuple] = None
+        self.variables: Optional[ListLike] = None
+
+        self._dataset_id: OptionalStr = None
+        self._variables: Dict = {}
+
+    def get_search_url(
+        self,
+        response: OptionalStr = None,
+        search_for: OptionalStr = None,
+        protocol: OptionalStr = None,
+        items_per_page: int = 1000,
+        page: int = 1,
+        **kwargs,
+    ) -> str:
+
+        base = (
+            "{server}/{protocol}/{dataset_id}.{response}"
+            "?cdm_data_type={cdm_data_type}"
+            "&institution={institution}"
+            "&ioos_category={ioos_category}"
+            "&keywords={keywords}"
+            "&long_name={long_name}"
+            "&standard_name={standard_name}"
+            "&variableName={variableName}"
+            "&Lat={Lat}"
+            "&Lon={Lon}"
+            "&Time={Time}"
+        )
+        if search_for:
+            search_for = quote_plus(search_for)
+            base += "&searchFor={searchFor}"
+
+        # Convert dates from datetime to `seconds since 1970-01-01T00:00:00Z`.
+        Time = kwargs.pop("Time", None)
+        if Time:
+            kwargs.update({"Time": parse_dates(Time)})
+        
+        protocol = protocol if protocol else self.protocol
+        response = response if response else self.response
+        if protocol:
+            kwargs.update({"protocol": protocol})
+
+        lower_case_search_terms = (
+            "cdm_data_type",
+            "institution",
+            "ioos_category",
+            "keywords",
+            "long_name",
+            "standard_name",
+            "variableName",
+        )
+        for search_term in lower_case_search_terms:
+            if search_term in kwargs.keys():
+                lowercase = kwargs[search_term].lower()
[…]
