### 2021-04-08 7c555ad
add list of projects that use erddapy and that are similiar to it
