### 2021-06-15 2c9cd28
using requests in place of urllib
