### 2021-06-19 c7ed90a
sanity check griddap variables and constraints
