### 2021-06-21 60e2166
escape griddap url build when constraints and variables are none
