### 2021-06-22 470c660
do not print griddap metadata
