### 2021-06-22 1583f69
use griddap default subsetting
