### 2021-06-22 d2a73af
Improve docstring
  - Improve docstring for 'get_categorize_url()' function.
  - Add a line explaining that valid attributes can be found in the categorize page.
