+++ erddapy/erddapy.py
@@ -306,6 +306,7 @@ class ERDDAP:
 
         Args:
             categorize_by: a valid attribute, e.g.: ioos_category or standard_name.
+                Valid attributes are shown in the '<YOUR-ERDDAP-URL>/erddap/categorize' page.
             value: an attribute value.
             response: default is HTML.
 
