### 2021-06-22 04164ef
use older mypy for now to avoid stub issue
