+++ .pre-commit-config.yaml
@@ -1,6 +1,6 @@
 repos:
 - repo: https://github.com/pre-commit/pre-commit-hooks
-  rev: v3.4.0
+  rev: v4.0.1
   hooks:
     - id: trailing-whitespace
     - id: check-ast
@@ -15,26 +15,27 @@ repos:
       files: requirements-dev.txt
 
 - repo: https://gitlab.com/pycqa/flake8
-  rev: 3.8.4
+  rev: 3.9.2
   hooks:
     - id: flake8
       exclude: docs/source/conf.py
       args: [--max-line-length=105]
 
 - repo: https://github.com/pycqa/isort
-  rev: 5.7.0
+  rev: 5.9.1
   hooks:
     - id: isort
       additional_dependencies: [toml]
       args: ["--profile", "black", "--filter-files"]
 
 - repo: https://github.com/psf/black
-  rev: 20.8b1
+  rev: 21.6b0
   hooks:
   - id: black
     language_version: python3
 
 - repo: https://github.com/pre-commit/mirrors-mypy
+  #rev: v0.902
   rev: v0.812
   hooks:
   - id: mypy
@@ -47,14 +48,14 @@ repos:
     - id: blackdoc
 
 - repo: https://github.com/econchick/interrogate
-  rev: 1.3.2
+  rev: 1.4.0
   hooks:
     - id: interrogate
       exclude: ^(docs|setup.py|tests)
       args: [--config=pyproject.toml]
 
 - repo: https://github.com/codespell-project/codespell
-  rev: v2.0.0
+  rev: v2.1.0
   hooks:
     - id: codespell
       exclude: >
@@ -65,7 +66,7 @@ repos:
         - --quiet-level=2
 
 - repo: https://github.com/asottile/pyupgrade
-  rev: v2.10.0
+  rev: v2.19.4
   hooks:
     - id: pyupgrade
       args:
@@ -75,3 +76,8 @@ repos:
   rev: v2.1.0
   hooks:
     - id: add-trailing-comma
+
+- repo: https://github.com/deathbeds/prenotebook
+  rev: f5bdb72a400f1a56fe88109936c83aa12cc349fa
+  hooks:
+    - id: prenotebook
