### 2021-06-23 535d208
fix type output of _griddap_get_constraints
