### 2021-06-23 a97dac1
will fix self.dim_names later
