### 2021-06-23 9b5bee1
configure pre-commit ci PRs
