+++ .pre-commit-config.yaml
@@ -81,3 +81,15 @@ repos:
   rev: f5bdb72a400f1a56fe88109936c83aa12cc349fa
   hooks:
     - id: prenotebook
+
+
+ci:
+    autofix_commit_msg: |
+        [pre-commit.ci] auto fixes from pre-commit.com hooks
+
+        for more information, see https://pre-commit.ci
+    autofix_prs: false
+    autoupdate_commit_msg: '[pre-commit.ci] pre-commit autoupdate'
+    autoupdate_schedule: monthly
+    skip: []
+    submodules: false
