### 2021-06-25 c905231
Use CoastWatch URL for docstring
