### 2021-06-27 5608ea3
case insensitive server name matching
