### 2021-06-27 a07901c
fix dim_names type error with check for None
