### 2021-06-27 e2959d8
check for dataset_id in griddap_initialize
