### 2021-06-28 e8d70dd
accept dataset_id or self.dataset_id
