### 2021-06-28 4d54d97
lower keys of server dict at creation
