+++ erddapy/servers.py
@@ -28,11 +28,10 @@ def servers_list():
         df = pd.read_json(path.joinpath("erddaps.json"))
     _server = namedtuple("server", ["description", "url"])
     return {
-        row["short_name"]: _server(row["name"], row["url"])
+        row["short_name"].lower(): _server(row["name"], row["url"])
         for k, row in df.iterrows()
         if row["short_name"]
     }
 
 
-servers_upper_case = servers_list()
-servers = {k.lower(): v for k, v in servers_upper_case.items()}
+servers = servers_list()
