### 2021-06-30 4f85f1e
default stride to 1 in griddap
