### 2021-07-02 6076396
update conda, use python -m, and debug
