+++ .github/workflows/publish.yml
@@ -22,7 +22,7 @@ jobs:
 
     - name: Install build tools
       run: |
-        python -m pip install --upgrade pip wheel setuptools setuptools_scm build twine pytest pandas pytz requests
+        python -m pip install --upgrade pip wheel setuptools setuptools_scm build twine
 
       shell: bash
 
