+++ .github/workflows/tests.yml
@@ -17,7 +17,8 @@ jobs:
     - uses: actions/checkout@v2
 
     - name: Setup Conda
-      uses: s-weigand/setup-conda@v1.0.5
+      #uses: s-weigand/setup-conda@v1.0.5
+      uses: s-weigand/setup-conda@fix-env-activation
       with:
         activate-conda: false
         update-conda: true
