### 2021-07-05 db20487
Update .github/workflows/tests.yml
Co-authored-by: Sebastian Weigand <s.weigand.phy@gmail.com>
