+++ .github/workflows/tests.yml
@@ -17,8 +17,7 @@ jobs:
     - uses: actions/checkout@v2
 
     - name: Setup Conda
-      #uses: s-weigand/setup-conda@v1.0.5
-      uses: s-weigand/setup-conda@fix-env-activation-dist
+      uses: s-weigand/setup-conda@v1.0.5
       with:
         activate-conda: false
         update-conda: true
