### 2021-07-05 cebae90
add griddap notebook to sphinx
