+++ .travis.yml
@@ -55,7 +55,8 @@ script:
     if [[ $TRAVIS_JOB_NAME == 'docs' ]]; then
       set -e
       travis_wait jupyter nbconvert --to notebook --execute notebooks/00-quick_intro.ipynb --output=00-quick_intro-output.ipynb
-      travis_wait jupyter nbconvert --to notebook --execute notebooks/01-longer_intro.ipynb --output=01-longer_intro-output.ipynb
+      travis_wait jupyter nbconvert --to notebook --execute notebooks/01a-griddap.ipynb --output=01a-griddap-output.ipynb
+      travis_wait jupyter nbconvert --to notebook --execute notebooks/01b-tabledap.ipynb --output=01b-tabledap.ipynb
       travis_wait jupyter nbconvert --to notebook --execute notebooks/02-extras.ipynb --output=02-extras-output.ipynb
       travis_wait jupyter nbconvert --to notebook --execute notebooks/03-searchfor.ipynb --output=03-searchfor-output.ipynb
       mv notebooks/*output.ipynb docs/source/
