+++ .pre-commit-config.yaml
@@ -35,12 +35,12 @@ repos:
     language_version: python3
 
 - repo: https://github.com/pre-commit/mirrors-mypy
-  #rev: v0.902
-  rev: v0.812
+  rev: v0.910
   hooks:
   - id: mypy
     exclude: docs/source/conf.py
-    args: [--ignore-missing-imports]
+    entry: bash -c 'exec env CONDA_PREFIX="$(python -c "import sys; print(sys.executable)")"'
+    args: ["--ignore-missing-imports", "--python-executable=CONDA_PREFIX"]
 
 - repo: https://github.com/keewis/blackdoc
   rev: v0.3.3
@@ -66,7 +66,7 @@ repos:
         - --quiet-level=2
 
 - repo: https://github.com/asottile/pyupgrade
-  rev: v2.19.4
+  rev: v2.20.0
   hooks:
     - id: pyupgrade
       args:
