+++ .pre-commit-config.yaml
@@ -82,7 +82,6 @@ repos:
   hooks:
     - id: prenotebook
 
-
 ci:
     autofix_commit_msg: |
         [pre-commit.ci] auto fixes from pre-commit.com hooks
