### 2021-07-05 56cc47f
[pre-commit.ci] pre-commit autoupdate
updates:
- https://gitlab.com/pycqa/flake8 → https://github.com/PyCQA/flake8
- [github.com/pre-commit/mirrors-mypy: v0.812 → v0.910](https://github.com/pre-commit/mirrors-mypy/compare/v0.812...v0.910)
- [github.com/asottile/pyupgrade: v2.19.4 → v2.20.0](https://github.com/asottile/pyupgrade/compare/v2.19.4...v2.20.0)
