+++ .pre-commit-config.yaml
@@ -14,7 +14,7 @@ repos:
     - id: file-contents-sorter
       files: requirements-dev.txt
 
-- repo: https://gitlab.com/pycqa/flake8
+- repo: https://github.com/PyCQA/flake8
   rev: 3.9.2
   hooks:
     - id: flake8
@@ -36,7 +36,7 @@ repos:
 
 - repo: https://github.com/pre-commit/mirrors-mypy
   #rev: v0.902
-  rev: v0.812
+  rev: v0.910
   hooks:
   - id: mypy
     exclude: docs/source/conf.py
@@ -66,7 +66,7 @@ repos:
         - --quiet-level=2
 
 - repo: https://github.com/asottile/pyupgrade
-  rev: v2.19.4
+  rev: v2.20.0
   hooks:
     - id: pyupgrade
       args:
