### 2021-07-19 4ab1e5a
pull search_all_servers out of ERDDAP class
