+++ erddapy/erddapy.py
@@ -23,7 +23,6 @@ try:
 except ImportError:
     from pandas._libs.tslibs.parsing import parse_time_string
 
-
 ListLike = Union[List[str], Tuple[str]]
 OptionalStr = Optional[str]
 
@@ -152,6 +151,34 @@ def parse_results(url: str, key: str, protocol="tabledap") -> Optional[Dict]:
         return None
     df["Server url"] = url.split("search")[0]
     return {key: df[["Title", "Institution", "Dataset ID", "Server url"]]}
+
+
+def search_all_servers(query="glider", servers_list=None, protocol="tabledap"):
+    """
+    Search all servers for a query string
+    Returns a dataframe of details for all matching datasets
+    Args:
+        query: string to search for
+        servers_list: optional list of servers. Defaults to searching all servers
+    """
+    if servers_list:
+        urls = {
+            server: f'{server}search/index.csv?page=1&itemsPerPage=100000&searchFor="{query}"'
+            for server in servers_list
+        }
+    else:
+        urls = {
+            key: f'{server.url}search/index.csv?page=1&itemsPerPage=100000&searchFor="{query}"'
+            for key, server in servers.items()
+        }
+    num_cores = multiprocessing.cpu_count()
+    returns = Parallel(n_jobs=num_cores)(
+        delayed(parse_results)(url, key, protocol=protocol) for key, url in urls.items()
+    )
+    dfs = [x for x in returns if x is not None]
+    df_all = pd.concat([list(df.values())[0] for df in dfs])
+    df_all.reset_index(drop=True, inplace=True)
+    return df_all
 
 
 class ERDDAP:
@@ -388,34 +415,6 @@ class ERDDAP:
         # Removing them entirely should be OK for older versions too.
         url = url.replace("&minTime=(ANY)", "").replace("&maxTime=(ANY)", "")
         return url
-
-    def search_all_servers(self, query="glider", servers_list=None):
-        """
-        Search all servers for a query string
-        Returns a dataframe of details for all matching datasets
-        Args:
-            query: string to search for
-            servers_list: optional list of servers. Defaults to searching all servers
-        """
-        if servers_list:
-            urls = {
-                server: f'{server}search/index.csv?page=1&itemsPerPage=100000&searchFor="{query}"'
-                for server in servers_list
-            }
-        else:
-            urls = {
-                key: f'{server.url}search/index.csv?page=1&itemsPerPage=100000&searchFor="{query}"'
-                for key, server in servers.items()
-            }
-        num_cores = multiprocessing.cpu_count()
-        returns = Parallel(n_jobs=num_cores)(
-            delayed(parse_results)(url, key, protocol="tabledap")
-            for key, url in urls.items()
-        )
-        dfs = [x for x in returns if x is not None]
-        df_all = pd.concat([list(df.values())[0] for df in dfs])
-        df_all.reset_index(drop=True, inplace=True)
-        return df_all
 
     def get_info_url(
         self,
