### 2021-07-23 949e4e0
make parallel a kwarg for all server search
