+++ erddapy/erddapy.py
@@ -153,14 +153,25 @@ def fetch_results(url: str, key: str, protocol="tabledap") -> Optional[Dict]:
     return {key: df[["Title", "Institution", "Dataset ID", "Server url"]]}
 
 
-def search_all_servers(query="glider", servers_list=None, protocol="tabledap"):
+def search_all_servers(
+    query="glider",
+    servers_list=None,
+    parallel=True,
+    protocol="tabledap",
+):
     """
     Search all servers for a query string
     Returns a dataframe of details for all matching datasets
     Args:
         query: string to search for
-        servers_list: optional list of servers. Defaults to searching all servers
+        servers_list: optional list of servers. if None, will search all servers in erddapy.servers
+        protocol: tabledap or griddap
+        parallel: If True, uses joblib to parallelize the search
     """
+    if protocol not in ["tabledap", "griddap"]:
+        raise ValueError(
+            f"Protocol must be tabledap or griddap, got {protocol}",
+        )
     if servers_list:
         urls = {
             server: f'{server}search/index.csv?page=1&itemsPerPage=100000&searchFor="{query}"'
@@ -171,11 +182,17 @@ def search_all_servers(query="glider", servers_list=None, protocol="tabledap"):
             key: f'{server.url}search/index.csv?page=1&itemsPerPage=100000&searchFor="{query}"'
             for key, server in servers.items()
         }
-    num_cores = multiprocessing.cpu_count()
-    returns = Parallel(n_jobs=num_cores)(
-        delayed(fetch_results)(url, key, protocol=protocol) for key, url in urls.items()
-    )
-    dfs = [x for x in returns if x is not None]
+    if parallel:
+        num_cores = multiprocessing.cpu_count()
+        returns = Parallel(n_jobs=num_cores)(
+            delayed(fetch_results)(url, key, protocol=protocol)
+            for key, url in urls.items()
+        )
+        dfs = [x for x in returns if x is not None]
+    else:
+        dfs = []
+        for key, url in urls.items():
+            dfs.append(fetch_results(url, key, protocol=protocol))
     df_all = pd.concat([list(df.values())[0] for df in dfs])
     df_all.reset_index(drop=True, inplace=True)
     return df_all
