### 2021-07-23 8faa0b4
search url generation in seperate function
