### 2021-07-23 e5641b7
seperate fetch and format of multi-server search
