### 2021-07-23 cf1d2a8
seperate multi server results formatting
