### 2021-07-23 ca4b024
refactor multi server search to seperate file
