### 2021-07-23 53ed965
added tests for new url_handling functions
