### 2021-07-23 74cbf0b
test for fetch_results
