+++ erddapy/multiple_server_search.py
@@ -3,8 +3,9 @@ from typing import Dict, Optional
 
 import pandas as pd
 from joblib import Parallel, delayed
-from servers import servers
 from url_handling import format_search_string, multi_urlopen
+
+from erddapy.servers import servers
 
 
 def parse_results(data: bytes, protocol, key, url) -> Dict[str, pd.DataFrame]:
