### 2021-07-26 e873e0e
Update erddapy/url_handling.py
Co-authored-by: Filipe <ocefpaf@gmail.com>
