+++ .pre-commit-config.yaml
@@ -22,14 +22,14 @@ repos:
       args: [--max-line-length=105]
 
 - repo: https://github.com/pycqa/isort
-  rev: 5.9.1
+  rev: 5.9.2
   hooks:
     - id: isort
       additional_dependencies: [toml]
       args: ["--profile", "black", "--filter-files"]
 
 - repo: https://github.com/psf/black
-  rev: 21.6b0
+  rev: 21.7b0
   hooks:
   - id: black
     language_version: python3
@@ -43,7 +43,7 @@ repos:
     args: ["--ignore-missing-imports", "--python-executable=CONDA_PREFIX"]
 
 - repo: https://github.com/keewis/blackdoc
-  rev: v0.3.3
+  rev: v0.3.4
   hooks:
     - id: blackdoc
 
@@ -66,7 +66,7 @@ repos:
         - --quiet-level=2
 
 - repo: https://github.com/asottile/pyupgrade
-  rev: v2.20.0
+  rev: v2.23.0
   hooks:
     - id: pyupgrade
       args:
