+++ erddapy/multiple_server_search.py
@@ -3,9 +3,9 @@ from typing import Dict, Optional
 
 import pandas as pd
 from joblib import Parallel, delayed
-from erddapy.url_handling import format_search_string, multi_urlopen
 
 from erddapy.servers import servers
+from erddapy.url_handling import format_search_string, multi_urlopen
 
 
 def parse_results(data: bytes, protocol, key, url) -> Dict[str, pd.DataFrame]:
