+++ .github/workflows/deploy-docs.yml
@@ -31,7 +31,7 @@ jobs:
         source activate TEST
         jupyter nbconvert --to notebook --execute notebooks/00-quick_intro.ipynb --output=00-quick_intro-output.ipynb
         jupyter nbconvert --to notebook --execute notebooks/01a-griddap.ipynb --output=01a-griddap-output.ipynb
-        jupyter nbconvert --to notebook --execute notebooks/01b-tabledap.ipynb --output=01b-tabledap.ipynb
+        jupyter nbconvert --to notebook --execute notebooks/01b-tabledap.ipynb --output=01b-tabledap-output.ipynb
         jupyter nbconvert --to notebook --execute notebooks/02-extras.ipynb --output=02-extras-output.ipynb
         jupyter nbconvert --to notebook --execute notebooks/03-searchfor.ipynb --output=03-searchfor-output.ipynb
         mv notebooks/*output.ipynb docs/source/
