+++ .pre-commit-config.yaml
@@ -66,7 +66,11 @@ repos:
         - --quiet-level=2
 
 - repo: https://github.com/asottile/pyupgrade
+<<<<<<< HEAD
   rev: v2.31.0
+=======
+  rev: v2.23.3
+>>>>>>> updated
   hooks:
     - id: pyupgrade
       args:
