### 2021-08-10 82cab72
reduce fetch depth
