+++ .github/workflows/deploy-docs.yml
@@ -2,14 +2,18 @@ name: Build and Deploy docs
 
 on:
   push:
-    branches: [master]
+    branches:
+      - master
 
 jobs:
   build-docs:
     runs-on: ubuntu-latest
 
     steps:
-    - uses: actions/checkout@v2
+    - name: checkout
+      uses: actions/checkout@v2
+      with:
+        fetch-depth: 0
 
     - name: Setup Conda
       uses: s-weigand/setup-conda@v1
