+++ .github/workflows/deploy-docs.yml
@@ -28,6 +28,10 @@ jobs:
         source activate TEST
         python -m pip install -e . --no-deps --force-reinstall
 
+    - name: Get the version
+      id: get_version
+      run: echo ::set-output name=VERSION::$(python setup.py --version)
+
     - name: Build documentation
       shell: bash -l {0}
       run: |
