+++ .github/workflows/deploy-docs.yml
@@ -48,9 +48,7 @@ jobs:
         popd
 
     - name: Deploy
-      uses: JamesIves/github-pages-deploy-action@3.7.1
+      uses: peaceiris/actions-gh-pages@v3.6.1
       with:
-        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-        BRANCH: gh-pages
-        FOLDER: webpage
-        CLEAN: false
+        github_token: ${{ secrets.GITHUB_TOKEN }}
+        publish_dir: docs/build/html
