+++ _sources/00-quick_intro-output.ipynb.txt
@@ -1,0 +1,249 @@
+{
+ "cells": [
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "# Quick introduction\n",
+    "\n",
+    "erddapy is a pure python package and can be installed with conda\n",
+    "\n",
+    "```shell\n",
+    "conda install --channel conda-forge erddapy\n",
+    "```\n",
+    "\n",
+    "or pip\n",
+    "\n",
+    "```shell\n",
+    "pip install erddapy\n",
+    "```\n"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "First we need to instantiate the ERDDAP URL constructor for a server. In this\n",
+    "example we will use\n",
+    "[https://gliders.ioos.us/erddap](https://gliders.ioos.us/erddap/index.html).\n"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 1,
+   "metadata": {
+    "execution": {
+     "iopub.execute_input": "2021-08-16T17:25:28.046265Z",
+     "iopub.status.busy": "2021-08-16T17:25:28.045678Z",
+     "iopub.status.idle": "2021-08-16T17:25:28.965466Z",
+     "shell.execute_reply": "2021-08-16T17:25:28.964984Z"
+    }
+   },
+   "outputs": [],
+   "source": [
+    "from erddapy import ERDDAP\n",
+    "\n",
+    "\n",
+    "e = ERDDAP(\n",
+    "    server=\"NGDAC\",  # \"NOAA IOOS NGDAC (National Glider Data Assembly Center)\"\n",
+    "    protocol=\"tabledap\",\n",
+    "    response=\"csv\",\n",
+    ")"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "Now we can populate the object a dataset id, variables of interest, and its\n",
+    "constraints. We can download the csvp response with the `.to_pandas` method.\n"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 2,
+   "metadata": {
+    "execution": {
+     "iopub.execute_input": "2021-08-16T17:25:28.970807Z",
+     "iopub.status.busy": "2021-08-16T17:25:28.970303Z",
+     "iopub.status.idle": "2021-08-16T17:25:30.608780Z",
+     "shell.execute_reply": "2021-08-16T17:25:30.609233Z"
+    }
+   },
+   "outputs": [
+    {
+     "data": {
+      "text/html": [
+       "<div>\n",
+       "<style scoped>\n",
+       "    .dataframe tbody tr th:only-of-type {\n",
+       "        vertical-align: middle;\n",
+       "    }\n",
+       "\n",
+       "    .dataframe tbody tr th {\n",
+       "        vertical-align: top;\n",
+       "    }\n",
+       "\n",
+       "    .dataframe thead th {\n",
+       "        text-align: right;\n",
+       "    }\n",
+       "</style>\n",
+       "<table border=\"1\" class=\"dataframe\">\n",
+       "  <thead>\n",
+       "    <tr style=\"text-align: right;\">\n",
+       "      <th></th>\n",
+       "      <th>depth (m)</th>\n",
+       "      <th>latitude (degrees_north)</th>\n",
+       "      <th>longitude (degrees_east)</th>\n",
+       "      <th>salinity (1)</th>\n",
+       "      <th>temperature (Celsius)</th>\n",
+       "    </tr>\n",
+       "    <tr>\n",
+       "      <th>time (UTC)</th>\n",
+       "      <th></th>\n",
+       "      <th></th>\n",
[…]
