+++ v0.8.0/_sources/erddapy.rst.txt
@@ -1,7 +1,0 @@
-:mod:`ERDDAP class API`
------------------------
-
-.. automodule:: erddapy.erddapy
-   :members:
-   :undoc-members:
-   :show-inheritance:

[…]
