+++ erddapy/url_handling.py
@@ -33,23 +33,6 @@ def urlopen(url: str, auth: Optional[tuple] = None, **kwargs: Dict) -> BinaryIO:
     return data
 
 
-def multi_urlopen(url: str) -> BinaryIO:
-    """
-    A more simple url open to work with multiprocessing
-    """
-    try:
-        response = requests.get(url, allow_redirects=True)
-    except requests.exceptions.ConnectionError:
-        return None
-    try:
-        response.raise_for_status()
-    except requests.exceptions.HTTPError:
-        return None
-    data = io.BytesIO(response.content)
-    data.seek(0)
-    return data
-
-
 @functools.lru_cache(maxsize=None)
 def check_url_response(url: str, **kwargs: Dict) -> str:
     """
@@ -89,10 +72,3 @@ def _distinct(url: str, **kwargs: Dict) -> str:
     if distinct is True:
         return f"{url}&distinct()"
     return url
-
-
-def format_search_string(server: str, query: str) -> str:
-    """
-    Generate a search string for an erddap server with user defined query
-    """
-    return f'{server}search/index.csv?page=1&itemsPerPage=100000&searchFor="{query}"'
