+++ tests/cassettes/test_fetch_results.yaml
@@ -218,4 +218,358 @@ interactions:
     status:
       code: 200
       message: ''
+- request:
+    body: null
+    headers:
+      Accept:
+      - '*/*'
+      Accept-Encoding:
+      - gzip, deflate, br
+      Connection:
+      - keep-alive
+      User-Agent:
+      - python-requests/2.26.0
+    method: GET
+    uri: https://gliders.ioos.us/erddap/search/index.csv?page=1&itemsPerPage=100000&searchFor=%22sst%22
+  response:
+    body:
+      string: !!binary |
+        H4sIAAAAAAAAAO2de28bSbLl/7+fojDA9rgB2a4qkiJ5gcHCY/f49sW43dvm9gCLxgolqiRxm2Rx
+        +LBb99NvZD3IVGaG2xUZvDMaH2DGrUfVyeCrzk+RURF328XNTbG5+HC43pX7i31xvSzN9++KX8vk
+        VfJ2W2zuLz6tdhe3i2W5u5gt9suSDl6tiu3DxV/evnl98f2H90k2zbLRxffr2+riz8X817ttdVjf
+        JPX3P334cPHdqlgs6de7/WJ/2C+q9cWbYl/Qesn3b/7t4n6/3+z+/eXLu+XiptzuXiyqavfisHtZ
+        bk1kL7uQXm4P+eR5nuZplqbTGS04erFropYrxCx+Vz83v6dQP2+B019eBKL5w5tys6weVuV6n1S3
+        SZF8WFbzwypptJN9lWzK7W21XSW7sthV62KZ7A7bj+XDzhx+s9jtquXH8iapfnu4K9fJvFrPSWpb
+        mOd8lyzWyf6+THb3xXJZfaLfFrs9KXwq9hS3Efih/JT8J31dPrxIktn9YpfcnOKh7zbFto7LqJyO
+        TSho+kUX9Hfrj4tttTbfkviP22pfzs36f9wlD2WxXT54CzfBrqr1Yl9tF+u7pLylx7jfdUG0j35e
+        bLcLenBF8nr25iL5UBbXi+1N8mG+oKUWt4t58t3r98l+u9gs6Z11uzxU23JHj39ePr+mN+VuXuxp
+        vefz+2W1rTb3D8slSe3K9a7aGrW/H6pFYav99fsPH2ZJnqa/1Yv9vKBH/Kn4WB7P+blczavkXXVN
+        L3Ey2xZrWm3xkSIt6M3/iv4pt0XhvyrN6S9+Wf+ynt+srm7oo3C1f9iUyZ+MyP+jZ6vaPtDzZt45
+        v6x/fvXT96/+/NfvPvz7L+aVbH+dPDsdmvxQrMpvf1l/WlVXi5tf1pvm1Prr/WJVJs9asWRG313Q
+        +vS+uNkluwU9NfTRHafP04z+N0vTf6//939IbFmYz+qNdfJf259c0JvibluWu6s1vUj35thqfece
+        3P3odDS9Yfd0ML2j9vfJsxV9eV3SIx+n6XqVPPszfZ00Xxf7JMvG3XkXyep5luy2z7NvzfNV0bGb
[…]

[…]
