+++ .github/workflows/publish.yml
@@ -1,9 +1,6 @@
 name: Publish to PyPI
 
-on:
-  release:
-    types:
-      - published
+on: ["push", "pull_request"]
 
 jobs:
   packages:
@@ -14,7 +11,7 @@ jobs:
     - name: Set up Python
       uses: actions/setup-python@v1
       with:
-        python-version: 3.8
+        python-version: 3.x
 
     - name: Get tags
       run: git fetch --depth=1 origin +refs/tags/*:refs/tags/*
@@ -41,6 +38,7 @@ jobs:
       shell: bash
 
     - name: Publish a Python distribution to PyPI
+      if: contains(github.ref, "release")
       uses: pypa/gh-action-pypi-publish@master
       with:
         user: __token__
