### 2021-09-16 5212a4f
filter non-public servers
