+++ tests/cassettes/test_to_pandas.yaml
@@ -5,11 +5,11 @@ interactions:
       Accept:
       - '*/*'
       Accept-Encoding:
-      - gzip, deflate
+      - gzip, deflate, br
       Connection:
       - keep-alive
       User-Agent:
-      - python-requests/2.24.0
+      - python-requests/2.26.0
     method: GET
     uri: https://upwell.pfeg.noaa.gov/erddap/tabledap/pmelTao5dayIso.csvp?ISO_6,time&time%3E=248011200.0&time%3C=1548504000.0&latitude%3E=10&latitude%3C=10&longitude%3E=265&longitude%3C=265
   response:
@@ -23,13 +23,15 @@ interactions:
       Content-Type:
       - text/plain; charset=UTF-8
       Date:
-      - Sun, 04 Oct 2020 15:45:10 GMT
+      - Thu, 16 Sep 2021 23:48:03 GMT
       Last-Modified:
-      - Sun, 04 Oct 2020 15:45:10 GMT
+      - Thu, 16 Sep 2021 23:48:03 GMT
       Location:
       - https://coastwatch.pfeg.noaa.gov/erddap/tabledap/pmelTao5dayIso.csvp?ISO_6%2Ctime&time%3E=248011200.0&time%3C=1548504000.0&latitude%3E=10&latitude%3C=10&longitude%3E=265&longitude%3C=265
       Strict-Transport-Security:
       - max-age=31536000; includeSubDomains
+      X-Frame-Options:
+      - SAMEORIGIN
       erddap-server:
       - '2.02'
       xdods-server:
@@ -43,57 +45,58 @@ interactions:
       Accept:
       - '*/*'
       Accept-Encoding:
-      - gzip, deflate
+      - gzip, deflate, br
       Connection:
       - keep-alive
       User-Agent:
-      - python-requests/2.24.0
+      - python-requests/2.26.0
     method: GET
     uri: https://coastwatch.pfeg.noaa.gov/erddap/tabledap/pmelTao5dayIso.csvp?ISO_6%2Ctime&time%3E=248011200.0&time%3C=1548504000.0&latitude%3E=10&latitude%3C=10&longitude%3E=265&longitude%3C=265
   response:
     body:
       string: !!binary |
-        H4sIAAAAAAAAAHWZva4ttw2Fez+FSxu4dyCSEkmlTZXGKeI0aYIUKVK4y/sjS2fPiEMqF9jdB0ra
[…]
