+++ tests/cassettes/test_urlopen_raise.yaml
@@ -5,187 +5,162 @@ interactions:
       Accept:
       - '*/*'
       Accept-Encoding:
-      - gzip, deflate
+      - gzip, deflate, br
       Connection:
       - keep-alive
       User-Agent:
-      - python-requests/2.24.0
+      - python-requests/2.26.0
     method: GET
     uri: https://developer.mozilla.org/en-US/404
   response:
     body:
       string: !!binary |
-        H4sIAAAAAAAAAwAAAP//7Ftrd9s20v7uX4GyZy1ra5Iidbctta7ttukmaU7tvn13254ciIQkNiSh
-        EqRlx/Z/32fAiyjJ9knibvvFTkQRwMxgAMwVgHZ2do4+O/3h5OLfb87YPI3C8c4RfbGQx7ORIWLz
-        p3OD+UEyMsI0MZgXcqVGRizN35VBsIL7bJGIaXA1MuTsADTSxYFty9nCioQdq88BxdhRJFLOvDlP
-        lEhHRpZOzUGtgZBM8UcWXI6M/zd/OjZPZLTgaTAJBbqUcSpiYL04G535M5HjKS8JFul4b5rFXhrI
-        eM9vshvmW5rB1zwSbFQvWYlYhNwTe/avE828vc8ajeYhu2vu+dLLIvRglS9noaCv5uGRXXRDPaZB
-        Gorxa5myb2QW++yWvTp9fWTn1TvVWGL0NjIuA7FcyCStsb8M/HQ+8sVl4AlTF/ZZEAdpwENTeTwU
-        I6c2JTmZRE5kqmpEYhnEvrjaZ7GcyjCUS0N3vXMUBvE7lohwZGA1Qsl9g82xLCPDVilm0rOnIKHs
-        UFJPyv5PEIb8POQT80cxy0KeWCqbYG2sycRrt6eTTt+b9qylnE5dg3GsOKEbLL1eiPzdztvQN2Ne
-        IpWSSTALYmaPn8DL1zL0S0bEsOc53Wlv4PUGn8gIVdWYmctIVJxoybaNLaAQqxOrCu7zskyAO8Wn
-        RFgf1CQLQh+FaxpT5MdWy287YtgfdqfDjuUpLKPuIYeYC1ENIxVXqa0B7HFOXz8+oBuRJDIxO62O
-        5bf7LacvOtzz2x/eWdFXLkIaNq/7imbWnEJh2E3BC2NFXRSE1wfsPS2awqIdrrf7gYKiAUAt+WLV
-        phLvgGVJuPdkcWyinyTi6V5DFxvN/aoT/fcJnbTa3f7UcXtOm/d1J+t9wEysj3Epgtk8PYAOAijc
-        aNSzuN52t/M3T+rjevVnzGi9B5+3PM8XHY93+MdM5wQ0PmQyj+xCUkvJ1aqystSM2TZ7JfVkstPX
-        F+w7ES5Ekrf8k13MA8XOZZZgHU6kL2DOk4ihDuz/LryUpZKlc8FSkUSKyakulOTeZBNYBPYytwr7
-        7NJirtWy2Ispg3+Ti+sK4c1LtuQKrKdwnypNgkmWCp8tg3QOAHQ3DUIQ+LfMmMdjJicph9WSsWA8
[…]
