+++ tests/cassettes/test_fetch_results.yaml
@@ -5,225 +5,6 @@ interactions:
       Accept:
       - '*/*'
       Accept-Encoding:
-      - gzip, deflate
-      Connection:
-      - keep-alive
-      User-Agent:
-      - python-requests/2.25.1
-    method: GET
-    uri: https://data.ioos.us/gliders/erddap/
-  response:
-    body:
-      string: "<html>\r\n<head><title>301 Moved Permanently</title></head>\r\n<body>\r\n<center><h1>301
-        Moved Permanently</h1></center>\r\n<hr><center>nginx/1.16.1</center>\r\n</body>\r\n</html>\r\n"
-    headers:
-      Connection:
-      - keep-alive
-      Content-Length:
-      - '169'
-      Content-Type:
-      - text/html
-      Date:
-      - Fri, 23 Jul 2021 16:58:45 GMT
-      Location:
-      - https://gliders.ioos.us/erddap/
-      Server:
-      - nginx/1.16.1
-    status:
-      code: 301
-      message: Moved Permanently
-- request:
-    body: null
-    headers:
-      Accept:
-      - '*/*'
-      Accept-Encoding:
-      - gzip, deflate
-      Connection:
-      - keep-alive
-      User-Agent:
-      - python-requests/2.25.1
-    method: GET
-    uri: https://gliders.ioos.us/erddap/
-  response:
-    body:
-      string: ''
-    headers:
-      Access-Control-Allow-Origin:
-      - '*'
-      Connection:
-      - keep-alive
-      Content-Length:
-      - '0'
-      Date:
-      - Fri, 23 Jul 2021 16:58:45 GMT
-      Location:
-      - https://gliders.ioos.us/erddap/index.html
-      Server:
-      - nginx/1.16.1
-      X-Cache:
-      - MISS
-    status:
-      code: 302
-      message: ''
-- request:
-    body: null
-    headers:
-      Accept:
-      - '*/*'
-      Accept-Encoding:
-      - gzip, deflate
-      Connection:
-      - keep-alive
-      User-Agent:
-      - python-requests/2.25.1
-    method: GET
[…]
