+++ .github/workflows/publish.yml
@@ -38,7 +38,7 @@ jobs:
       shell: bash
 
     - name: Publish a Python distribution to PyPI
-      if: contains(github.ref, "release")
+      if: ${{ github.event_name == 'create' && github.event.ref_type == 'tag' }}
       uses: pypa/gh-action-pypi-publish@master
       with:
         user: __token__
