### 2021-09-24 0fa1127
better support for relative constraints
