+++ .pre-commit-config.yaml
@@ -66,7 +66,7 @@ repos:
         - --quiet-level=2
 
 - repo: https://github.com/asottile/pyupgrade
-  rev: v2.26.0
+  rev: v2.27.0
   hooks:
     - id: pyupgrade
       args:
