+++ erddapy/erddapy.py
@@ -328,8 +328,8 @@ class ERDDAP:
             base += "&searchFor={searchFor}"
 
         # Convert dates from datetime to `seconds since 1970-01-01T00:00:00Z`.
-        min_time = kwargs.pop("min_time", None)
-        max_time = kwargs.pop("max_time", None)
+        min_time = kwargs.pop("min_time", "")
+        max_time = kwargs.pop("max_time", "")
         if min_time and not _check_substrings(min_time):
             kwargs.update({"min_time": parse_dates(min_time)})
         else:
