+++ erddapy/multiple_server_search.py
@@ -4,6 +4,7 @@ from typing.io import BinaryIO
 
 import pandas as pd
 import requests
+
 try:
     joblib = True
     from joblib import Parallel, delayed
@@ -92,7 +93,9 @@ def search_servers(
     if parallel:
         num_cores = multiprocessing.cpu_count()
         if not joblib:
-            raise ImportError("Missing joblib. Please install it to use parallel searches.")
+            raise ImportError(
+                "Missing joblib. Please install it to use parallel searches.",
+            )
         returns = Parallel(n_jobs=num_cores)(
             delayed(fetch_results)(url, key, protocol=protocol)
             for key, url in urls.items()
@@ -128,7 +131,9 @@ def advanced_search_servers(
     if parallel:
         num_cores = multiprocessing.cpu_count()
         if not joblib:
-            raise ImportError("Missing joblib. Please install it to use parallel searches.")
+            raise ImportError(
+                "Missing joblib. Please install it to use parallel searches.",
+            )
         returns = Parallel(n_jobs=num_cores)(
             delayed(fetch_results)(url, key, protocol=protocol)
             for key, url in urls.items()
