+++ .pre-commit-config.yaml
@@ -15,7 +15,7 @@ repos:
       files: requirements-dev.txt
 
 - repo: https://github.com/PyCQA/flake8
-  rev: 3.9.2
+  rev: 4.0.1
   hooks:
     - id: flake8
       exclude: docs/source/conf.py
@@ -35,7 +35,7 @@ repos:
     language_version: python3
 
 - repo: https://github.com/pre-commit/mirrors-mypy
-  rev: v0.910
+  rev: v0.910-1
   hooks:
   - id: mypy
     exclude: docs/source/conf.py
@@ -73,7 +73,7 @@ repos:
         - --py36-plus
 
 - repo: https://github.com/asottile/add-trailing-comma
-  rev: v2.1.0
+  rev: v2.2.0
   hooks:
     - id: add-trailing-comma
 
