### 2021-12-06 72b9f85
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/pycqa/isort: 5.9.3 → 5.10.1](https://github.com/pycqa/isort/compare/5.9.3...5.10.1)
- [github.com/psf/black: 21.10b0 → 21.12b0](https://github.com/psf/black/compare/21.10b0...21.12b0)
- [github.com/asottile/pyupgrade: v2.29.0 → v2.29.1](https://github.com/asottile/pyupgrade/compare/v2.29.0...v2.29.1)
- [github.com/asottile/add-trailing-comma: v2.2.0 → v2.2.1](https://github.com/asottile/add-trailing-comma/compare/v2.2.0...v2.2.1)
