+++ .pre-commit-config.yaml
@@ -22,14 +22,14 @@ repos:
       args: [--max-line-length=105]
 
 - repo: https://github.com/pycqa/isort
-  rev: 5.9.3
+  rev: 5.10.1
   hooks:
     - id: isort
       additional_dependencies: [toml]
       args: ["--profile", "black", "--filter-files"]
 
 - repo: https://github.com/psf/black
-  rev: 21.10b0
+  rev: 21.12b0
   hooks:
   - id: black
     language_version: python3
@@ -66,14 +66,14 @@ repos:
         - --quiet-level=2
 
 - repo: https://github.com/asottile/pyupgrade
-  rev: v2.29.0
+  rev: v2.29.1
   hooks:
     - id: pyupgrade
       args:
         - --py36-plus
 
 - repo: https://github.com/asottile/add-trailing-comma
-  rev: v2.2.0
+  rev: v2.2.1
   hooks:
     - id: add-trailing-comma
 
