### 2022-01-06 f56c8ec
move tests to pre-commit instead of GHA
