+++ .pre-commit-config.yaml
@@ -82,6 +82,12 @@ repos:
   hooks:
     - id: prenotebook
 
+-   repo: https://github.com/pycqa/pydocstyle
+    rev: 6.1.1
+    hooks:
+      - id: pydocstyle
+        exclude: ^(docs|setup.py)
+
 ci:
     autofix_commit_msg: |
         [pre-commit.ci] auto fixes from pre-commit.com hooks
