+++ .github/workflows/tests.yml
@@ -42,10 +42,3 @@ jobs:
       run: |
         source activate TEST
         python -m pytest -rxs --cov=erddapy --vcr-record=none tests
-
-    - name: Pydocstyle
-      shell: bash -l {0}
-      run: |
-        source activate TEST
-        python -m pytest -rxs --cov=erddapy --vcr-record=none tests
-        pydocstyle erddapy
