+++ tests/test_to_objects.py
@@ -1,3 +1,5 @@
+"""Test converting to other data models objects."""
+
 import sys
 
 import iris
@@ -10,6 +12,7 @@ from erddapy import ERDDAP
 @pytest.fixture
 @pytest.mark.web
 def e():
+    """Instantiate ERDDAP class for testing."""
     yield ERDDAP(
         server="https://upwell.pfeg.noaa.gov/erddap",
         response="htmlTable",
@@ -18,6 +21,7 @@ def e():
 
 @pytest.fixture
 def bermuda_1_msl(e):
+    """Load bermuda_1_msl for testing."""
     e.dataset_id = "noaa_ngdc_9cfa_244d_0065"
     e.constraints = None
     e.protocol = "griddap"
@@ -27,6 +31,7 @@ def bermuda_1_msl(e):
 @pytest.fixture
 @pytest.mark.vcr()
 def taodata(e):
+    """Load taodata for testing."""
     e.dataset_id = "pmelTao5dayIso"
     e.protocol = "tabledap"
     e.constraints = {
@@ -44,6 +49,7 @@ def taodata(e):
 @pytest.mark.web
 @pytest.mark.vcr()
 def test_to_pandas(taodata):
+    """Test converting tabledap to a pandas DataFrame."""
     import pandas as pd
 
     df = taodata.to_pandas(index_col="time (UTC)", parse_dates=True).dropna()
@@ -57,6 +63,7 @@ def test_to_pandas(taodata):
 @pytest.mark.web
 @pytest.mark.vcr()
 def test_to_xarray_tabledap(taodata):
+    """Test converting tabledap to an xarray Dataset."""
     ds = taodata.to_xarray()
 
     assert isinstance(ds, xr.Dataset)
@@ -68,6 +75,7 @@ def test_to_xarray_tabledap(taodata):
 @pytest.mark.web
 @pytest.mark.vcr()
 def test_to_xarray_griddap(bermuda_1_msl):
+    """Test converting griddap to an xarray Dataset."""
     ds = bermuda_1_msl.to_xarray()
 
     assert isinstance(ds, xr.Dataset)
@@ -77,6 +85,7 @@ def test_to_xarray_griddap(bermuda_1_msl):
 @pytest.mark.vcr()
 @pytest.mark.skipif(sys.platform == "win32", reason="does not run on windows")
 def test_to_iris_tabledap(taodata):
+    """Test converting tabledap to an iris cube."""
     cubes = taodata.to_iris()
 
     assert isinstance(cubes, iris.cube.CubeList)
@@ -88,5 +97,6 @@ def test_to_iris_tabledap(taodata):
 @pytest.mark.vcr()
 @pytest.mark.skipif(sys.platform == "win32", reason="does not run on windows")
 def test_to_iris_griddap(bermuda_1_msl):
+    """Test converting griddap to an iris cube."""
     cubes = bermuda_1_msl.to_iris()
     assert isinstance(cubes, iris.cube.CubeList)
