### 2022-01-31 82b86fd
Remove Pandas FutureWarning
  This removes the Pandas FutureWarning triggered by the use of
  frame.append()

  Summary of changes:
    - Replace `frame.append()` with `pd.concat()`
    - Replace `table.index =` with `.set_index(drop=True)`
