### 2022-01-31 bc1ca0f
Fix set_index method in function _get_griddap_constraints
