+++ setup.cfg
@@ -8,6 +8,10 @@ long_description_content_type = text/markdown
 long_description = file: README.md
 license = BSD-3-Clause
 license_file = LICENSE.txt
+project_urls =
+    Documentation = https://ioos.github.io/erddapy
+    Bug Tracker = https://github.com/ioos/erddapy/issues
+    Source Code = hhttps://github.com/ioos/erddapy
 classifiers =
     Development Status :: 5 - Production/Stable
     Intended Audience :: Science/Research
