### 2022-03-14 2934c8a
retry vcr and skip win+macOS
