+++ tests/cassettes/test_search_servers_with_a_list[False].yaml
@@ -2316,4 +2316,963 @@ interactions:
     status:
       code: 200
       message: ''
+- request:
+    body: null
+    headers:
+      Accept:
+      - '*/*'
+      Accept-Encoding:
+      - gzip, deflate, br
+      Connection:
+      - keep-alive
+      User-Agent:
+      - python-requests/2.27.1
+    method: GET
+    uri: https://coastwatch.pfeg.noaa.gov/erddap/search/index.csv?page=1&itemsPerPage=100000&searchFor=%22sst%22
+  response:
+    body:
+      string: !!binary |
+        H4sIAAAAAAAAAOy9a3PjOJYt+n1+BSIj7owdLcmWRWUq60Tfc1V+pWvSj7aclTN36oSDFmmLnSSo
+        IiW73B/ub797g09JlC0BG5TsREd1lS0Lmw8Aay0ACxsPkec49rgxmN7F7qQxse98F38/t3+4rM9O
+        I3s8ajwFcePe89240R8O3Tj24EuNG28C/x5Mg8COnhsnp0eHjbPBJWt/bre7jTN+HzZ+tYc/HqJw
+        yh0mfr8eDBrHge358Od44k2mEy/kjSN7YsOl2dnRv40mk3H8y97eMLTjyZM9GY5a43v3ocVD2249
+        hI97boQ3u/eQ3PTeP8f+aXswuGk0GrJFWw/iCVctDm8iL7oXuX9O3XjSaIynd743bHw4/QKPeMNO
+        /fDO9lm7+SNgA9dmg2l0bw9dduMGYzeyJ9PIZTsixG4j/XKD7bf22+zIfYhct8EO9tv7TfjXpwY7
+        gvf1/KHxoc8c/KnBHpLoSwNjWPjuxGb4Vr2YjaPQmQ5dh9mT5J52bD8O2Q8ePnFmx2zqTyK7OfIe
+        Rixy49AX1bLL7p7ZZOSy366+suvL8wHbuXYf4A9w6cuha3N2Hjqu7/EHNniOJ26wy7Cqxy12A4U8
+        Pp5OGL6L7EZiZsPNOW7kPcKd3EdhIKKnr+oLXLx5nV/8lYe78vxwwq6i8J/ucAJvUrz15tXVboud
+        hBFcZAJvKm7AZd0/eFaz49Cx7WELKq/F7TipVCdpenvwjLeX364vD4/7F82v1rcv99xpnn69/LUp
+        aom1/uA3I3iT6ffFS7WjCQvvxUOcwxv0mrHLY7j6WQDvG59xyTPEbOf8TDzGOHmABnsaecPRfNBT
+        fJvsHiKu/HKSF5EHhrv+gx+dDQ4vL27OLr4dH7GmqJxh5NqTMIqzK2XtOX+8eBKOx/AEScMR1/Q4
+        8+2JC1U/wcb5SVQzXBMrNQjhX5E7dPmECZRg9iNUACJJUtHwehvsDhoEXO1ZFLn37Se4ALTC/CMO
[…]
