+++ tests/cassettes/test_fetch_results.yaml
@@ -358,4 +358,1218 @@ interactions:
     status:
       code: 200
       message: ''
+- request:
+    body: null
+    headers:
+      Accept:
+      - '*/*'
+      Accept-Encoding:
+      - gzip, deflate, br
+      Connection:
+      - keep-alive
+      User-Agent:
+      - python-requests/2.27.1
+    method: GET
+    uri: https://gliders.ioos.us/erddap/search/index.csv?page=1&itemsPerPage=100000&searchFor=%22sst%22
+  response:
+    body:
+      string: !!binary |
+        H4sIAAAAAAAAAOy9/Y8bR5L3+fv9FYUFblYGWlJl8f0BFgeNPevzYmT5LN4scDBOYHdT3TyzyR6+
+        SNb+9ReZVSSrMzM4ysioGdP9fbCPR1Kzvhl8q++nIyMj7jaL29vZ49X7/fV2vrvaza6Xc/v3t7Nf
+        58Wb4vvN7PH+6vPD9urjYjnfXk0Xu+WcHvzwMNt8ufrP77/79uqH9+8KMzFmcPXD6uP66s+zm1/v
+        Nuv96rZwf//5/furvzzMFkv68Xa32O13i/Xq6rvZbkbrFT98979d3e92j9v/9fr13XJxO99sXy3W
+        6+2r/fb1fGMje30I6fVmX41fVmVVmrKcTGnBwattHbVcIWfxO/fa/CMF97pFLn99FYnm376bPy7X
+        Xx7mq12x/ljMivfL9c3+oai1i926eJxvPq43D8V2PtuuV7Nlsd1vPs2/bO3Dbxfb7Xr5aX5brH/7
+        cjdfFTfr1Q1JbWb2Nd8Wi1Wxu58X2/vZcrn+TD+dbXek8Hm2o7itwI/zz8V/0Z/nX14VxfR+sS1u
+        T/HQ3x5nGxeXVTk9tqCg6QeHoP+y+rTYrFf2ryT+02a9m9/Y9f99W3yZzzbLL8HCdbAP69Vit94s
+        VnfF/CM9x932EETz7G9mm82Cntys+Hb63VXxfj67Xmxui/c3C1pq8XFxU/zl23fFbrN4XNIn6+Ny
+        v97Mt/T8b+Yvr+lDub2Z7Wi9lzf3y/Vm/Xj/Zbkkqe18tV1vrNrf9+vFrK321x/ev58WVVn+5hb7
+        24Ke8efZp/nxmr/NH27Wxdv1Nb3FxXQzW9Fqi08U6Yw+/G/oP/PNbBa+K/Xlr35Z/bK6uX34cEtf
+        hQ+7L4/z4j+syP9Hr9Z684VeN/vJ+WX1tzc///Dmz3/9y/v/9Yt9J5sfFy9ODy1+nD3Mv/ll9flh
+        /WFx+8vqsb7U/Xm3eJgXLxqxYkp/u6L16XNxuy22C3pp6Ks7Kl+Whv5vWpb/y/3f/0Niy5n9rt62
+        Lv5r8y9X9KG428zn2w8repPu7WPXqzv/wYd/Oj2aPrA7ejB9onb3xYsH+uP1nJ75qCxXD8WLP9Of
[…]

[…]
