+++ tests/test_multiple_server_search.py
@@ -32,42 +32,54 @@ def test_fetch_no_results():
     assert data is None
 
 
-@pytest.mark.vcr()
+# I guess we cannot record vcrs with parallel requests.
+# @pytest.mark.vcr()
 @pytest.mark.web
 @pytest.mark.skipif(
-    sys.platform == "win32",
+    sys.platform in ["win32", "darwin"],
     reason="run only on linux to avoid extra load on the server",
 )
-@pytest.mark.skipif(
-    sys.platform == "darwin",
-    reason="run only on linux to avoid extra load on the server",
-)
-@pytest.mark.parametrize("parallel", [True, False])
-def test_search_awesome_erddap_servers(parallel):
-    """Test multiple server search on awesome ERDDAP list."""
+def test_search_awesome_erddap_servers_True():
+    """Test multiple server search on awesome ERDDAP list parallel=True."""
     query = "glider"
     protocol = "tabledap"
     df = search_servers(
         query=query,
         protocol=protocol,
-        parallel=parallel,
+        parallel=True,
     )
     assert df is not None
     assert not df.empty
 
 
+# https://github.com/kevin1024/vcrpy/issues/533
+# @pytest.mark.vcr()
 @pytest.mark.web
-@pytest.mark.vcr()
 @pytest.mark.skipif(
-    sys.platform == "win32",
+    sys.platform in ["win32", "darwin"],
     reason="run only on linux to avoid extra load on the server",
 )
+def test_search_awesome_erddap_servers_False():
+    """Test multiple server search on awesome ERDDAP list with parallel=False."""
+    query = "glider"
+    protocol = "tabledap"
+    df = search_servers(
+        query=query,
+        protocol=protocol,
+        parallel=False,
+    )
+    assert df is not None
+    assert not df.empty
+
+
+# I guess we cannot record vcrs with parallel requests.
+# @pytest.mark.vcr()
+@pytest.mark.web
 @pytest.mark.skipif(
-    sys.platform == "darwin",
+    sys.platform in ["win32", "darwin"],
     reason="run only on linux to avoid extra load on the server",
 )
-@pytest.mark.parametrize("parallel", [True, False])
-def test_search_servers_with_a_list(parallel):
+def test_search_servers_with_a_list_True():
     """
     Check that downloads are made and that serial and parallel results are similar.
 
@@ -85,7 +97,38 @@ def test_search_servers_with_a_list(parallel):
         query=query,
         servers_list=servers_list,
         protocol=protocol,
-        parallel=parallel,
+        parallel=True,
+    )
+
+    assert df is not None
+    assert not df.empty
+
+
+@pytest.mark.vcr()
+@pytest.mark.web
+@pytest.mark.skipif(
+    sys.platform in ["win32", "darwin"],
+    reason="run only on linux to avoid extra load on the server",
+)
+def test_search_servers_with_a_list_False():
+    """
+    Check that downloads are made and that serial and parallel results are similar.
+
+    Ideally they should be identical but the servers are live
+    and changes from one request to another can happen.
+
+    """
+    servers_list = [
+        "https://coastwatch.pfeg.noaa.gov/erddap/",
+        "https://gliders.ioos.us/erddap/",
+    ]
+    query = "sst"
+    protocol = "griddap"
+    df = search_servers(
+        query=query,
+        servers_list=servers_list,
+        protocol=protocol,
+        parallel=False,
     )
 
     assert df is not None
