+++ tests/test_multiple_server_search.py
@@ -76,8 +76,8 @@ def test_search_awesome_erddap_servers_False():
 # @pytest.mark.vcr()
 @pytest.mark.web
 @pytest.mark.skipif(
-    sys.platform in ["win32", "darwin"],
-    reason="run only on linux to avoid extra load on the server",
+    (sys.platform in ["win32", "darwin"] or sys.version_info < (3, 9)),
+    reason="run only on linux and latest to avoid extra load on the server",
 )
 def test_search_servers_with_a_list_True():
     """
