+++ .pre-commit-config.yaml
@@ -35,7 +35,7 @@ repos:
     language_version: python3
 
 - repo: https://github.com/pre-commit/mirrors-mypy
-  rev: v0.931
+  rev: v0.941
   hooks:
   - id: mypy
     exclude: docs/source/conf.py
@@ -66,7 +66,7 @@ repos:
         - --quiet-level=2
 
 - repo: https://github.com/asottile/pyupgrade
-  rev: v2.23.3
+  rev: v2.31.1
   hooks:
     - id: pyupgrade
       args:
