+++ erddapy/erddapy.py
@@ -557,7 +557,10 @@ class ERDDAP:
             for k, v in _constraints.items():
                 if _check_substrings(v):
                     continue
-                if k.startswith("time"):
+                # The valid operators are
+                # =, != (not equals), =~ (a regular expression test), <, <=, >, and >=
+                seps = ["=", "!=", "=~", "<", "<=", ">", ">="]
+                if k.startswith("time") and any([sep in k for sep in seps]):
                     _constraints.update({k: parse_dates(v)})
             _constraints = _quote_string_constraints(_constraints)
             _constraints_url = _format_constraints_url(_constraints)
