### 2022-03-15 a61ed93
safer way to ensure the constraint is time
