+++ setup.cfg
@@ -11,7 +11,7 @@ license_file = LICENSE.txt
 project_urls =
     Documentation = https://ioos.github.io/erddapy
     Bug Tracker = https://github.com/ioos/erddapy/issues
-    Source Code = hhttps://github.com/ioos/erddapy
+    Source Code = https://github.com/ioos/erddapy
 classifiers =
     Development Status :: 5 - Production/Stable
     Intended Audience :: Science/Research
