### 2022-03-16 7e7c4f4
drop py36 and pandas <1
