+++ .pre-commit-config.yaml
@@ -29,13 +29,13 @@ repos:
       args: ["--profile", "black", "--filter-files"]
 
 - repo: https://github.com/psf/black
-  rev: 22.1.0
+  rev: 22.3.0
   hooks:
   - id: black
     language_version: python3
 
 - repo: https://github.com/pre-commit/mirrors-mypy
-  rev: v0.941
+  rev: v0.942
   hooks:
   - id: mypy
     exclude: docs/source/conf.py
@@ -73,7 +73,7 @@ repos:
         - --py36-plus
 
 - repo: https://github.com/asottile/add-trailing-comma
-  rev: v2.2.1
+  rev: v2.2.2
   hooks:
     - id: add-trailing-comma
 
