+++ tests/test_url_handling.py
@@ -2,8 +2,8 @@
 
 import io
 
+import httpx
 import pytest
-from requests.exceptions import HTTPError, ReadTimeout
 
 from erddapy.url_handling import check_url_response, urlopen
 
@@ -22,21 +22,21 @@ def test_urlopen():
 def test_urlopen_raise():
     """Assure that urlopen will raise for bad URLs."""
     url = "https://developer.mozilla.org/en-US/404"
-    with pytest.raises(HTTPError):
+    with pytest.raises(httpx.HTTPError):
         urlopen(url)
 
 
 @pytest.mark.web
 def test_urlopen_requests_kwargs():
     """Test that urlopen can pass kwargs to requests."""
-    base_url = "http://erddap.sensors.ioos.us/erddap/tabledap/"
+    base_url = "https://erddap.sensors.ioos.us/erddap/tabledap/"
     timeout_seconds = 1  # request timeout in seconds
     slowwly_milliseconds = (timeout_seconds + 1) * 1000
     slowwly_url = (
         f"https://flash.siwalik.in/delay/{slowwly_milliseconds}/url/{base_url}"
     )
 
-    with pytest.raises(ReadTimeout):
+    with pytest.raises(httpx.ReadTimeout):
         urlopen(slowwly_url, timeout=timeout_seconds)
 
 
@@ -51,5 +51,5 @@ def test_check_url_response():
         "&time>=2017-08-29T00:00:00Z"
         "&time<=2015-09-05T19:00:00Z"
     )
-    with pytest.raises(HTTPError):
+    with pytest.raises(httpx.HTTPError):
         check_url_response(bad_request)
