+++ erddapy/netcdf_handling.py
@@ -9,11 +9,11 @@ from urllib.parse import urlparse
 from erddapy.url_handling import urlopen
 
 
-def _nc_dataset(url, auth, **request_kwargs: Dict):
+def _nc_dataset(url, auth, **requests_kwargs: Dict):
     """Return a netCDF4-python Dataset from memory and fallbacks to disk if that fails."""
     from netCDF4 import Dataset
 
-    data = urlopen(url=url, auth=auth, **request_kwargs)
+    data = urlopen(url=url, auth=auth, **requests_kwargs)
     try:
         return Dataset(Path(urlparse(url).path).name, memory=data.read())
     except OSError:
