+++ tests/test_url_handling.py
@@ -32,9 +32,7 @@ def test_urlopen_requests_kwargs():
     base_url = "https://erddap.sensors.ioos.us/erddap/tabledap/"
     timeout_seconds = 1  # request timeout in seconds
     slowwly_milliseconds = (timeout_seconds + 1) * 1000
-    slowwly_url = (
-        f"https://flash-the-slow-api.herokuapp.com/delay/{slowwly_milliseconds}/url/{base_url}"
-    )
+    slowwly_url = f"https://flash-the-slow-api.herokuapp.com/delay/{slowwly_milliseconds}/url/{base_url}"
 
     with pytest.raises(httpx.ReadTimeout):
         urlopen(slowwly_url, timeout=timeout_seconds)
