+++ .pre-commit-config.yaml
@@ -35,7 +35,7 @@ repos:
     language_version: python3
 
 - repo: https://github.com/pre-commit/mirrors-mypy
-  rev: v0.950
+  rev: v0.960
   hooks:
   - id: mypy
     exclude: docs/source/conf.py
