### 2022-06-13 ef9477a
Remove pre-commit hook 'prenotebook'
  This hook will be disabled while the CI problems persist.
  See https://github.com/deathbeds/prenotebook/issues/6 for discussion.
