+++ .pre-commit-config.yaml
@@ -77,11 +77,6 @@ repos:
   hooks:
     - id: add-trailing-comma
 
-- repo: https://github.com/deathbeds/prenotebook
-  rev: f5bdb72a400f1a56fe88109936c83aa12cc349fa
-  hooks:
-    - id: prenotebook
-
 -   repo: https://github.com/pycqa/pydocstyle
     rev: 6.1.1
     hooks:
