### 2022-06-20 84a2449
Modify 'get_search_url' to return all items if non-html response (#244)
  - Set items_per_page variable to 1e6 if response is either csv or json
