### 2022-06-20 05fb3d6
Add test for CSV and JSON searches
  - Add new fixture 'gliders' - has more than 1000 datasets
  - Adds tests 'test_csv_search' and 'test_json_search' that check if the response
  has more than 1000 records
