### 2022-07-18 09139f8
Create core subpackage
  - Create erddapy/core subpackage
  - Move module 'url_handling' to core subpackage
  - Refactor imports
