### 2022-07-18 9346051
Mark more response types as 'non_paginated'
  - On method ERDDAP.get_search_url, add all CSV, JSON and TSV-type responses to variable 'non_paginated_responses'
