### 2022-07-18 bce48ee
Move module 'netcdf_handling' to subpackage 'core'
  - Move module to subpackage
  - Refactor imports
