+++ tests/test_netcdf_handling.py
@@ -4,8 +4,8 @@ import os
 
 import pytest
 
+from erddapy.core.netcdf_handling import _nc_dataset, _tempnc
 from erddapy.core.url_handling import urlopen
-from erddapy.netcdf_handling import _nc_dataset, _tempnc
 
 
 @pytest.mark.web
