### 2022-07-25 899b2c3
Refactor tests/test_erddapy.py (#258)
  - Refactor test URLs
    - Change search string
  - Record new cassette
