+++ tests/cassettes/test_erddap2_10.yaml
@@ -9,28 +9,40 @@ interactions:
       connection:
       - keep-alive
       host:
-      - coastwatch.pfeg.noaa.gov
+      - erddap.ioos.us
       user-agent:
-      - python-httpx/0.22.0
+      - python-httpx/0.23.0
     method: HEAD
-    uri: https://coastwatch.pfeg.noaa.gov/erddap/search/advanced.csv?page=1&itemsPerPage=1000&protocol=(ANY)&cdm_data_type=(ANY)&institution=(ANY)&ioos_category=(ANY)&keywords=(ANY)&long_name=(ANY)&standard_name=(ANY)&variableName=(ANY)&minLon=(ANY)&maxLon=(ANY)&minLat=(ANY)&maxLat=(ANY)&minTime=&maxTime=&searchFor=whoi
+    uri: http://erddap.ioos.us/erddap/search/advanced.csv?page=1&itemsPerPage=1000&protocol=(ANY)&cdm_data_type=(ANY)&institution=(ANY)&ioos_category=(ANY)&keywords=(ANY)&long_name=(ANY)&standard_name=(ANY)&variableName=(ANY)&minLon=(ANY)&maxLon=(ANY)&minLat=(ANY)&maxLat=(ANY)&minTime=&maxTime=&searchFor=NOAA
   response:
     content: ''
     headers:
       Connection:
-      - close
+      - keep-alive
       Content-Disposition:
       - attachment;filename=AdvancedSearch.csv
       Content-Encoding:
       - gzip
       Content-Length:
-      - '5908'
+      - '3960'
       Content-Type:
       - text/csv;charset=ISO-8859-1
       Date:
-      - Tue, 26 Apr 2022 16:45:16 GMT
+      - Mon, 25 Jul 2022 15:05:00 GMT
+      Keep-Alive:
+      - timeout=20
+      Server:
+      - Apache
       Strict-Transport-Security:
-      - max-age=31536000; includeSubDomains
+      - max-age=0
+      X-Content-Type-Options:
+      - nosniff
+      X-Frame-Options:
+      - SAMEORIGIN
+      X-XSS-Protection:
+      - 1; mode=block
+      vary:
+      - Origin
     http_version: HTTP/1.1
     status_code: 200
 version: 1
