### 2022-07-27 fa7345c
Refactor tests/test_url_builder.py (#258)
  - Refactor test URLs
    - Change dataset_id, constraints, and variables
  - "Turn off" griddap search for now
  - Record new cassettes
