+++ tests/cassettes/test_search_url_change_protocol.yaml
@@ -9,33 +9,26 @@ interactions:
       connection:
       - keep-alive
       host:
-      - upwell.pfeg.noaa.gov
+      - standards.sensors.ioos.us
       user-agent:
-      - python-httpx/0.22.0
+      - python-httpx/0.23.0
     method: HEAD
-    uri: https://upwell.pfeg.noaa.gov/erddap/search/advanced.htmlTable?page=1&itemsPerPage=1000&protocol=tabledap&cdm_data_type=(ANY)&institution=(ANY)&ioos_category=(ANY)&keywords=(ANY)&long_name=(ANY)&standard_name=(ANY)&variableName=(ANY)&minLon=(ANY)&maxLon=(ANY)&minLat=(ANY)&maxLat=(ANY)&minTime=&maxTime=&searchFor=salinity
+    uri: https://standards.sensors.ioos.us/erddap/search/advanced.htmlTable?page=1&itemsPerPage=1000&protocol=tabledap&cdm_data_type=(ANY)&institution=(ANY)&ioos_category=(ANY)&keywords=(ANY)&long_name=(ANY)&standard_name=(ANY)&variableName=(ANY)&minLon=(ANY)&maxLon=(ANY)&minLat=(ANY)&maxLat=(ANY)&minTime=&maxTime=&searchFor=salinity
   response:
     content: ''
     headers:
       Connection:
-      - close
+      - keep-alive
       Content-Encoding:
       - gzip
-      Content-Security-Policy:
-      - script-src 'self' https://apis.google.com https://code.jquery.com/ https://www.google-analytics.com
-        stackpath.bootstrapcdn.com https://polarwatch.noaa.gov https://www.googletagmanager.com
-        https://upwell.pfeg.oaa.gov 'unsafe-inline' 'unsafe-eval'; font-src 'self'
-        https://fonts.googleapis.com  https://fonts.gstatic.com;
+      Content-Length:
+      - '1909'
       Content-Type:
       - text/html;charset=UTF-8
       Date:
-      - Tue, 26 Apr 2022 16:47:08 GMT
-      Strict-Transport-Security:
-      - max-age=31536000; includeSubDomains
-      Transfer-Encoding:
-      - chunked
-      X-Frame-Options:
-      - SAMEORIGIN
+      - Wed, 27 Jul 2022 14:53:55 GMT
+      Server:
+      - nginx/1.20.1
     http_version: HTTP/1.1
     status_code: 200
 - request:
@@ -48,72 +41,26 @@ interactions:
       connection:
       - keep-alive
       host:
-      - upwell.pfeg.noaa.gov
+      - standards.sensors.ioos.us
       user-agent:
-      - python-httpx/0.22.0
+      - python-httpx/0.23.0
     method: HEAD
-    uri: https://upwell.pfeg.noaa.gov/erddap/search/advanced.htmlTable?page=1&itemsPerPage=1000&protocol=griddap&cdm_data_type=(ANY)&institution=(ANY)&ioos_category=(ANY)&keywords=(ANY)&long_name=(ANY)&standard_name=(ANY)&variableName=(ANY)&minLon=(ANY)&maxLon=(ANY)&minLat=(ANY)&maxLat=(ANY)&minTime=&maxTime=&searchFor=salinity
+    uri: https://standards.sensors.ioos.us/erddap/search/advanced.htmlTable?page=1&itemsPerPage=1000&protocol=(ANY)&cdm_data_type=(ANY)&institution=(ANY)&ioos_category=(ANY)&keywords=(ANY)&long_name=(ANY)&standard_name=(ANY)&variableName=(ANY)&minLon=(ANY)&maxLon=(ANY)&minLat=(ANY)&maxLat=(ANY)&minTime=&maxTime=&searchFor=salinity
   response:
     content: ''
     headers:
[…]
