+++ tests/test_url_handling.py
@@ -12,7 +12,7 @@ from erddapy.url_handling import check_url_response, urlopen
 @pytest.mark.vcr()
 def test_urlopen():
     """Assure that urlopen is always a BytesIO object."""
-    url = "http://erddap.sensors.ioos.us/erddap/tabledap/"
+    url = "https://standards.sensors.ioos.us/erddap/tabledap/"
     ret = urlopen(url)
     isinstance(ret, io.BytesIO)
 
@@ -29,7 +29,7 @@ def test_urlopen_raise():
 @pytest.mark.web
 def test_urlopen_requests_kwargs():
     """Test that urlopen can pass kwargs to requests."""
-    base_url = "https://erddap.sensors.ioos.us/erddap/tabledap/"
+    base_url = "https://standards.sensors.ioos.us/erddap/tabledap/"
     timeout_seconds = 1  # request timeout in seconds
     slowwly_milliseconds = (timeout_seconds + 1) * 1000
     slowwly_url = f"https://flash-the-slow-api.herokuapp.com/delay/{slowwly_milliseconds}/url/{base_url}"
@@ -43,7 +43,7 @@ def test_urlopen_requests_kwargs():
 def test_check_url_response():
     """Test if a bad request returns HTTPError."""
     bad_request = (
-        "http://erddap.sensors.ioos.us/erddap/tabledap/"
+        "https://standards.sensors.ioos.us/erddap/tabledap/"
         "gov_usgs_waterdata_340800117235901.htmlTable?"
         "time,"
         "&time>=2017-08-29T00:00:00Z"
