### 2022-07-27 5950cb3
Refactor tests/test_multiple_server_search.py (#258)
  - Refactor test URLs
  - Create fixture 'servers'
  - Record new cassettes
