### 2022-07-27 8e1418c
Rename fixture 'servers' to 'servers_list'
