### 2022-07-27 e59a850
Refactor 'test_check_url_response'
  - Modify dataset_id so correct error status will be raised (400, not 404)
  - Rewrite cassette
