### 2022-07-28 ec79c31
Code review for #255
  - Improve comment in function 'get_search_url'
  - Move duplicated comment to fixture 'gliders'
