### 2022-07-28 bbf11db
Move helper functions to core/url_handling
  - Move helper functions from multiple_server_search.py to core/url_handling.py

  These functions may be reutilised in the future so they will be moved to the url module.
  The '_format_results' function was kept in place to avoid a pandas import in the url module.
