### 2022-07-28 80d49a2
Create 'servers' directory
  - Move module 'servers.py' to new directory
  - Move file 'erddaps.json' to new directory
  - Add __init__ module
  - Refactor imports (in erddapy.__init__ and erddapy.erddapy)
