+++ erddapy/erddapy.py
@@ -12,7 +12,7 @@ from pandas._libs.tslibs.parsing import parse_time_string
 
 from erddapy.core.netcdf_handling import _nc_dataset, _tempnc
 from erddapy.core.url_handling import _distinct, urlopen
-from erddapy.servers import servers
+from erddapy.servers.servers import servers
 
 ListLike = Union[List[str], Tuple[str]]
 OptionalStr = Optional[str]
