+++ erddapy/servers/__init__.py
@@ -1,0 +1,6 @@
+"""
+Servers subpackage for the erddapy parent package.
+
+Contains the servers.py module and the erddaps.json file,
+which is a fallback when the awesome-erddap URL cannot be reached.
+"""
