+++ erddapy/__init__.py
@@ -1,7 +1,7 @@
 """Easier access to scientific data."""
 
 from erddapy.erddapy import ERDDAP
-from erddapy.servers import servers
+from erddapy.servers.servers import servers
 
 __all__ = ["ERDDAP", "servers"]
 
