### 2022-08-01 59291fd
Make servers object available from init module
  This allows importing the servers dict with 'from erddapy.servers import servers'
