+++ .pre-commit-config.yaml
@@ -15,7 +15,7 @@ repos:
       files: requirements-dev.txt
 
 - repo: https://github.com/PyCQA/flake8
-  rev: 4.0.1
+  rev: 5.0.2
   hooks:
     - id: flake8
       exclude: docs/source/conf.py
@@ -35,7 +35,7 @@ repos:
     language_version: python3
 
 - repo: https://github.com/pre-commit/mirrors-mypy
-  rev: v0.961
+  rev: v0.971
   hooks:
   - id: mypy
     exclude: docs/source/conf.py
@@ -43,7 +43,7 @@ repos:
     args: ["--ignore-missing-imports", "--python-executable=CONDA_PREFIX"]
 
 - repo: https://github.com/keewis/blackdoc
-  rev: v0.3.4
+  rev: v0.3.5
   hooks:
     - id: blackdoc
 
@@ -66,7 +66,7 @@ repos:
         - --quiet-level=2
 
 - repo: https://github.com/asottile/pyupgrade
-  rev: v2.34.0
+  rev: v2.37.3
   hooks:
     - id: pyupgrade
       args:
