### 2022-08-05 18bb075
Rename URL module
  - Rename erddapy.core.url_handling to erddapy.core.url
  - Refactor imports
