+++ erddapy/erddapy.py
@@ -11,7 +11,7 @@ import pytz
 from pandas._libs.tslibs.parsing import parse_time_string
 
 from erddapy.core.netcdf_handling import _nc_dataset, _tempnc
-from erddapy.core.url_handling import _distinct, urlopen
+from erddapy.core.url import _distinct, urlopen
 from erddapy.servers.servers import servers
 
 ListLike = Union[List[str], Tuple[str]]
