### 2022-08-05 1582282
Refactor functions out of erddapy.erddapy module
  - Move URL handling functions to erddapy.core.url
  - Move GRIDDAP handling functions to erddapy.core.griddap
  - Refactor imports
