### 2022-08-05 9aff0f6
Rename netcdf module
  - Rename erddapy.core.netcdf_handling to erddapy.core.netcdf
  - Refactor imports
