+++ tests/test_netcdf_handling.py
@@ -4,7 +4,7 @@ import os
 
 import pytest
 
-from erddapy.core.netcdf_handling import _nc_dataset, _tempnc
+from erddapy.core.netcdf import _nc_dataset, _tempnc
 from erddapy.core.url import urlopen
 
 
