### 2022-08-10 f5ef393
Refactor private '_search_url' function as public
  - Rename from '_search_url' to 'get_search_url'
  - Refactor imports
  - Add docstring
