+++ erddapy/erddapy.py
@@ -17,7 +17,7 @@ from erddapy.core.url import (
     _distinct,
     _format_constraints_url,
     _quote_string_constraints,
-    _search_url,
+    get_search_url,
     parse_dates,
     urlopen,
 )
@@ -196,26 +196,7 @@ class ERDDAP:
         protocol = protocol if protocol else self.protocol
         response = response if response else self.response
 
-        # These responses should not be paginated b/c that hinders the correct amount of data silently
-        # and can surprise users when the number of items is greater than ERDDAP's defaults (1000 items).
-        # Ideally there should be no pagination for this on the ERDDAP side but for now we settled for a
-        # "really big" `items_per_page` number.
-        non_paginated_responses = [
-            "csv",
-            "csvp",
-            "csv0",
-            "json",
-            "jsonlCSV1",
-            "jsonlCSV",
-            "jsonlKVP",
-            "tsv",
-            "tsvp",
-            "tsv0",
-        ]
-        if response in non_paginated_responses:
-            items_per_page = int(1e6)
-
-        return _search_url(
+        return get_search_url(
             self.server,
             response=response,
             search_for=search_for,
