### 2022-08-11 da9ceb1
Refactor 'get_info_url' and 'get_categorize_url'
  - Move functions to erddapy.core.url
  - Use functions in ERDDAP class
  - Refactor docstrings to use gold standard servers
