### 2022-08-29 04310c7
Refactor 'get_download_url'
  - Refactor method as a function in erddap.core.url
  - Use standalone function in ERDDAP class
  - Refactor imports accordingly (in tests as well)
    - Add '__all__' variable to erddapy
