### 2022-08-29 66545d1
Refactor netCDF tests
  - Use Gold Standard ERDDAP server for netCDF tests
  - Use small 'allDatasets' file
