### 2022-09-01 #1 @lwasser
Good morning, everyone. Because this issue has remained open since 2019 with not much movement, I am going to close it. @choldgraf @jlpalomino @ocefpaf @xmnlab @npch thank you all so much for helping us refine our review process through this review. If in the future Filipe, you wish to resubmit this you are welcome to. In the meantime pyopensci will be running in full force as of today (september 1) now that I am full time on the project. I can't wait to continue working with all of you on this effort.

For now... closing this issue. :)
