### 2022-09-01 #1 @ocefpaf
> In the meantime pyopensci will be running in full force as of today (september 1) now that I am full time on the project. I can't wait to continue working with all of you on this effort.

Those are great news! Glad to hear it!!

PS: the erddapy application was never really meant to be a real one, just a test-drive of the workflow. I'd be happy to re-submit a real one as soon as possible,
