### 2022-09-01 #1 @lwasser
@ocefpaf  hi!! 👋  yes i'm very excited and even more excited to get to work with you and the conda-forge team as it makes sense :) i totally understand erddapy was intended to be submitted to test our review process and we did learn a lot from it! you are welcome to submit to us any time you wish!! there is no rush - we aren't going anywhere!! Please do feel free to reach out if pyopensci can support any of the work you've been doing as well in the python ecosystem. 🎆
